### 2025-07-21 #8624 @editorialbot
[PRE REVIEW]: ggpedigree: Visualizing Pedigrees with 'ggplot2' and 'plotly'
**Submitting author:** @smasongarrison (<a href="http://orcid.org/0000-0002-4804-6003">S. Mason Garrison</a>)
**Repository:** https://github.com/R-Computing-Lab/ggpedigree
**Branch with paper.md** (empty if default branch):
**Version:** v0.8.0
