### 2025-07-21 03e4786
removed data processing to reduce paper length
