+++ pkgdown.yml
@@ -9,7 +9,7 @@ articles:
   v20_matrix: v20_matrix.html
   articles/v21_phenotypicbydegree: v21_phenotypicbydegree.html
   articles/v22_plots_morecomplexity: v22_plots_morecomplexity.html
-last_built: 2025-07-21T13:47Z
+last_built: 2025-07-21T13:51Z
 urls:
   reference: https://r-computing-lab.github.io/ggpedigree/reference
   article: https://r-computing-lab.github.io/ggpedigree/articles
