+++ vignettes/articles/_paper.Rmd
@@ -41,29 +41,14 @@ options(rmarkdown.html_vignette.check_title = FALSE)
 library(rticles)
 ```
 
-```{=html}
-<!--Guidance 
-JOSS welcomes submissions from broadly diverse research areas. For this reason, we require that authors include in the paper some sentences that explain the software functionality and domain of use to a non-specialist reader. We also require that authors explain the research applications of the software. The paper should be between 250-1000 words. Authors submitting papers significantly longer than 1000 words may be asked to reduce the length of their paper.
-Your paper should include:
-
-A list of the authors of the software and their affiliations, using the correct format (see the example below).
-A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience.
-A Statement of need section that clearly illustrates the research purpose of the software and places it in the context of related work.
-A list of key references, including to other software addressing related needs. Note that the references should include full names of venues, e.g., journals and conferences, not abbreviations only understood in the context of a specific discipline.
-Mention (if applicable) a representative set of past or ongoing research projects using the software and recent scholarly publications enabled by it.
-Acknowledgment of any financial support.
--->
-```
 
 # Summary
 
-<!--  A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience. -->
 
 Pedigree diagrams underpin research and practice across genetics, animal breeding, genealogy, forensics, and counseling. They help medical geneticists trace the inheritance of Mendelian diseases and identify at-risk relatives; enable dairy breeders to plan matings that improve milk yield; support genealogists in reconstructing ancestry; assist forensic scientists in establishing familial connections in criminal investigations; and facilitate family therapists and counselors in understanding their clients' relationships [@genograms_2020]. Early R tools such as kinship2 [@kinship2] plot simple nuclear families effectively, but they do not scale to today’s pedigrees that can exceed 1,000s of individuals. As datasets have grown, researchers now work with increasingly complex family structures, including large-scale plant breeding pedigrees [@shaw2014], web-based pedigree management systems [@Ranaweera2018open], interactive pedigree editors [@pedigreejs], and behavior genetic studies of extended family structures [@hunter_analytic_2021; @garrison_analyzing_2023]. That complexity exposes the limitations of existing tools, which often struggle with large and complex datasets. `ggpedigree` addresses this need by combining a vectorised layout algorithm, ggplot2 output, and optional plotly interactivity.
 
 # Statement of need
 
-<!-- A Statement of need section that clearly illustrates the research purpose of the software and places it in the context of related work. -->
 
 Pedigree visualization has traditionally relied on proprietary software (e.g., Progeny, GenoPro, Pedigree Viewer) or R packages like kinship2 [@kinship2], pedtools [@pedtools], or pedtricks [@pedtricks]. While these tools are functional for many use cases, their limitations become pronounced when working with complex, modern pedigree datasets or when more detailed customization is required. Most R packages focus on base graphics or simple ggplot2 implementations.
 
@@ -93,9 +78,9 @@ I describe the main features of the `ggpedigree` package below. More detailed de
 
[…]
