### 2025-07-21 842ec02
Update paper.bib
