+++ vignettes/articles/paper.bib
@@ -39,6 +39,7 @@
     title = {kinship2: Pedigree Functions},
     author = {{Sinnwell}, J. P. and {Therneau}, T. M.},
     year = {2022},
+	doi = {10.32614/CRAN.package.kinship2},
     note = {R package version 1.9.6},
     url = {https://CRAN.R-project.org/package=kinship2},
   }
@@ -331,7 +332,7 @@
 	journal = {Behavior Genetics},
 	pages = {488{\textendash}488},
 	volume = {54},
-	url = {https://scholar.google.com/scholar?cluster=4710819130852314260&hl=en&oi=scholarr},
+	doi = {10.1007/s10519-024-10197-8},
        note = {Abstract presented at the 54th Behavior Genetics Association Annual Meeting}
 }
 
@@ -342,7 +343,7 @@
 	journal = {Behavior Genetics},
 	pages = {531{\textendash}531},
 	volume = {54},
-	url = {https://scholar.google.com/scholar?cluster=16878992125450714238&hl=en&oi=scholarr},
+	doi = {10.1007/s10519-024-10197-8},
        note = {Abstract presented at the 54th Behavior Genetics Association Annual Meeting}
 
 }
@@ -451,7 +452,8 @@ PMCID: PMC4133633}
 	publisher = {Academic Press, an imprint of Elsevier},
 	url = {https://shop.elsevier.com/books/pedigree-analysis-in-r/vigeland/978-0-12-824430-2},
 	address = {London},
-	langid = {canadian}
+	langid = {canadian},
+	isbn = {978-0128244302}
 }
 
 @article{vigeland2022,
@@ -473,7 +475,7 @@ PMCID: PMC4133633}
 	author = {{Vigeland}, Magnus D.},
 	year = {2021},
 	doi = {10.32614/CRAN.package.pedtools},
-		url = {https://CRAN.R-project.org/package=pedtools},
+	url = {https://CRAN.R-project.org/package=pedtools},
 	date = {2021}
 }
 
