### 2025-07-21 #8624 @smasongarrison
The [suggested articles](https://github.com/openjournals/joss-reviews/issues/8624#issuecomment-3096870321) are all good hits for different aspects of the package. (Ok, rTASSEL is a little far afield) The one that stands out is schochastics of [graphlayouts](https://github.com/openjournals/joss-reviews/issues/5238), given that they're familiar with working with network structures.

mkerin from the reviewer database has interests in data visualization and familiar with large genetically informed samples [https://reviewers.joss.theoj.org/reviewers/3333]. They might be a good fit as well.

> **Five most similar historical JOSS papers:**
>
> ⚠️ **Important note to authors:** Please do not directly contact the authors of these papers. If you believe these papers are relevant to your submission, please mention this in your submission or communicate with your handling editor.
>
> [ggalluvial: Layered Grammar for Alluvial Plots](https://github.com/openjournals/joss-reviews/issues/2017) Submitting author: `corybrunson` Handling editor: `akeshavan (Retired)` Reviewers: `clauswilke, tweed1e` Similarity score: 0.7062
>
> [GGoutlieR: an R package to identify and visualize unusual geo-genetic patterns of biological samples](https://github.com/openjournals/joss-reviews/issues/5687) Submitting author: `CheWeiChang1992` Handling editor: `martinfleis (Retired)` Reviewers: `tkchafin, btmartin721` Similarity score: 0.7035
>
> [rTASSEL: An R interface to TASSEL for analyzing genomic diversity](https://github.com/openjournals/joss-reviews/issues/4530) Submitting author: `btmonier` Handling editor: `fboehm (Active)` Reviewers: `tkchafin, tomsing1` Similarity score: 0.6940
>
> [graphlayouts: Layout algorithms for network visualizations in R](https://github.com/openjournals/joss-reviews/issues/5238) Submitting author: `schochastics` Handling editor: `crvernon (Active)` Reviewers: `bpbond, ndjackso` Similarity score: 0.6875
>
> [Visualizations with statistical details: The ggstatsplot approach](https://github.com/openjournals/joss-reviews/issues/3167) Submitting author: `IndrajeetPatil` Handling editor: `csoneson (Active)` Reviewers: `njtierney, kevinrue` Similarity score: 0.6806
>
> ⚠️ **Note to editors:** If these papers look like they might be a good match, click through to the review issue for that paper and invite one or more of the authors _before_ considering asking the reviewers of these papers to review again for JOSS.
