### 2025-07-21 7ec01f9
push vectoized
