### 2025-07-21 b4afb5a
Update DESCRIPTION
Apply suggestions from code review

Update calcCoordinates.R

Co-Authored-By: Copilot <175728472+Copilot@users.noreply.github.com>
