+++ vignettes/v00_plots.R
@@ -50,7 +50,7 @@ ggPedigree(
     segment_linewidth = .5,
     point_size = 3,
     outline_multiplier = 1.5,
-   # outline_additional_size = -1,
+    # outline_additional_size = -1,
     sex_shape_female = "💸",
     sex_shape_male = "🖤",
     segment_spouse_color = viridis_pal()(5)[1],
@@ -77,12 +77,12 @@ ggPedigree(
   famID = "famID",
   personID = "personID",
   config = list(
-#    label_column = "name",
+    #    label_column = "name",
     sex_color_palette = c("pink", "blue"),
     label_text_angle = -45,
     label_nudge_y = .15,
     label_nudge_x = 0.45,
-    label_method = "ggrepel",#"geom_text",
+    label_method = "ggrepel", # "geom_text",
     #   sex_color_palette = c("black", "black"),
     sex_color_include = TRUE
   )
