+++ README.Rmd
@@ -68,21 +68,21 @@ library(tidyverse)
 # if you don't have the most recent version of BGmisc, you may need to install it first as a stop-gap I've added the data loading here
 data("potter") # load example data from BGmisc
 if (!"twinID" %in% names(potter)) {
-# Add twinID and zygosity columns for demonstration purposes
-potter <- potter %>%
-  mutate(
-    twinID = case_when(
-      name == "Fred Weasley" ~ 13,
-      name == "George Weasley" ~ 12,
-      TRUE ~ NA_real_
-    ),
-    zygosity = case_when(
-      name == "Fred Weasley" ~ "mz",
-      name == "George Weasley" ~ "mz",
-      TRUE ~ NA_character_
+  # Add twinID and zygosity columns for demonstration purposes
+  potter <- potter %>%
+    mutate(
+      twinID = case_when(
+        name == "Fred Weasley" ~ 13,
+        name == "George Weasley" ~ 12,
+        TRUE ~ NA_real_
+      ),
+      zygosity = case_when(
+        name == "Fred Weasley" ~ "mz",
+        name == "George Weasley" ~ "mz",
+        TRUE ~ NA_character_
+      )
     )
-  )
- }
+}
 ```
 
 Here is a basic example of how to use `ggpedigree` to visualize a pedigree structure. The `potter` dataset contains simulated pedigree data for the Weasley family from the Harry Potter series.

[…]
