### 2025-07-22 fd893ec
Add alignPedigreeWithRelations function and refactor
Introduces the alignPedigreeWithRelations function to handle pedigree alignment using additional relations if provided, and refactors calculateCoordinates to use this new function. Updates NEWS.md to reflect these changes and adds corresponding documentation for alignPedigreeWithRelations.
