+++ R/calcCoordinates.R
@@ -61,9 +61,22 @@ calculateCoordinates <- function(ped,
   )
   config <- utils::modifyList(default_config, config)
 
-  # Recode sex values in case non-standard codes are used (e.g., "M"/"F")
-  ped_recode <- BGmisc::recodeSex(ped, code_male = code_male)
+  # Construct a pedigree object to compute layout coordinates
 
+  # use relations if provided, otherwise use default settings
+  ped_ped <-  alignPedigreeWithRelations(ped = ped,
+                                         personID=personID,
+                                         dadID=dadID,
+                                         momID=momID,
+                                         code_male=code_male,
+                                         sexVar=sexVar,
+                                         config=config)
+
+  # use hints if provided
+  pos <- alignPedigreeWithHints(ped_ped = ped_ped,
+                                config = config)
+
+<<<<<<< Updated upstream
   if("relation" %in% names(config) && !is.null(config$relation)) {
   # Construct a pedigree object to compute layout coordinates
   ped_ped <- kinship2::pedigree(
@@ -82,6 +95,8 @@ calculateCoordinates <- function(ped,
   )
 }
   pos <- alignPedigreeWithHints(ped_ped=ped_ped, config=config)
+=======
+>>>>>>> Stashed changes
   #  assign("DEBUG_pos", pos, envir = .GlobalEnv)
 
   # Extract layout information
@@ -119,7 +134,12 @@ calculateCoordinates <- function(ped,
   y_coords <- nid_pos[, "row"]
   x_coords <- nid_pos[, "col"]
   x_pos <- pos$pos[cbind(y_coords, x_coords)]
+
+
+  # Spouse information
   spouse_vector <- pos$spouse[cbind(y_coords, x_coords)]
+
+  # Parent information
   parent_fam <- pos$fam[cbind(y_coords, x_coords)]
   y_fam <- y_coords - 1
 
@@ -127,12 +147,14 @@ calculateCoordinates <- function(ped,
   parent_col_left <- parent_fam
   parent_col_right <- parent_fam + 1
 
-  valid <- parent_row >= 1 &
+  # Ensure parent columns are within bounds
+  valid_parent <- parent_row >= 1 &
     parent_col_left >= 1 &
     parent_col_right <= ncol(pos$pos)
 
-  parent_left_vector[valid] <- pos$pos[cbind(parent_row[valid], parent_col_left[valid])]
-  parent_right_vector[valid] <- pos$pos[cbind(parent_row[valid], parent_col_right[valid])]
+  parent_left_vector[valid_parent] <- pos$pos[cbind(parent_row[valid_parent], parent_col_left[valid_parent])]
+
+  parent_right_vector[valid_parent] <- pos$pos[cbind(parent_row[valid_parent], parent_col_right[valid_parent])]
 
 # Vectorized above, maintained for now in case need to revert
 if(FALSE){
@@ -182,7 +204,8 @@ if(FALSE){
 
   # Create duplicate rows for extra appearances
   # For each duplicate nid
-  for (nid_val in duplicate_nids) {
+
+    for (nid_val in duplicate_nids) {
     # All appearance positions
     appearance_indices <- which(nid_vector == nid_val)
 
@@ -200,6 +223,7 @@ if(FALSE){
       )
     }
   }
+
   if (length(extra_info) > 0) {
     # Bind into single data.frame of all extras
     extra_df <- do.call(rbind, extra_info)
@@ -238,6 +262,7 @@ if(FALSE){
     ped_extra <- NULL
     ped$extra <- FALSE
   }
+  # clean up
[…]
