### 2025-07-22 89ef072
Update calcCoordinates.R
