+++ .github/workflows/draft-pdf.yaml
@@ -1,7 +1,8 @@
 on:
   push:
+    branches: [main]
     paths:
-      - '**.md'
+      - '*paper.md'
 jobs:
   paper:
     runs-on: ubuntu-latest
