### 2025-07-22 6d53042
final vectorization!?
