+++ R/calcCoordinates.R
@@ -139,21 +139,6 @@ calculateCoordinates <- function(ped,
 
   parent_right_vector[valid_parent] <- pos$pos[cbind(parent_row[valid_parent], parent_col_right[valid_parent])]
 
-  # Vectorized above, maintained for now in case need to revert
-  if (FALSE) {
-    # Populate coordinates from nid positions
-    for (i in seq_along(nid_vector)) {
-      y_coords[i] <- nid_pos[i, "row"]
-      x_coords[i] <- nid_pos[i, "col"]
-      x_pos[i] <- pos$pos[nid_pos[i, "row"], nid_pos[i, "col"]]
-      spouse_vector[i] <- pos$spouse[nid_pos[i, "row"], nid_pos[i, "col"]]
-      parent_fam[i] <- pos$fam[nid_pos[i, "row"], nid_pos[i, "col"]]
-      y_fam[i] <- BGmisc:::tryNA(nid_pos[i, "row"] - 1)
-      parent_left_vector[i] <- BGmisc:::tryNA(pos$pos[nid_pos[i, "row"] - 1, parent_fam[i] + 0])
-      parent_right_vector[i] <- BGmisc:::tryNA(pos$pos[nid_pos[i, "row"] - 1, parent_fam[i] + 1])
-    }
-  }
-
 
   # -----
   # Fill in the data frame with coordinates
@@ -176,6 +161,7 @@ calculateCoordinates <- function(ped,
 
   # Detect multiple layout positions for the same individual
   # This can happen if the same individual appears multiple times in the pedigree
+
   # For each nid, count how many times it appears
   appearance_counts <- table(nid_vector)
 
@@ -186,66 +172,54 @@ calculateCoordinates <- function(ped,
   extra_info <- list()
 
   # Create duplicate rows for extra appearances
-  # For each duplicate nid
 
-  for (nid_val in duplicate_nids) {
-    # All appearance positions
-    appearance_indices <- which(nid_vector == nid_val)
+  # All appearance positions
+  appearance_indices <- which(nid_vector %in% duplicate_nids)
 
-    # Find which index was already used in tmp
-    # tmp is a mapping from person index to nid_vector position
-    used_index <- tmp[which(seq_along(ped_ped$id) == nid_val)]
+  # Find which index was already used in tmp
+  # tmp is a mapping from person index to nid_vector position
+  used_indices <- tmp[duplicate_nids]
 
-    # Extra indices are the appearances NOT used by match()
-    extra_indices <- setdiff(appearance_indices, used_index)
+  # Extra indices are the appearances NOT used by match()
+  extra_indices <- setdiff(appearance_indices, used_indices)
 
-    if (length(extra_indices) > 0) {
-      extra_info[[as.character(nid_val)]] <- data.frame(
-        nid = nid_val,
-        idx = extra_indices
-      )
-    }
-  }
+  # If there are extra indices, we need to create additional rows
+  if (length(extra_indices) > 0) {
 
-  if (length(extra_info) > 0) {
-    # Bind into single data.frame of all extras
-    extra_df <- do.call(rbind, extra_info)
-    # Initialize list to collect extra rows
-    extra_rows <- vector("list", nrow(extra_df))
-    # Single loop over the flat index/nid map
+    extra_df <- data.frame(
+      nid = nid_vector[extra_indices],
+      idx = extra_indices
+    )
 
-    for (i in seq_len(nrow(extra_df))) {
-      idx <- extra_df$idx[i]
-      nid_val <- extra_df$nid[i]
[…]
