+++ vignettes/v00_plots.Rmd
@@ -138,12 +138,12 @@ ggPedigree(
   famID = "famID",
   personID = "personID",
   config = list(
-    label_column = "first_name",
+    # label_column = "first_name",
     sex_color_palette = c("pink", "blue"),
     label_text_angle = -35,
     label_nudge_y = .215,
     label_nudge_x = 0.45,
-    label_method = "geom_text",#"ggrepel", # 
+    label_method = "geom_text", # "ggrepel", #
     #   sex_color_palette = c("black", "black"),
     sex_color_include = TRUE
   )
