### 2025-07-23 ab5f735
fixed odd verbose behavior
