### 2025-07-25 61500d7
Refactor to avoid sapply in ggPedigreeInteractive
Replaced sapply() with vapply() in ggPedigreeInteractive for improved code safety and clarity. Minor formatting improvements in calcCoordinates.R for readability. Added goodpractice package usage in benchmarking script.
