### 2025-07-26 eb4ea30
add twins subfunction
