+++ R/calcConnectionsHelpers.R
@@ -252,8 +252,8 @@ getRelativeCoordinates <- function(ped,
 #' @return A string representing the symmetric key
 #' @keywords internal
 #' @importFrom dplyr if_else
-
-makeSymmetricKey <- function(id1, id2, sep = ".") {
+#' @aliases makeSymmetricKey
+.makeSymmetricKey <- function(id1, id2, sep = ".") {
   if (missing(id1) || missing(id2)) {
     stop("Both id1 and id2 must be provided.")
   }
@@ -286,3 +286,6 @@ makeSymmetricKey <- function(id1, id2, sep = ".") {
     paste0(id2, sep, id1)
   )
 }
+
+#' @rdname dot-makeSymmetricKey
+makeSymmetricKey <- .makeSymmetricKey
