### 2025-07-28 0eb0f0e
Increase default label and axis text sizes in plot config
Updated the default values for label_text_size from 3 to 4 and axis_text_size from 8 to 9 in getDefaultPlotConfig. Documentation was updated accordingly to reflect these changes.
