### 2025-07-28 a4a587e
add non-dot version to evaluate complexity
