+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: ggpedigree
 Title: Visualizing Pedigrees with 'ggplot2' and 'plotly'
-Version: 0.8.1.9000
+Version: 0.9
 Date/Publication: 2025
 Authors@R: c(
     person("S. Mason", "Garrison", email= "garrissm@wfu.edu", role = c("aut", "cre", "cph"),
