### 2025-08-01 3cd9f13
add log option
