### 2025-08-01 4710eab
added a couple of tests
