+++ R/ggpedigree.R
@@ -899,7 +899,8 @@ addSelfSegment <- .addSelfSegment
         c = config$focal_fill_chroma,
         l = config$focal_fill_lightness,
         direction = config$focal_fill_hue_direction,
-        na.value = config$focal_fill_na_value,
+        na.value = config$focal_fill_na_value#,
+      #  transform = ifelse(config$focal_fill_use_log,"log2","identity")
       )
     } else if (config$focal_fill_method %in% c("viridis_c")) {
       plotObject <- plotObject + ggplot2::scale_colour_viridis_c(
@@ -907,7 +908,8 @@ addSelfSegment <- .addSelfSegment
         begin = config$focal_fill_viridis_begin,
         end = config$focal_fill_viridis_end,
         direction = config$focal_fill_viridis_direction,
-        na.value = config$focal_fill_na_value
+        na.value = config$focal_fill_na_value,
+        transform = ifelse(config$focal_fill_use_log,"log2","identity")
       )
     } else if (config$focal_fill_method %in% c("viridis_d")) {
       plotObject <- plotObject + ggplot2::scale_colour_viridis_d(
@@ -915,7 +917,8 @@ addSelfSegment <- .addSelfSegment
         begin = config$focal_fill_viridis_begin,
         end = config$focal_fill_viridis_end,
         direction = config$focal_fill_viridis_direction,
-        na.value = config$focal_fill_na_value
+        na.value = config$focal_fill_na_value#,
+     #   transform = ifelse(config$focal_fill_use_log,"log2","identity")
       )
     } else if (config$focal_fill_method %in% c("viridis_b")) {
       plotObject <- plotObject + ggplot2::scale_colour_viridis_b(
@@ -923,7 +926,8 @@ addSelfSegment <- .addSelfSegment
         begin = config$focal_fill_viridis_begin,
         end = config$focal_fill_viridis_end,
         direction = config$focal_fill_viridis_direction,
-        na.value = config$focal_fill_na_value
+        na.value = config$focal_fill_na_value,
+        transform = ifelse(config$focal_fill_use_log,"log2","identity")
       )
     } else if (config$focal_fill_method %in% c("manual")) {
       plotObject <- plotObject + ggplot2::scale_color_manual(
