### 2025-08-01 f91f32c
add expect_error
