### 2025-08-04 6a2d55e
EXTREMELY ROUGH GGPROTO
