+++ tests/testthat/test-ggPedigree.R
@@ -15,9 +15,9 @@ test_that("broken hints doesn't cause a fatal error", {
   # Test with hints
   expect_warning(
     ggPedigree(potter,
-      famID = "famID",
-      personID = "personID",
-      config = list(hints = TRUE)
+               famID = "famID",
+               personID = "personID",
+               config = list(hints = TRUE)
     )
   ) %>% suppressWarnings()
 
@@ -43,18 +43,18 @@ test_that("broken hints doesn't cause a fatal error", {
     )
   expect_warning(
     ggPedigree(potter,
-      famID = "famID",
-      #  phantoms = TRUE, # not  in CRAN version
-      personID = "personID",
-      config = list(
-        hints = TRUE,
-        generation_width = 2,
-        generation_height = 2,
-        status_code_affected = "deceased",
-        status_code_unaffected = "alive",
-        status_include = TRUE
-      ),
-      status_column = "status"
+               famID = "famID",
+               #  phantoms = TRUE, # not  in CRAN version
+               personID = "personID",
+               config = list(
+                 hints = TRUE,
+                 generation_width = 2,
+                 generation_height = 2,
+                 status_code_affected = "deceased",
+                 status_code_unaffected = "alive",
+                 status_include = TRUE
+               ),
+               status_column = "status"
     )
   ) %>% suppressWarnings()
 })
@@ -74,8 +74,8 @@ test_that("ggPedigree returns a ggplot object", {
 
   # Test with hints
   p <- ggPedigree(potter,
-    famID = "famID",
-    personID = "personID"
+                  famID = "famID",
+                  personID = "personID"
   )
   expect_s3_class(p, "gg")
 
@@ -155,11 +155,11 @@ test_that("ggPedigree handles non-standard names", {
     )
 
   p <- ggPedigree(potter,
-    famID = "family_id",
-    personID = "individual_id",
-    momID = "mother_id",
-    dadID = "father_id",
-    spouseID = "spouse_id"
+                  famID = "family_id",
+                  personID = "individual_id",
+                  momID = "mother_id",
+                  dadID = "father_id",
+                  spouseID = "spouse_id"
   )
   expect_s3_class(p, "gg")
   expect_true(all(p$data$individual_id %in% potter$individual_id)) # ID retention
@@ -230,27 +230,27 @@ test_that("focal fill works with ID", {
   data("potter") # load example data from BGmisc
 
   p <- ggPedigree(potter,
-    famID = "famID",
-    personID = "personID",
-    config = list(
-      focal_fill_include = TRUE,
-      sex_color_include = FALSE,
-      focal_fill_personID = 1
-    )
+                  famID = "famID",
+                  personID = "personID",
+                  config = list(
+                    focal_fill_include = TRUE,
+                    sex_color_include = FALSE,
+                    focal_fill_personID = 1
+                  )
   )
   expect_s3_class(p, "gg") # Should return a ggplot object
   expect_true("focal_fill" %in% names(p$data)) # focal_fill column should be present
   expect_true(all(p$data$focal_fill >= 0 & p$data$focal_fill <= 1)) # focal_fill values should be between 0 and 1
 
   p2 <- ggPedigree(potter,
-    famID = "famID",
-    personID = "personID",
-    config = list(
-      focal_fill_include = TRUE,
-      sex_color_include = FALSE,
-      focal_fill_force_zero = TRUE,
-      focal_fill_personID = 1
-    )
+                   famID = "famID",
+                   personID = "personID",
+                   config = list(
+                     focal_fill_include = TRUE,
+                     sex_color_include = FALSE,
+                     focal_fill_force_zero = TRUE,
+                     focal_fill_personID = 1
+                   )
   )
   expect_s3_class(p2, "gg") # Should return a ggplot object
[…]
