### 2025-09-04 20c941b
Update bibliography with new article on mtDNA and longevity
Removed outdated tech report entry and added a new article entry regarding the contributions of inherited mtDNA to longevity.
