+++ vignettes/articles/paper.bib
@@ -364,13 +364,6 @@
   year         = {2025}
 }
 
-@techreport{burt2025mtDNA,
-  author = {{Burt}, S. A. and {Garrison}, S. M. and {Lyu}, X. and {Rodgers}, J. L. and {Carroll}, S. L. and {Smith}, Ken R. and {Hunter}, Michael D.},
-  title = {Inherited mtDNA contributes to longevity: Evidence from extended pedigrees with 176 million kinship pairs},
-  year = {2025},
-  journal      = {Revise and resubmit at \textit{eBioMedicine}}
-}
-
 @misc{mcfarlane2015,
 	title = {Data from: Very low levels of direct additive genetic variance in fitness and fitness components in a red squirrel population},
 	author = {{McFarlane}, S. Eryn and {Boutin}, Stan and {Humphries}, Murray M. and {McAdam}, Andrew G. and {Gorrell}, Jamieson C. and {Coltman}, David W.},
@@ -512,4 +505,27 @@ PMCID: PMC4133633}
 	doi = {10.1111/j.1755-0998.2009.02817.x},
 	note = {PMID: 21565076},
 	langid = {eng}
+}
+@article{burt2025mtDNA,
+title = {Contributions of inherited mtDNA to longevity: evidence from extended pedigrees with 176 million kinship pairs},
+journal = {eBioMedicine},
+volume = {119},
+pages = {105911},
+year = {2025},
+issn = {2352-3964},
+doi = {https://doi.org/10.1016/j.ebiom.2025.105911},
+url = {https://www.sciencedirect.com/science/article/pii/S235239642500355X},
+author = {S. Alexandra Burt and S. Mason Garrison and Xuanyu Lyu and Joseph L. Rodgers and Sarah L. Carroll and Ken R. Smith and Michael D. Hunter},
+keywords = {mtDNA, Longevity, Extended pedigrees, Matrilineal relatedness, Heritability},
+abstract = {Summary
+Background
+Mitochondria are bacteria-like organelles with their own DNA (mtDNA) that exist in the cellular cytoplasm of almost every cell in the human body. Because mitochondria are critical for sustaining life, it follows that inherited mtDNA could be a key aetiologic element underlying longevity. Unfortunately, biometric approaches able to quantify heritable contributions of mtDNA have not been available.
+Methods
+We directly leveraged the unique matrilineal inheritance pattern of mtDNA to estimate its effects on longevity (defined as the top 10% oldest survivors within their birth cohort). We employed the Utah Population Database (UPDB) to identify 176,348,110 unique kinship links amongst 1,018,929 individuals born between 1700 and 1925 with information on matrilineal versus patrilineal relatedness.
+Findings
+Across 1st, 2nd, 3rd, 4th, and 5th degree kin, matrilineal relatives were more similar in their longevity outcomes than were non-maternal relatives. Variance component analyses indicated nuclear DNA heritability of 23–26% and mtDNA heritability of at least 5% − despite mtDNA constituting only ∼16.6 k base pairs (versus 2,875,002 k base pairs for nuclear DNA). Moreover, sharing the maternal line of a longevous relative translated to an average of 11.3 months extra years of life.
+Interpretation
+Results collectively suggest that mtDNA may be an important element of unusually long lifespans.
+Funding
+This project was supported by RF1-AG073189 and R01-AG022095 from the National Institute on Aging (NIA). We also acknowledge partial support through grant P30-CA2014 from the National Cancer Institute, University of Utah, and from the University of Utah’s program in Personalized Health and Utah Clinical and Translational Science Institute.}
 }
