### 2025-09-04 e1b9f26
Clarify .computeCurvedMidpoint documentation
Expanded and clarified the documentation for .computeCurvedMidpoint in both R and Rd files. The new description explains the quadratic Bézier construction, parameter meanings, and return value, improving clarity for users and maintainers.
