### 2025-09-12 cc5b014
Add plotly optimization for pedigree plots
Introduced optimizePedigree and optimizePlotlyPedigree functions to reduce plotly object size by rounding coordinates. Added optimize_plotly and value_rounding_digits options to plot config, updated ggPedigreeInteractive to use these optimizations, and adjusted documentation and tests accordingly.

Update ggPedigreeInteractive.R
