+++ R/ggPedigreeInteractive.R
@@ -27,7 +27,8 @@ ggPedigreeInteractive <- function(ped,
                                   tooltip_columns = NULL,
                                   focal_fill_column = NULL,
                                   overlay_column = NULL,
-                                  config = list(),
+                                  config = list(optimize_plotly = TRUE
+                                    ),
                                   debug = FALSE,
                                   return_widget = TRUE,
                                   phantoms = FALSE,
@@ -108,6 +109,10 @@ ggPedigreeInteractive <- function(ped,
     stop("None of the specified tooltip_columns found in `ped`.")
   }
 
+#  if (config$optimize_static == TRUE) {
+#    static_plot <- optimizeStaticPedigree(static_plot, config=config)
+#  }
+   # assign("DEBUG_static_plot", static_plot, envir = .GlobalEnv)
 
   ## 3. Convert ggplot → plotly ---------------------------------------------
   #   Add the tooltip text to the data frame
@@ -120,7 +125,6 @@ ggPedigreeInteractive <- function(ped,
 
     if (length(point_layers) == 0L) {
       warnings("No GeomPoint layer found for tooltips.")
-
       static_plot <- static_plot + ggplot2::aes(text = formatTooltip(
         df = static_plot$data,
         config$tooltip_columns
@@ -137,11 +141,14 @@ ggPedigreeInteractive <- function(ped,
       }
     }
 
+  if (config$optimize_plotly == TRUE) {
+    static_plot <- optimizePedigree(static_plot, config=config, plot_type = "static")
+  }
     plt <- tryCatch(
       plotly::ggplotly(static_plot,
         tooltip = "text",
         width   = NULL,
-        height  = NULL
+        height  = NULL #, originalData = !config$optimize_plotly # retain original data for hover if not optimizing
       ),
       error = function(e) {
         warning("Error in ggplotly conversion: ", e$message)
@@ -150,11 +157,15 @@ ggPedigreeInteractive <- function(ped,
       }
     )
   } else {
+      if (config$optimize_plotly == TRUE) {
+    static_plot <- optimizePedigree(static_plot, config=config, plot_type = "static")
+  }
     plt <- tryCatch(
       plotly::ggplotly(static_plot,
         tooltip = NULL,
         width = NULL,
-        height = NULL
+        height = NULL ,
+        originalData = !config$optimize_plotly # retain original data for hover if not optimizing
       ),
       error = function(e) {
         warning("Error in ggplotly conversion: ", e$message)
@@ -163,6 +174,12 @@ ggPedigreeInteractive <- function(ped,
       }
     )
   }
+
+ #     assign("DEBUG_static_plot", static_plot, envir = .GlobalEnv)
+  if (config$optimize_plotly == TRUE) {
+    plt <- optimizePedigree(plt, config=config, plot_type = "plotly")
+  }
+
   if (config$return_static == TRUE) {
     return(static_plot) # return the static plot
   } else if (config$return_widget == TRUE) {
@@ -171,6 +188,7 @@ ggPedigreeInteractive <- function(ped,
     class(plt) <- c("plotly", class(plt)) # ensure proper S3 dispatch
     return(plt)
   }
+
 }
 
 
@@ -190,6 +208,97 @@ formatTooltip <- function(df, tooltip_columns, sep = ": ") {
   apply(df[tooltip_columns], 1, function(row) {
     paste(paste(tooltip_columns, row, sep = sep), collapse = "<br>")
   })
+}
+
+
+#' @title Optimize Pedigree Plot
+#' @description
+#' Optimize a pedigree plot by rounding coordinates to reduce file size.
+#' @param p A plotly or ggplot object representing the pedigree plot.
+#' @param config A list of configuration parameters, including `value_rounding_digits`.
[…]
