+++ man/optimizePlotlyPedigree.Rd
@@ -1,0 +1,11 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/ggPedigreeInteractive.R
+\name{optimizePlotlyPedigree}
+\alias{optimizePlotlyPedigree}
+\title{Optimize Plotly Pedigree Plot}
+\usage{
+optimizePlotlyPedigree(p, config = list())
+}
+\description{
+Optimize a Plotly pedigree plot by rounding coordinates to reduce file size.
+}

[…]
