### 2025-09-12 ac11a05
Add kinship2 internal and alignment source files
Introduced kinship2_internals.R and multiple kinshipsource R scripts for pedigree alignment and manipulation, including align.pedigree, alignped1-4, and related helpers. Also fixed a minor documentation formatting issue in ggpedigreeHelpers.R. These changes modularize and expose internal pedigree alignment logic for further development and integration.
