+++ R/kinshipsource/plot.pedigree.R
@@ -1,0 +1,373 @@
+
+plot.pedigree <- function(x, id = x$id, status = x$status, 
+                          affected = x$affected, 
+                          cex = 1, col = 1,
+                          symbolsize = 1, branch = 0.6, 
+                          packed = TRUE, align = c(1.5,2), width = 8, 
+                          density=c(-1, 35,65,20), #mar=c(3.1, 1, 3.1, 1),
+                          angle=c(90,65,40,0), keep.par=FALSE,
+                          subregion, pconnect=.5, ...)
+{
+    ## As of 2020-09, documention in no-web directory is moved to here and a vignette.
+    ## Relevant sections in the vignette are marked in this code with
+       ## Doc: followed by the section title
+    
+    Call <- match.call()
+    n <- length(x$id)
+    if(n < 3) {
+        stop("Cannot plot pedigree with fewer than 3 subjects")
+    }
+    ## Doc: This portion is documented as the setup-data
+    
+    if(is.null(status))
+      status <- rep(0, n)
+    else {
+        if(!all(status == 0 | status == 1))
+          stop("Invalid status code")
+        if(length(status) != n)
+          stop("Wrong length for status")
+    }
+    if(!missing(id)) {
+        if(length(id) != n)
+          stop("Wrong length for id")
+    }
+    ## Doc: still part of setup/data
+    ## affected is a 0/1 matrix of any marker data.  It may be attached to the pedigree or added
+    ## here.  It can be a vector of length [[n]] or a matrix with [[n]] rows.
+    ## If not present, the default is to plot open symbols without shading or color  
+
+    ## If affected is a matrix, then the shading and/or density value for ith column is
+    ## taken from the ith element of the angle/density arguments.
+
+    ## For purposes within the plot method, NA values in ``affected'' are
+    ## coded to -1, and plotted as a question mark (?) in the plot symbol
+    ## region for that affected status
+
+    if(is.null(affected)){
+      affected <- matrix(0,nrow=n)
+    }
+    else {
+        if (is.matrix(affected)){
+            if (nrow(affected) != n) stop("Wrong number of rows in affected")
+            if (is.logical(affected)) affected <- 1* affected
+            if (ncol(affected) > length(angle) || ncol(affected) > length(density))
+                stop("More columns in the affected matrix than angle/density values")
+            } 
+        else {
+            if (length(affected) != n)
+                stop("Wrong length for affected")
+
+            if (is.logical(affected)) affected <- as.numeric(affected)
+            if (is.factor(affected))  affected <- as.numeric(affected) -1
+            }
+        if(max(affected, na.rm=TRUE) > min(affected, na.rm=TRUE)) {
+          affected <- matrix(affected - min(affected, na.rm=TRUE),nrow=n)
+         ## affected[is.na(affected)] <- -1
+        } else {
+          affected <- matrix(affected,nrow=n)
+        }
+        ## JPS 4/28/17 bug fix b/c some cases NAs are not set to -1
+        affected[is.na(affected)] <- -1
+        if (!all(affected==0 | affected==1 | affected == -1))
+                stop("Invalid code for affected status")
+    }
+
+    if (length(col) ==1) col <- rep(col, n)
[…]
