### 2025-09-12 1dc4474
Update R/ggPedigreeInteractive.R
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
