### 2025-09-12 d9175ac
Refactor code formatting and indentation in R scripts
Improves code readability and consistency by adjusting indentation, spacing, and alignment across multiple R files. No functional changes were made; only formatting, comment alignment, and minor stylistic updates to function calls and argument lists.
