+++ R/ggPedigreeInteractive.R
@@ -27,8 +27,7 @@ ggPedigreeInteractive <- function(ped,
                                   tooltip_columns = NULL,
                                   focal_fill_column = NULL,
                                   overlay_column = NULL,
-                                  config = list(optimize_plotly = TRUE
-                                    ),
+                                  config = list(optimize_plotly = TRUE),
                                   debug = FALSE,
                                   return_widget = TRUE,
                                   phantoms = FALSE,
@@ -109,19 +108,18 @@ ggPedigreeInteractive <- function(ped,
     stop("None of the specified tooltip_columns found in `ped`.")
   }
 
-#  if (config$optimize_static == TRUE) {
-#    static_plot <- optimizeStaticPedigree(static_plot, config=config)
-#  }
-   # assign("DEBUG_static_plot", static_plot, envir = .GlobalEnv)
+  #  if (config$optimize_static == TRUE) {
+  #    static_plot <- optimizeStaticPedigree(static_plot, config=config)
+  #  }
+  # assign("DEBUG_static_plot", static_plot, envir = .GlobalEnv)
 
   ## 3. Convert ggplot → plotly ---------------------------------------------
   #   Add the tooltip text to the data frame
   if (config$tooltip_include == TRUE) {
     # add tooltips to geom_point layers
-    point_layers <- which(vapply(static_plot$layers, FUN= function(l) {
+    point_layers <- which(vapply(static_plot$layers, FUN = function(l) {
       inherits(l$geom, "GeomPoint")
-    }, FUN.VALUE = logical(1)
-    ))
+    }, FUN.VALUE = logical(1)))
 
     if (length(point_layers) == 0L) {
       warnings("No GeomPoint layer found for tooltips.")
@@ -141,14 +139,14 @@ ggPedigreeInteractive <- function(ped,
       }
     }
 
-  if (config$optimize_plotly == TRUE) {
-    static_plot <- optimizePedigree(static_plot, config=config, plot_type = "static")
-  }
+    if (config$optimize_plotly == TRUE) {
+      static_plot <- optimizePedigree(static_plot, config = config, plot_type = "static")
+    }
     plt <- tryCatch(
       plotly::ggplotly(static_plot,
         tooltip = "text",
         width   = NULL,
-        height  = NULL #, originalData = !config$optimize_plotly # retain original data for hover if not optimizing
+        height  = NULL # , originalData = !config$optimize_plotly # retain original data for hover if not optimizing
       ),
       error = function(e) {
         warning("Error in ggplotly conversion: ", e$message)
@@ -157,14 +155,14 @@ ggPedigreeInteractive <- function(ped,
       }
     )
   } else {
-      if (config$optimize_plotly == TRUE) {
-    static_plot <- optimizePedigree(static_plot, config=config, plot_type = "static")
-  }
+    if (config$optimize_plotly == TRUE) {
+      static_plot <- optimizePedigree(static_plot, config = config, plot_type = "static")
+    }
     plt <- tryCatch(
       plotly::ggplotly(static_plot,
         tooltip = NULL,
         width = NULL,
-        height = NULL ,
+        height = NULL,
         originalData = !config$optimize_plotly # retain original data for hover if not optimizing
       ),
       error = function(e) {
@@ -175,9 +173,9 @@ ggPedigreeInteractive <- function(ped,
     )
   }
 
- #     assign("DEBUG_static_plot", static_plot, envir = .GlobalEnv)
+  #     assign("DEBUG_static_plot", static_plot, envir = .GlobalEnv)
   if (config$optimize_plotly == TRUE) {
-    plt <- optimizePedigree(plt, config=config, plot_type = "plotly")
[…]
