+++ README.md
@@ -84,9 +84,8 @@ cite the following:
     citation(package = "ggpedigree")
 
 Garrison S (2025). *ggpedigree: Visualizing Pedigrees with ‘ggplot2’ and
-‘plotly’*. R package version 0.8.0.1, commit
-d9a268ce50bb5318c1c9187e3362a0a606004808,
-<https://github.com/R-Computing-Lab/ggpedigree>.
+‘plotly’*. R package version 0.9.0,
+<https://github.com/R-Computing-Lab/ggpedigree/>.
 
 A BibTeX entry for LaTeX users is
 
@@ -94,8 +93,8 @@ A BibTeX entry for LaTeX users is
       title = {ggpedigree: Visualizing Pedigrees with 'ggplot2' and 'plotly'},
       author = {S. Mason Garrison},
       year = {2025},
-      note = {R package version 0.8.0.1, commit d9a268ce50bb5318c1c9187e3362a0a606004808},
-      url = {https://github.com/R-Computing-Lab/ggpedigree},
+      note = {R package version 0.9.0},
+      url = {https://github.com/R-Computing-Lab/ggpedigree/},
     }
 
 ## Contributing
