### 2025-10-14 7280583
Add smart scale transformation for relatedness matrix
- Add color_scale_trans parameter to control scale transformation
- Default to sqrt transformation for ggRelatednessMatrix for better visual distinction
- Support identity (linear), sqrt, log, and other transformations
- Add tests to verify transformation behavior

Co-authored-by: smasongarrison <6001608+smasongarrison@users.noreply.github.com>
