### 2025-10-14 0e42142
Update documentation and add examples for scale transformation
Co-authored-by: smasongarrison <6001608+smasongarrison@users.noreply.github.com>
