+++ R/ggRelatednessMatrix.R
@@ -15,6 +15,7 @@
 #'  \describe{
 #'   \item{matrix_color_palette}{A vector of colors for the heatmap (default: Reds scale)}
 #'   \item{color_scale_midpoint}{Numeric midpoint for diverging color scale (default: 0.25)}
+#'   \item{color_scale_trans}{Transformation applied to color scale. Default is "sqrt" for better visual distinction of genetic relatedness levels. Use "identity" for linear scale, or "log" for logarithmic.}
 #'   \item{plot_title}{Plot title}
 #'   \item{matrix_cluster}{Logical; should rows/cols be clustered (default: TRUE)}
 #'   \item{axis_x_label, axis_y_label}{Axis labels}
@@ -39,6 +40,23 @@
 #'     axis_x_label = "Individuals",
 #'     axis_y_label = "Individuals",
 #'     axis_text_size = 8
+#'   )
+#' )
+#'
+#' # Plot with different scale transformations
+#' # Sqrt (default) - better for genetic relatedness
+#' ggRelatednessMatrix(mat,
+#'   config = list(
+#'     color_scale_trans = "sqrt",
+#'     plot_title = "Relatedness Matrix (sqrt scale)"
+#'   )
+#' )
+#'
+#' # Linear scale
+#' ggRelatednessMatrix(mat,
+#'   config = list(
+#'     color_scale_trans = "identity",
+#'     plot_title = "Relatedness Matrix (linear scale)"
 #'   )
 #' )
 ggRelatednessMatrix <- function(
