### 2025-10-14 e7ce12b
Update NEWS.md with scale transformation feature
Co-authored-by: smasongarrison <6001608+smasongarrison@users.noreply.github.com>
