### 2025-10-18 de9553a
Refactor to use consistent naming and apply to all gradient scales
- Replace color_scale_trans with tile_color_scale_trans (consistent naming)
- Add focal_fill_trans parameter to replace focal_fill_use_log
- Apply sqrt transformation to both ggpedigree and ggRelatednessMatrix
- Maintain backward compatibility with focal_fill_use_log
- Add comprehensive tests for scale transformations

Co-authored-by: smasongarrison <6001608+smasongarrison@users.noreply.github.com>
