### 2025-11-03 2d82593
consolidate 2 and 6 by moving around
