+++ vignettes/articles/paper.md
@@ -27,7 +27,7 @@ affiliations:
 citation_author: Garrison
 csl: apa.csl
 journal: JOSS
-date: "12 September, 2025"
+date: "03 November, 2025"
 bibliography: paper.bib
 vignette: >
   %\VignetteEncoding{UTF-8}
@@ -36,214 +36,198 @@ vignette: >
   %\VignetteEngine{knitr::rmarkdown}
 ---
 
-
-
-
 # Summary
 
-
-Pedigree diagrams underpin research and practice across genetics, animal breeding, genealogy, forensics, and counseling. They help medical geneticists trace the inheritance of Mendelian diseases and identify at-risk relatives; enable dairy breeders to plan matings that improve milk yield; support genealogists in reconstructing ancestry; assist forensic scientists in establishing familial connections in criminal investigations; and facilitate family therapists and counselors in understanding their clients' relationships [@genograms_2020]. Early R tools such as kinship2 [@kinship2] plot simple nuclear families effectively, but they do not scale to today’s pedigrees that can exceed 1,000s of individuals. As datasets have grown, researchers now work with increasingly complex family structures, including large-scale plant breeding pedigrees [@shaw2014], web-based pedigree management systems [@Ranaweera2018open], interactive pedigree editors [@pedigreejs], and behavior genetic studies of extended family structures [@hunter_analytic_2021; @garrison_analyzing_2023]. That complexity exposes the limitations of existing tools, which often struggle with large and complex datasets. `ggpedigree` addresses this need by combining a vectorised layout algorithm, ggplot2 output, and optional plotly interactivity.
+Pedigree diagrams underpin research and practice across genetics, animal
+breeding, genealogy, forensics, and counseling. They help medical
+geneticists trace the inheritance of Mendelian diseases and identify
+at-risk relatives; enable dairy breeders to plan matings that improve
+milk yield; support genealogists in reconstructing ancestry; assist
+forensic scientists in establishing familial connections in criminal
+investigations; and facilitate family therapists and counselors in
+understanding their clients' relationships [@genograms_2020]. Early R
+tools such as kinship2 [@kinship2] plot simple nuclear families
+effectively, but they do not scale to today's pedigrees that can exceed
+1,000s of individuals. As datasets have grown, researchers now work with
+increasingly complex family structures, including large-scale plant
+breeding pedigrees [@shaw2014], web-based pedigree management systems
+[@Ranaweera2018open], interactive pedigree editors [@pedigreejs], and
+behavior genetic studies of extended family structures
+[@hunter_analytic_2021; @garrison_analyzing_2023]. That complexity
+exposes the limitations of existing tools, which often struggle with
+large and complex datasets. `ggpedigree` addresses this need by
+combining a vectorised layout algorithm, ggplot2 output, and optional
+plotly interactivity.
 
 # Statement of need
 
+Pedigree visualization has traditionally relied on proprietary software
+(e.g., Progeny, GenoPro, Pedigree Viewer) or R packages like `kinship2`
+[@kinship2], pedtools [@pedtools], or `pedtricks` [@pedtricks]. While
+these tools are functional for many use cases, their limitations become
+pronounced when working with complex, modern pedigree datasets or when
+more detailed customization is required. Most R packages focus on base
+graphics or simple `ggplot2` implementations.
 
[…]
