### 2025-11-03 65282e8
Update package references and remove log file
Replaced plain package names with code formatting in _paper.Rmd and paper.md for consistency. Deleted the obsolete paper.log file.
