+++ vignettes/articles/_paper.Rmd
@@ -54,11 +54,11 @@ Pedigree visualization has traditionally relied on proprietary software (e.g., P
 
 Existing R solutions face three main challenges. First, current solutions are often poorly integrated with tidyverse workflows and do not expose the full theming and layering capabilities familiar to `ggplot2` users [@wickham_ggplot2_2016]. In animal-focused workflows, rapid rendering seems to takes precedence over aesthetic flexibility. I suspect that this is because users tend to work with more uniform data and fewer phenotypes. By contrast, human-focused workflows—particularly in behavior genetics and genetic epidemiology [@garrison_analyzing_2023; @lyu2025; @mcardleRAM]—require closer integration with tidyverse pipelines and more flexible plotting systems to accommodate complex pedigree structures and harmonization of phenotypes across data sources. In other words, the needs are different.
 
-Second, most R-based tools offer no interactivity. Static graphics are often sufficient for publication, but interactivity improves exploration and communication during model development or data cleaning. A notable exception is pedtools [@vigeland2021], which offers a sister shiny app, QuickPed [@vigeland2022]. While the R ecosystem includes libraries, like plotly, that support interactive plotting, these features have yet to be integrated into pedigree functions.
+Second, most R-based tools offer no interactivity. Static graphics are often sufficient for publication, but interactivity improves exploration and communication during model development or data cleaning. A notable exception is `pedtools` [@vigeland2021], which offers a sister shiny app, `QuickPed` [@vigeland2022]. While the R ecosystem includes libraries, like `plotly`, that support interactive plotting, these features have yet to be integrated into pedigree functions.
 
 Third, scalability and extensibility remain limited across existing tools. Several R packages attempt to address these challenges with built-in pedigree plotting functions. `kinship2` [@kinship2] remains widely used but produces static base graphics and relies on non-vectorized recursive layout functions that do not scale well to large families. A partial `ggplot2` implementation exists in a modernized `kinship2` [called Pedixplorer, @pedixplorer], but is non-vectorized and incompatible with other ggplot2 layers. `pedtricks`, a revival of `pedantics` [@morrissey2010], provides a ggplot2-based implementation for large animal pedigrees but lacks extensibility and interactivity. The `geneHapR` [@Zhang2023] package focuses on haplotype visualization rather than general pedigree structure. The `pedgene` package [@pedgene] offers some plotting functions but is primarily designed for association testing. The `pedigreejs` package [@pedigreejs] provides an interactive pedigree editor but does not integrate with R or ggplot2, limiting its utility for R users.
 
-None of these packages offers the combination of modern ggplot2 integration, interactive capabilities, and extensibility that ggpedigree provides. ggpedigree addresses these limitations by providing a comprehensive visualization framework built on modern R graphics infrastructure. It leverages the extensive customization capabilities of ggplot2 while adding specialized functionality for pedigree-specific visualization challenges.
+None of these packages offers the combination of modern `ggplot2` integration, interactive capabilities, and extensibility that `ggpedigree` provides. `ggpedigree` addresses these limitations by providing a comprehensive visualization framework built on modern R graphics infrastructure. It leverages the extensive customization capabilities of `ggplot2` while adding specialized functionality for pedigree-specific visualization challenges.
 
 ## Software Architecture
 
