+++ vignettes/articles/_paper.Rmd
@@ -91,9 +91,9 @@ ggPedigree(potter,
 )
 ```
 
-\includegraphics[width=0.95\textwidth,keepaspectratio]{potter_pedigree.png}
+\includegraphics[width=0.55\textwidth,keepaspectratio]{potter_pedigree.png}
 
-I demonstrate several advanced features by restyling the figure from [@hunter2025tracing],  to follow the Wake Forest brand identity guidelines; source code is in the appendix.
+I demonstrate several advanced features by restyling the figure from [@hunter2025tracing], to follow the Wake Forest brand identity guidelines; source code is in the appendix.
 
 \includegraphics[width=0.95\textwidth,keepaspectratio]{wfu_potter_pedigree.png}
 
