### 2025-11-03 0eb36cb
Clarify and expand feature descriptions in paper
Improved wording and detail in the feature section of the ggpedigree package documentation. Expanded explanation of customization capabilities and provided a link to the full example source code.
