+++ vignettes/articles/v01_plots_extended.Rmd
@@ -255,9 +255,6 @@ data("inbreeding")
 df <- inbreeding # multigenerational pedigree with consanguinity
 
 
-# df  <- dplyr::filter(df, famID %in% c(5, 7))
-
-
 p <- ggPedigree(
   df,
   famID = "famID",
@@ -286,153 +283,135 @@ p + facet_wrap(~famID, scales = "free") #+ scale_color_viridis(
 
 ## Potter Pedigree Plots
 
-These figures are used in the manuscript to illustrate the use of the `ggpedigree` package for plotting complex pedigrees. The first figure shows a simplified version of the Potter family pedigree, while the second figure includes additional family members and relationships. 
+
+The package supports extensive customization of visual aesthetics. The following 
+example is a figure from Hunter et al that used the Potter pedigree data. The figure has been restyled according to Wake Forest  University brand identity guidelines to demonstrate ggpedigree's customization capabilities. The figure combines two panels:  panel (a) highlights unique mitochondrial lines in the Dursley and Evans families, while panel (b) shows the full pedigree with Molly Weasley’s mitochondrial descendants in gold. 
+
 
 ```{r}
-library(BGmisc)
 library(ggpedigree)
+library(BGmisc) # helper utilities & example data
 library(tidyverse)
-library(patchwork) # for combining plots
+library(showtext)
+library(sysfonts)
+library(patchwork) # combining plots
 data("potter") # load the potter pedigree data
 
-df_potter <- potter %>%
-  mutate(
-    name = case_when(
-      personID == 1 ~ "Vernon",
-      personID == 2 ~ "Marjorie",
-      personID == 3 ~ "Petunia",
-      personID == 4 ~ "Lily",
-      personID == 5 ~ "James",
-      personID == 6 ~ "Dudley",
-      personID == 7 ~ "Harry",
-      personID == 8 ~ "Ginny",
-      personID == 9 ~ "Arthur",
-      personID == 10 ~ "Molly",
-      personID == 11 ~ "Ron",
-      personID == 12 ~ "Fred",
-      personID == 13 ~ "George",
-      personID == 14 ~ "Percy",
-      personID == 15 ~ "Charlie",
-      personID == 16 ~ "Bill",
-      personID == 17 ~ "Hermione",
-      personID == 18 ~ "Fleur",
-      personID == 19 ~ "Gabrielle",
-      personID == 20 ~ "Audrey",
-      personID == 21 ~ "James",
-      personID == 22 ~ "Albus",
-      personID == 23 ~ "Lily",
-      personID == 24 ~ "Rose",
-      personID == 25 ~ "Hugo",
-      personID == 26 ~ "Victoire",
-      personID == 27 ~ "Dominique",
-      personID == 28 ~ "Louis",
-      personID == 29 ~ "Molly",
-      personID == 30 ~ "Lucy",
-      personID == 101 ~ "Mother",
-      personID == 102 ~ "Father",
-      personID == 103 ~ "Mother",
-      personID == 104 ~ "Father",
-      personID == 105 ~ "Mother",
-      personID == 106 ~ "Father"
-    )
-  )
+# Load Google fonts for styling
+# Load Google fonts for styling
+font_add_google(name = "Cormorant", family = "cormorant")
+showtext_auto() # render Google fonts
 
-m1 <- ggPedigree(df_potter %>% filter(personID %in% c(1:7, 101:104)),
+# Set WFU style guidelines
+text_color_wfu <- "#222222"
+focal_fill_color_values_wfu <- c(
[…]
