+++ vignettes/articles/_paper.Rmd
@@ -102,7 +102,7 @@ example is a figure from @hunter2025tracing that used the Potter pedigree data. 
 \includegraphics[width=0.95\textwidth,keepaspectratio]{wfu_potter_pedigree.png}
 
 The complete source code for this example is available in the package documentation at 
-<https://r-computing-lab.github.io/ggpedigree/articles/>.
+<https://r-computing-lab.github.io/ggpedigree/articles/v01_plots_extended.html>.
 
 Collectively, these tools provide a valuable resource for behavior geneticists and others who work with extended family data. They were developed as part of a grant and have been used in several ongoing projects, presentations [@garrison2024; @hunter2024], and forthcoming papers [@lyu2025; @hunter2025tracing; @burt2025mtDNA].
 
