+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: ggpedigree
 Title: Visualizing Pedigrees with 'ggplot2' and 'plotly'
-Version: 0.9.0
+Version: 0.9.1
 Date/Publication: 2025
 Authors@R: c(
     person("S. Mason", "Garrison", email= "garrissm@wfu.edu", role = c("aut", "cre", "cph"),
@@ -41,6 +41,8 @@ Suggests:
     rmarkdown,
     tibble,
     corrplot,
+    showtext,
+    sysfonts,
     readxl,
     htmlwidgets,
     testthat (>= 3.0.0)
