### 2025-11-03 0e926fe
Add workflow_dispatch trigger to test coverage workflow
Add workflow_dispatch trigger to draft-pdf.yaml
