+++ .github/workflows/test-coverage.yaml
@@ -5,7 +5,7 @@ on:
     branches: [main, master,dev_main,dev]
   pull_request:
     branches: [main, master,dev_main,dev]
-
+  workflow_dispatch:
 name: test-coverage
 
 jobs:
