### 2025-11-03 9374122
Add comprehensive tests for buildSpouseSegments function
Co-authored-by: smasongarrison <6001608+smasongarrison@users.noreply.github.com>
