+++ tests/testthat/test-calcConnections.R
@@ -271,3 +271,219 @@ test_that("calculateConnections resolves 'extra' rows properly", {
   expect_equal(sum(conns$personID == "A"), 1)
   expect_equal(sum(conns$personID == "A_2"), 1)
 })
+
+# Tests for buildSpouseSegments function
+test_that("buildSpouseSegments with use_hash=TRUE creates correct segments", {
+  # Create test data with parent_hash
+  ped <- data.frame(
+    personID = c("A", "B", "C"),
+    parent_hash = c("A.B", "A.B", NA),
+    stringsAsFactors = FALSE
+  )
+  
+  connections_for_FOO <- data.frame(
+    personID = c("A", "B", "C"),
+    x_pos = c(1, 3, 2),
+    y_pos = c(1, 1, 2),
+    stringsAsFactors = FALSE
+  )
+  
+  result <- ggpedigree:::buildSpouseSegments(ped, connections_for_FOO, use_hash = TRUE)
+  
+  expect_true(is.data.frame(result))
+  expect_true("x_start" %in% names(result))
+  expect_true("y_start" %in% names(result))
+  expect_true("x_end" %in% names(result))
+  expect_true("y_end" %in% names(result))
+  
+  # Check that segments were created for parent_hash entries
+  result_with_coords <- result[!is.na(result$x_start) & !is.na(result$x_end), ]
+  expect_true(nrow(result_with_coords) > 0)
+})
+
+test_that("buildSpouseSegments with use_hash=FALSE creates correct segments", {
+  # Create test data with spouseID
+  ped <- data.frame(
+    personID = c("A", "B", "C"),
+    spouseID = c("B", "A", NA),
+    x_pos = c(1, 3, 2),
+    y_pos = c(1, 1, 2),
+    stringsAsFactors = FALSE
+  )
+  
+  connections_for_FOO <- data.frame(
+    personID = c("A", "B", "C"),
+    x_pos = c(1, 3, 2),
+    y_pos = c(1, 1, 2),
+    stringsAsFactors = FALSE
+  )
+  
+  result <- ggpedigree:::buildSpouseSegments(ped, connections_for_FOO, use_hash = FALSE)
+  
+  expect_true(is.data.frame(result))
+  expect_true("x_start" %in% names(result))
+  expect_true("y_start" %in% names(result))
+  expect_true("x_end" %in% names(result))
+  expect_true("y_end" %in% names(result))
+  
+  # Check that segments were created for spouses
+  expect_equal(nrow(result), 2) # Only A and B have spouses
+})
+
+test_that("buildSpouseSegments with use_hash=FALSE filters out NA spouseID", {
+  # Create test data where some have NA spouseID
+  ped <- data.frame(
+    personID = c("A", "B", "C", "D"),
+    spouseID = c("B", "A", NA, NA),
+    x_pos = c(1, 3, 2, 4),
+    y_pos = c(1, 1, 2, 2),
+    stringsAsFactors = FALSE
+  )
+  
+  connections_for_FOO <- data.frame(
+    personID = c("A", "B", "C", "D"),
+    x_pos = c(1, 3, 2, 4),
+    y_pos = c(1, 1, 2, 2),
+    stringsAsFactors = FALSE
+  )
+  
+  result <- ggpedigree:::buildSpouseSegments(ped, connections_for_FOO, use_hash = FALSE)
+  
+  # Only A and B should have segments (they're each other's spouses)
+  expect_equal(nrow(result), 2)
+  expect_true(all(result$personID %in% c("A", "B")))
+})
+
+test_that("buildSpouseSegments with use_hash=TRUE handles NA parent_hash", {
[…]
