### 2025-11-03 720ae47
Add edge case tests for unmatched IDs in buildSpouseSegments
Co-authored-by: smasongarrison <6001608+smasongarrison@users.noreply.github.com>
