### 2025-11-04 8869e65
ensure reproducibility of the figure
tweak font size for the vignette

Update v01_plots_extended.Rmd
