### 2025-11-04 ff543da
Fix buildSpouseSegments tests - add spouseID to connections_for_FOO
Co-authored-by: smasongarrison <6001608+smasongarrison@users.noreply.github.com>
