### 2025-11-04 1533acb
Add tests for buildSpouseSegments and update docs
Added comprehensive test coverage for buildSpouseSegments, including cases for use_hash and spouseID logic, NA handling, and intermediate column removal. Updated documentation in R and Rd files to clarify the connections_for_FOO parameter. NEWS.md updated to reflect new test coverage.
