+++ .github/workflows/test-coverage.yaml
@@ -3,7 +3,7 @@
 on:
   push:
     branches: [main, master,dev_main,dev]
-  pull_request:
+  pull_request_target:
     branches: [main, master,dev_main,dev]
   workflow_dispatch:
 name: test-coverage
@@ -15,7 +15,7 @@ jobs:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
 
     steps:
-      - uses: actions/checkout@v4
+      - uses: actions/checkout@v5
 
       - uses: r-lib/actions/setup-r@v2
         with:
@@ -44,7 +44,7 @@ jobs:
 
       - name: Upload test results
         if: failure()
-        uses: actions/upload-artifact@v4
+        uses: actions/upload-artifact@v5
         with:
           name: coverage-test-failures
           path: ${{ runner.temp }}/package
