### 2025-11-04 #8624 @smasongarrison
@csoneson I was able to trim down the word count to 1334 by consolidating the statement of need a bit, as well as directing readers to the external documentation. Thanks for the suggestions.
