### 2025-11-04 6e98097
Add Codecov upload step to test coverage workflow
