### 2025-11-05 6db86c2
Add pedigree object implementation and internals
Introduces kinship2_pedigree and related internal functions for pedigree and pedigreeList object creation, subsetting, and printing to prepare for making kin2 alignments internal Updates NAMESPACE to export S3 methods for pedigree classes. Refactors internal functions to use kinship2_* naming and replaces calls to kinship2::pedigree with kinship2_pedigree. Adds documentation for pedigree objects.
