+++ NAMESPACE
@@ -1,5 +1,9 @@
 # Generated by roxygen2: do not edit by hand
 
+S3method("[",pedigree)
+S3method("[",pedigreeList)
+S3method(print,pedigree)
+S3method(print,pedigreeList)
 export(calculateConnections)
 export(calculateCoordinates)
 export(countOffspring)

[…]
