### 2025-11-06 #8624 @csoneson
👋🏻 @schochastics, @mkerin, @LouisLeNezet - would you be interested in reviewing this submission for the Journal of Open Source Software (JOSS)?

> ggpedigree: Visualizing Pedigrees with 'ggplot2' and 'plotly'
#8624

We carry out our checklist-driven reviews on GitHub, more details about the process are available [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html)

Thanks in advance!
