### 2025-11-07 f39996c
Update kinship2_check.hint.R
