+++ R/kinship2_check.hint.R
@@ -7,6 +7,19 @@
 #after the recent rewrite.
 #Users can introduce problems as well if they modify the hints.
 
+#' @title Check kinship2 hints for consistency
+#'
+#' @description
+#' This function checks the consistency of kinship2 hints, particularly
+#' focusing on the `order` and `spouse` components. It ensures that the
+#' order is numeric and matches the length of the `sex` vector, and that
+#' marriages are valid male/female pairs without duplicates.
+#' @param hints A list containing kinship2 hints, including `order` and optionally `spouse`.
+#' @param sex A character vector indicating the sex of each individual
+#' ('male' or 'female').
+#' @return The original `hints` list if all checks pass; otherwise, an error is raised.
+
+
 kinship2_check.hint <- function(hints, sex) {
     if (is.null(hints$order)) stop("Missing order component")
     if (!is.numeric(hints$order)) stop("Invalid order component")
@@ -14,17 +27,18 @@ kinship2_check.hint <- function(hints, sex) {
     if (length(hints$order) != n) stop("Wrong length for order component")
 
     spouse <- hints$spouse
-    if (is.null(spouse)) hints
-    else {
+    if (is.null(spouse)) {
+      hints
+    } else {
         lspouse <- spouse[,1]
         rspouse <- spouse[,2]
         if (any(lspouse <1 | lspouse >n | rspouse <1 | rspouse > n))
-            stop("Invalid spouse value")
+          warning("Invalid spouse value")
 
         temp1 <- (sex[lspouse]== 'female' & sex[rspouse]=='male')
         temp2 <- (sex[rspouse]== 'female' & sex[lspouse]=='male')
         if (!all(temp1 | temp2))
-            stop("A marriage is not male/female")
+            warning("A marriage is not male/female")
 
         hash <- n*pmax(lspouse, rspouse) + pmin(lspouse, rspouse)
         #Turn off this check for now - is set off if someone is married to two siblings
