### 2025-11-07 92fe7bc
Update NEWS.md
