### 2025-11-07 576a97f
Add roxygen2 documentation for all kinship2 functions
Co-authored-by: smasongarrison <6001608+smasongarrison@users.noreply.github.com>
