+++ R/kinship2_alignped4.R
@@ -1,4 +1,21 @@
 # Automatically generated from all.nw using noweb
+
+#' Compute optimal horizontal spacing for pedigree alignment
+#'
+#' This is an internal helper function for pedigree alignment. It uses quadratic
+#' programming to find optimal horizontal positions for subjects that minimize
+#' the distance between parents and children while keeping spouses together and
+#' respecting spacing constraints. Requires the quadprog package.
+#'
+#' @param rval Aligned pedigree structure from previous alignment steps
+#' @param spouse Logical matrix indicating spouse connections
+#' @param level Integer vector of generation levels
+#' @param width Numeric, maximum width of the pedigree plot
+#' @param align Logical or numeric vector. If logical, uses default alignment parameters.
+#'   If numeric, should be a vector c(a1, a2) where a1 controls parent-child penalties
+#'   and a2 controls spouse penalties
+#' @return Matrix of optimized horizontal positions for each subject
+#' @keywords internal
 kinship2_alignped4 <- function(rval, spouse, level, width, align) {
 
     ## Doc: alignped4 -part1, spacing across page
