### 2025-11-07 c6ea94f
Fix typo in kinship2_bitSize error message
Co-authored-by: smasongarrison <6001608+smasongarrison@users.noreply.github.com>
