### 2025-11-12 #8624 @flyaflya
> 👋🏻 [@flyaflya](https://github.com/flyaflya), [@magnusdv](https://github.com/magnusdv), [@mkerin](https://github.com/mkerin), [@LouisLeNezet](https://github.com/LouisLeNezet) - would you be interested in reviewing this submission for the Journal of Open Source Software (JOSS)?
>
> > ggpedigree: Visualizing Pedigrees with 'ggplot2' and 'plotly'
> > [#8624](https://github.com/openjournals/joss-reviews/issues/8624)
>
> We carry out our checklist-driven reviews on GitHub, more details about the process are available [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html)
>
> Thanks in advance!

I can get this done between Dec. 11th and Dec 15th.  If that timeline works, I'd be happy to review this work.  Looks like a nice package.
