### 2025-11-16 #9434 @editorialbot
[REVIEW]: ggpedigree: Visualizing Pedigrees with 'ggplot2' and 'plotly'
**Submitting author:** @smasongarrison (<a href="http://orcid.org/0000-0002-4804-6003">S. Mason Garrison</a>)
**Repository:** https://github.com/R-Computing-Lab/ggpedigree
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.0.3
