+++ vignettes/articles/paper.bib
@@ -513,22 +513,11 @@ volume = {119},
 pages = {105911},
 year = {2025},
 issn = {2352-3964},
-doi = {https://doi.org/10.1016/j.ebiom.2025.105911},
+doi = {10.1016/j.ebiom.2025.105911},
 url = {https://www.sciencedirect.com/science/article/pii/S235239642500355X},
 author = {S. Alexandra Burt and S. Mason Garrison and Xuanyu Lyu and Joseph L. Rodgers and Sarah L. Carroll and Ken R. Smith and Michael D. Hunter},
 keywords = {mtDNA, Longevity, Extended pedigrees, Matrilineal relatedness, Heritability},
-abstract = {Summary
-Background
-Mitochondria are bacteria-like organelles with their own DNA (mtDNA) that exist in the cellular cytoplasm of almost every cell in the human body. Because mitochondria are critical for sustaining life, it follows that inherited mtDNA could be a key aetiologic element underlying longevity. Unfortunately, biometric approaches able to quantify heritable contributions of mtDNA have not been available.
-Methods
-We directly leveraged the unique matrilineal inheritance pattern of mtDNA to estimate its effects on longevity (defined as the top 10% oldest survivors within their birth cohort). We employed the Utah Population Database (UPDB) to identify 176,348,110 unique kinship links amongst 1,018,929 individuals born between 1700 and 1925 with information on matrilineal versus patrilineal relatedness.
-Findings
-Across 1st, 2nd, 3rd, 4th, and 5th degree kin, matrilineal relatives were more similar in their longevity outcomes than were non-maternal relatives. Variance component analyses indicated nuclear DNA heritability of 23–26% and mtDNA heritability of at least 5% − despite mtDNA constituting only ∼16.6 k base pairs (versus 2,875,002 k base pairs for nuclear DNA). Moreover, sharing the maternal line of a longevous relative translated to an average of 11.3 months extra years of life.
-Interpretation
-Results collectively suggest that mtDNA may be an important element of unusually long lifespans.
-Funding
-This project was supported by RF1-AG073189 and R01-AG022095 from the National Institute on Aging (NIA). We also acknowledge partial support through grant P30-CA2014 from the National Cancer Institute, University of Utah, and from the University of Utah’s program in Personalized Health and Utah Clinical and Translational Science Institute.}
-}
+abstract = {}
 @article{pedixplorer,
     author = {Le Nézet, Louis and Sinnwell, Jason and Letko, Anna and André, Catherine and Quignon, Pascale},
     title = {Pedixplorer: a Bioconductor package to streamline pedigree design and visualization},
