### 2025-11-16 #9434 @csoneson
@smasongarrison - could you check the invalid DOI detected above? It seemed that the DOIs were fixed in the pre-review issue, was there another reference added or is this a regression?
