+++ reference/optimizePlotlyPedigree.md
@@ -1,0 +1,25 @@
+# Optimize Plotly Pedigree Plot
+
+Optimize a Plotly pedigree plot by rounding coordinates to reduce file
+size.
+
+## Usage
+
+``` r
+optimizePlotlyPedigree(p, config = list())
+```
+
+## Arguments
+
+- p:
+
+  A plotly object representing the pedigree plot.
+
+- config:
+
+  A list of configuration parameters, including
+  \`value_rounding_digits\`.
+
+## Value
+
+The optimized plotly object with rounded coordinates.
