+++ vignettes/articles/paper.bib
@@ -518,6 +518,7 @@ url = {https://www.sciencedirect.com/science/article/pii/S235239642500355X},
 author = {S. Alexandra Burt and S. Mason Garrison and Xuanyu Lyu and Joseph L. Rodgers and Sarah L. Carroll and Ken R. Smith and Michael D. Hunter},
 keywords = {mtDNA, Longevity, Extended pedigrees, Matrilineal relatedness, Heritability},
 abstract = {}
+}
 @article{pedixplorer,
     author = {Le Nézet, Louis and Sinnwell, Jason and Letko, Anna and André, Catherine and Quignon, Pascale},
     title = {Pedixplorer: a Bioconductor package to streamline pedigree design and visualization},
