+++ reference/ggPedigreeInteractive.md
@@ -213,5 +213,5 @@ library(BGmisc)
 data("potter")
 ggPedigreeInteractive(potter, famID = "famID", personID = "personID")
 
-{"x":{"data":[{"x":[2,1,null,null,0,null,1,2,null,3,4,null,4,3,null,null,1.5,null,4,3,null,3,4,null,8.5,7.5,null,7.5,8.5,null,6,5,null,null,7,null,null,8,null,10,9,null,null,11,null,13,12,null,5,6,null,12,13,null,null,14,null,9,10,null,null,2.2250000000000001,null,null,3.2250000000000001,null,null,4.2249999999999996,null,null,5.2249999999999996,null,null,6.2249999999999996,null,null,11.5,null,null,12.5,null,null,13.5,null,null,9,null,null,10,null,0,1,null,1,0,null,3.5,2.5,null,2.5,3.5,null,14,13,null,13,14],"y":[-2,-2,null,null,-2,null,-2,-2,null,-2,-2,null,-2,-2,null,null,-3,null,-3,-3,null,-3,-3,null,-2,-2,null,-2,-2,null,-3,-3,null,null,-3,null,null,-3,null,-3,-3,null,null,-3,null,-3,-3,null,-3,-3,null,-3,-3,null,null,-3,null,-3,-3,null,null,-4,null,null,-4,null,null,-4,null,null,-4,null,null,-4,null,null,-4,null,null,-4,null,null,-4,null,null,-4,null,null,-4,null,-1,-1,null,-1,-1,null,-1,-1,null,-1,-1,null,-2,-2,null,-2,-2],"text":"","type":"scatter","mode":"lines","line":{"width":1.8897637795275593,"color":"rgba(0,0,0,1)","dash":"solid"},"hoveron":"points","showlegend":false,"xaxis":"x","yaxis":"y","hoverinfo":"text","frame":null},{"x":[0.5,0.5,null,0.5,0.5,null,3,3,null,3,3,null,null,null,null,1.5,1.5,null,3,3.5,null,8,8,null,null,null,null,null,null,null,8,8,null,8,8,null,8,8,null,8,8,null,8,8,null,8,8,null,null,null,null,13.5,13.5,null,13.5,13.5,null,null,null,null,3.2250000000000001,3.5,null,3.2250000000000001,3.5,null,3.2250000000000001,3.5,null,5.7249999999999996,5.5,null,5.7249999999999996,5.5,null,12.5,12.5,null,12.5,12.5,null,12.5,12.5,null,9.5,9.5,null,9.5,9.5,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null],"y":[-1.5,-1,null,-1.5,-1,null,-1.5,-1,null,-1.5,-1,null,null,null,null,-2.5,-2,null,-2.5,-2,null,-2.5,-2,null,null,null,null,null,null,null,-2.5,-2,null,-2.5,-2,null,-2.5,-2,null,-2.5,-2,null,-2.5,-2,null,-2.5,-2,null,null,null,null,-2.5,-2,null,-2.5,-2,null,null,null,null,-3.5,-3,null,-3.5,-3,null,-3.5,-3,null,-3.5,-3,null,-3.5,-3,null,-3.5,-3,null,-3.5,-3,null,-3.5,-3,null,-3.5,-3,null,-3.5,-3,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null],"text":"","type":"scatter","mode":"lines","line":{"width":1.8897637795275593,"color":"rgba(0,0,0,1)","dash":"solid"},"hoveron":"points","showlegend":false,"xaxis":"x","yaxis":"y","hoverinfo":"text","frame":null},{"x":[1,0.5,null,0,0.5,null,2,3,null,4
[…]
