### 2025-11-24 39e5cf0
Update kinship2_pedigree.R
seem to have properly eliminated dependency
