+++ NAMESPACE
@@ -2,8 +2,6 @@
 
 S3method("[",pedigree)
 S3method("[",pedigreeList)
-S3method(print,pedigree)
-S3method(print,pedigreeList)
 export(calculateConnections)
 export(calculateCoordinates)
 export(countOffspring)
