### 2025-11-25 618f3e3
Refactor kinship2 alignment functions for clarity
Reformatted kinship2_alignped1, kinship2_alignped2, kinship2_alignped3, kinship2_alignped4, kinship2_bitSize, kinship2_check.hint, and kinship2_internals functions for improved readability and consistency. Changes include indentation, spacing, and minor logic clarifications, but no functional changes to the algorithms. This improves maintainability and code style across the kinship2 alignment and plotting internals.
