+++ R/kinship2_bitSize.R
@@ -23,21 +23,23 @@
 kinship2_bitSize <- function(ped) {
   ## calculate bit size of a pedigree
 
-  if(!("pedigree" %in% class(ped)))
+  if (!("pedigree" %in% class(ped))) {
     stop("Must be a pedigree object.\n")
+  }
 
   father <- ped$findex
   mother <- ped$mindex
   id <- ped$id
 
-  founder <- father==0 & mother==0
+  founder <- father == 0 & mother == 0
   pedSize <- length(father)
   nFounder <- sum(founder)
   nNonFounder <- pedSize - nFounder
-  bitSize <- 2*nNonFounder - nFounder
+  bitSize <- 2 * nNonFounder - nFounder
 
-  return(list(bitSize=bitSize,
+  return(list(
+    bitSize = bitSize,
     nFounder = nFounder,
-    nNonFounder = nNonFounder))
+    nNonFounder = nNonFounder
+  ))
 }
-
