### 2025-11-25 8aed00c
refactor to reduce complexity
