### 2025-11-25 71dbb19
split kinship2 internals
