+++ R/kinship2_pedigree.R
@@ -171,107 +171,16 @@ sex <- pedigree.sexrepair(sex=sex)
     temp$status <- pedigree.process_status(status, n)
   }
 
-  if (!missing(relation)) {
-    if (!missing(famid)) {
-      if (is.matrix(relation)) {
-        if (ncol(relation) != 4) {
-          stop("Relation matrix must have 3 columns + famid")
-        }
-        id1 <- relation[, 1]
-        id2 <- relation[, 2]
-        code <- relation[, 3]
-        famid <- relation[, 4]
-      } else if (is.data.frame(relation)) {
-        id1 <- relation$id1
-        id2 <- relation$id2
-        code <- relation$code
-        famid <- relation$famid
-        if (is.null(id1) || is.null(id2) || is.null(code) || is.null(famid)) {
-          stop("Relation data must have id1, id2, code, and family id")
-        }
-      } else {
-        stop("Relation argument must be a matrix or a dataframe")
-      }
-    } else {
-      if (is.matrix(relation)) {
-        if (ncol(relation) != 3) {
-          stop("Relation matrix must have 3 columns: id1, id2, code")
-        }
-        id1 <- relation[, 1]
-        id2 <- relation[, 2]
-        code <- relation[, 3]
-      } else if (is.data.frame(relation)) {
-        id1 <- relation$id1
-        id2 <- relation$id2
-        code <- relation$code
-        if (is.null(id1) || is.null(id2) || is.null(code)) {
-          stop("Relation data frame must have id1, id2, and code")
-        }
-      } else {
-        stop("Relation argument must be a matrix or a list")
-      }
-    }
-
-    if (!is.numeric(code)) {
-      code <- match(code, c("MZ twin", "DZ twin", "UZ twin", "spouse"))
-    } else {
-      code <- factor(code,
-        levels = 1:4,
-        labels = c("MZ twin", "DZ twin", "UZ twin", "spouse")
-      )
-    }
-    if (any(is.na(code))) {
-      stop("Invalid relationship code")
-    }
-
-    # Is everyone in this relationship in the pedigree?
-    if (!missing(famid)) {
-      temp1 <- match(paste(as.character(famid), as.character(id1), sep = "/"),
-        id,
-        nomatch = 0
-      )
-      temp2 <- match(paste(as.character(famid), as.character(id2), sep = "/"),
-        id,
-        nomatch = 0
-      )
-    } else {
-      temp1 <- match(id1, id, nomatch = 0)
-      temp2 <- match(id2, id, nomatch = 0)
-    }
-
-    if (any(temp1 == 0 | temp2 == 0)) {
-      stop("Subjects in relationships that are not in the pedigree")
-    }
-    if (any(temp1 == temp2)) {
-      who <- temp1[temp1 == temp2]
-      stop(paste("Subject", id[who], "is their own spouse or twin"))
-    }
-
-    # Check, are the twins really twins?
-    ncode <- as.numeric(code)
-    if (any(ncode < 3)) {
-      twins <- (ncode < 3)
-      if (any(momid[temp1[twins]] != momid[temp2[twins]])) {
-        stop("Twins found with different mothers")
-      }
-      if (any(dadid[temp1[twins]] != dadid[temp2[twins]])) {
-        stop("Twins found with different fathers")
-      }
-    }
-    # Check, are the monozygote twins the same gender?
-    if (any(code == "MZ twin")) {
-      mztwins <- (code == "MZ twin")
-      if (any(sex[temp1[mztwins]] != sex[temp2[mztwins]])) {
-        stop("MZ twins with different sexes")
-      }
-    }
-
[…]
