+++ R/kinship2_internals.R
@@ -1,309 +1,3 @@
-# From autohint.R file
-# Automatically generated from all.nw using noweb
-
-#' Automatically generate alignment hints for pedigree plotting
-#'
-#' This function automatically generates alignment hints for pedigree plotting.
-#' Hints control the relative horizontal positioning of subjects within their
-#' generation and the placement of spouse pairs. The function handles twins,
-#' multiple marriages, and complex pedigree structures.
-#'
-#' @param ped A pedigree object
-#' @param hints Optional existing hints (list with `order` and optionally `spouse` components)
-#' @param packed Logical, if TRUE uses compact packing algorithm (default TRUE)
-#' @param align Logical, if TRUE attempts to align spouses on the same level (default FALSE)
-#' @return A list containing:
-#'   \item{order}{Numeric vector of relative ordering hints for subjects}
-#'   \item{spouse}{Matrix of spouse pair information}
-#' @keywords internal
-#' @details
-#' The function is called automatically by kinship2_align.pedigree if no hints
-#' are provided. It analyzes the pedigree structure, identifies twins, handles
-#' multiple marriages, and determines optimal subject ordering to minimize
-#' crossing lines and produce aesthetically pleasing plots.
-#'
-#' Full documentation is available in the align_code_details vignette.
-kinship2_autohint <- function(ped, hints, packed = TRUE, align = FALSE) {
-  ## full documentation now in vignette: align_code_details.Rmd
-  ## REferences to those sections appear here as:
-  ## Doc: AutoHint
-  if (!is.null(ped$hints)) {
-    return(ped$hints)
-  } # nothing to do
-  n <- length(ped$id)
-  depth <- kinship2_kindepth(ped, align = TRUE)
-
-  if (is.null(ped$relation)) {
-    relation <- NULL
-  } else {
-    relation <- cbind(
-      as.matrix(ped$relation[, 1:2]),
-      as.numeric(ped$relation[, 3])
-    )
-  }
-  if (!is.null(relation) && any(relation[, 3] < 4)) {
-    temp <- (relation[, 3] < 4)
-    twinlist <- unique(c(relation[temp, 1:2])) # list of twin id's
-    twinrel <- relation[temp, , drop = FALSE]
-
-    twinset <- rep(0, n)
-    twinord <- rep(1, n)
-    for (i in 2:length(twinlist)) {
-      # Now, for any pair of twins on a line of twinrel, give both
-      #  of them the minimum of the two ids
-      # For a set of triplets, it might take two iterations for the
-      #  smallest of the 3 numbers to "march" across the threesome.
-      #  For quads, up to 3 iterations, for quints, up to 4, ....
-      newid <- pmin(twinrel[, 1], twinrel[, 2])
-      twinset[twinrel[, 1]] <- newid
-      twinset[twinrel[, 2]] <- newid
-      twinord[twinrel[, 2]] <- pmax(
-        twinord[twinrel[, 2]],
-        twinord[twinrel[, 1]] + 1
-      )
-    }
-  } else {
-    twinset <- rep(0, n)
-    twinrel <- NULL
-  }
-  ## Doc: Shift
-  shift <- function(id, sibs, goleft, hint, twinrel, twinset) {
-    if (twinset[id] > 0) {
-      shift.amt <- 1 + diff(range(hint[sibs])) # enough to avoid overlap
-      twins <- sibs[twinset[sibs] == twinset[id]]
-      if (goleft) {
[…]
