### 2025-11-27 c13a484
add tests
