+++ R/kinship2_pedigree.R
@@ -36,23 +36,26 @@
 #' @name pedigree
 
 
-pedigree <- function(id, dadid, momid, sex, affected, status, relation,
+pedigree <- function(id, dadid, momid,
+                     sex, affected,
+                     status, relation,
                      famid, missid) {
   n <- length(id)
   ## Code transferred from noweb to markdown vignette.
   ## Sections from the noweb/vignettes are noted here with
   ## Doc: Error and Data Checks
 
-  pedigree.idcheck(id=id,
-                   momid = momid,
-                   dadid = dadid,
-                   sex=sex
-                 )
+  pedigree.idcheck(
+    id = id,
+    momid = momid,
+    dadid = dadid,
+    sex = sex
+  )
 
 
-id <- pedigree.idrepair(id=id)
+  id <- pedigree.idrepair(id = id)
 
-sex <- pedigree.sexrepair(sex=sex)
+  sex <- pedigree.sexrepair(sex = sex)
 
   ## Doc:  Errors2
   if (missing(missid)) {
@@ -79,6 +82,7 @@ sex <- pedigree.sexrepair(sex=sex)
     dadid <- paste(as.character(famid), as.character(dadid), sep = "/")
     momid <- paste(as.character(famid), as.character(momid), sep = "/")
   }
+  has_famid <- !missing(famid)
 
   if (any(duplicated(id))) {
     duplist <- id[duplicated(id)]
@@ -165,21 +169,19 @@ sex <- pedigree.sexrepair(sex=sex)
   }
 
 
-
-
   if (!missing(status)) {
     temp$status <- pedigree.process_status(status, n)
   }
 
-if (!missing(relation)) {
+  if (!missing(relation)) {
     temp$relation <- pedigree.process_relation(
       relation = relation,
       has_famid = has_famid,
-      famid     = if (has_famid) famid else NULL,
-      id        = id,
-      momid     = momid,
-      dadid     = dadid,
-      sex       = sex
+      famid = if (has_famid) famid else NULL,
+      id = id,
+      momid = momid,
+      dadid = dadid,
+      sex = sex
     )
   }
   ## Doc: Finish
@@ -200,10 +202,12 @@ pedigree.process_relation <- function(relation,
                                       momid,
                                       dadid,
                                       sex) {
-  rel_parsed <- pedigree.parse_relation(relation, has_famid = has_famid)
-  id1       <- rel_parsed$id1
-  id2       <- rel_parsed$id2
-  rel_code  <- pedigree.coerce_relation_code(rel_parsed$code)
+  rel_parsed <- pedigree.parse_relation(relation,
+    has_famid = has_famid
+  )
+  id1 <- rel_parsed$id1
+  id2 <- rel_parsed$id2
+  rel_code <- pedigree.coerce_relation_code(rel_parsed$code)
   rel_famid <- rel_parsed$famid
 
   # Ensure everyone in the relationship is in the pedigree
@@ -272,18 +276,18 @@ pedigree.parse_relation <- function(relation, has_famid) {
         stop("Relation matrix must have 3 columns: id1, id2, code")
       }
     }
-    id1  <- relation[, 1]
-    id2  <- relation[, 2]
+    id1 <- relation[, 1]
+    id2 <- relation[, 2]
     code <- relation[, 3]
-    fam  <- if (has_famid) relation[, 4] else NULL
+    fam <- if (has_famid) relation[, 4] else NULL
   } else if (is.data.frame(relation)) {
-    id1  <- relation$id1
-    id2  <- relation$id2
+    id1 <- relation$id1
+    id2 <- relation$id2
[…]
