+++ man/kinship2_autohint.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/kinship2_internals.R
+% Please edit documentation in R/kinship2_autohints.R
 \name{kinship2_autohint}
 \alias{kinship2_autohint}
 \title{Automatically generate alignment hints for pedigree plotting}

[…]
