### 2025-11-27 7b267f6
Refactor duporder to kinship2_duporder and add docs
Moved the duporder function out of kinship2_autohints.R and renamed it to kinship2_duporder for clarity and maintainability. Added roxygen2 documentation and a corresponding Rd file for kinship2_duporder. Updated references in kinship2_autohints.R and expanded NEWS.md and tests to reflect these changes.
