+++ NEWS.md
@@ -1,7 +1,10 @@
 # ggpedigree (development version 1.00)
-* Folded in kinship2 functions to eliminate the depedency optional
+* Folded in kinship2 functions to eliminate the dependency
 * Made kinship2 package optional, and only needed for plotPedigree (a kinship2 wrapper)
 * Importing relevant tests from kinship2 package
+* Updated documentation to reflect changes
+* Refactored code to improve maintainability
+* Added more unit tests for the new functions
 
 # ggpedigree 0.9.1
 * Minor documentation revisions to complement JOSS paper update
