+++ tests/testthat/_snaps/kinship2_aligns.md
@@ -1,86 +1,4 @@
-# align.pedigree works
-
-    Code
-      kinship2_align.pedigree(ped)
-    Output
-      $n
-      [1]  8 19 22  8
-      
-      $nid
-           [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9]
-      [1,]    5    6    7    8   35   36   42   43    0
-      [2,]    1    2   15   17   16   18   19   20    3
-      [3,]    9   10   29   30   31   32   33   34   10
-      [4,]   21   22   23   24   25   26   27   28    0
-           [,10] [,11] [,12] [,13] [,14] [,15] [,16]
-      [1,]     0     0     0     0     0     0     0
-      [2,]     4    37    38    44    45    46    47
-      [3,]    11    12    18    14    15    39    40
-      [4,]     0     0     0     0     0     0     0
-           [,17] [,18] [,19] [,20] [,21] [,22]
-      [1,]     0     0     0     0     0     0
-      [2,]    48    50    49     0     0     0
-      [3,]    41    51    52    53    54    55
-      [4,]     0     0     0     0     0     0
-      
-      $pos
-                    [,1] [,2] [,3] [,4]     [,5]
-      [1,]  3.750000e+00 4.75 6.50 7.50 11.75536
-      [2,] -4.541957e-13 1.00 2.00 3.00  4.00000
-      [3,]  1.000000e-02 1.01 2.01 3.01  4.01000
-      [4,]  1.939898e-16 1.00 2.00 3.00 10.01000
-               [,6]  [,7]  [,8]     [,9]    [,10] [,11]
-      [1,] 12.75536 17.71 18.71  0.00000  0.00000  0.00
-      [2,]  5.00000  6.00  7.00 10.50072 11.50072 13.01
-      [3,]  5.01000  6.01  7.01  8.01000  9.01000 10.01
-      [4,] 11.01000 12.01 13.01  0.00000  0.00000  0.00
-           [,12] [,13] [,14] [,15] [,16] [,17] [,18]
-      [1,]  0.00  0.00  0.00  0.00  0.00  0.00  0.00
-      [2,] 14.01 15.01 16.01 17.01 18.01 19.01 20.01
-      [3,] 11.01 12.01 13.01 14.01 15.01 16.01 17.01
-      [4,]  0.00  0.00  0.00  0.00  0.00  0.00  0.00
-           [,19] [,20] [,21] [,22]
-      [1,]  0.00  0.00  0.00  0.00
-      [2,] 21.01  0.00  0.00  0.00
-      [3,] 18.01 19.01 20.01 21.01
-      [4,]  0.00  0.00  0.00  0.00
-      
-      $fam
-           [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9]
-      [1,]    0    0    0    0    0    0    0    0    0
-      [2,]    0    0    1    0    1    1    1    3    5
-      [3,]    1    0    4    7    7    7    7    7    9
-      [4,]    1    1    1    1   11   11   13   13    0
-           [,10] [,11] [,12] [,13] [,14] [,15] [,16]
-      [1,]     0     0     0     0     0     0     0
-      [2,]     0     0     5     0     7     7     7
-      [3,]     9     9     0     9     0    11    11
-      [4,]     0     0     0     0     0     0     0
-           [,17] [,18] [,19] [,20] [,21] [,22]
-      [1,]     0     0     0     0     0     0
-      [2,]     7     0     7     0     0     0
-      [3,]    11    13    13    18    18    18
-      [4,]     0     0     0     0     0     0
-      
-      $spouse
-           [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9]
-      [1,]    1    0    1    0    1    0    1    0    0
[…]
