### 2025-11-27 fc5ebed
Update test-kinship2_aligns.R
eliminated snaps
