+++ tests/testthat/test-kinship2_aligns.R
@@ -21,7 +21,11 @@ test_that("test autohint works with sample.ped", {
     packed = TRUE, align = TRUE,
     width = 8, hints = newhint
   )
-  expect_snapshot(plist)
+  #  expect_snapshot(plist)
+  expect_equal(plist$n, c(8, 19, 22, 8))
+  expect_equal(dim(plist$nid), c(4, 22))
+  expect_equal(dim(plist$pos), c(4, 22))
+  expect_equal(dim(plist$fam), c(4, 22))
 })
 
 test_that("align.pedigree works with ASOIAF", {
@@ -63,7 +67,13 @@ test_that("test autohint works with ASOIAF", {
     packed = TRUE, align = TRUE,
     width = 8, hints = newhint
   )
-  expect_snapshot(plist)
+  expect_equal(plist$n, c(
+    34, 63, 68, 52, 22, 11, 19, 37, 93,
+    92, 41, 23, 30, 24, 26, 20, 30, 26, 19, 43, 36
+  ))
+  expect_equal(dim(plist$nid), c(21, 93))
+  expect_equal(dim(plist$pos), c(21, 93))
+  expect_equal(dim(plist$fam), c(21, 93))
 })
 
 
@@ -90,5 +100,8 @@ test_that("test autohint works when packed false", {
     packed = FALSE, align = TRUE,
     width = 8, hints = newhint
   )
-  expect_snapshot(plist)
+  expect_equal(plist$n, c(8, 19, 22, 8))
+  expect_equal(dim(plist$nid), c(4, 22))
+  expect_equal(dim(plist$pos), c(4, 22))
+  expect_equal(dim(plist$fam), c(4, 22))
 })
