### 2025-11-27 4aa77b9
Remove 'dev' branch from pull request triggers
