+++ .github/workflows/pkgdown.yaml
@@ -4,7 +4,7 @@ on:
   push:
     branches: [main, master]
   pull_request:
-    branches: [main, master, dev]
+    branches: [main, master]
   release:
     types: [published]
   workflow_dispatch:
