### 2025-11-28 11ccd2c
internals
