### 2025-11-29 9fec2cb
Expand label and group handling in ggPedigree.core
Added additional synonyms for maternal, paternal, and family groupings to improve flexibility in group recognition. Refactored label method checks to use a variable for ggrepel methods, and improved error handling for invalid label methods. Updated join_by usage to be explicit with dplyr namespace.
