+++ reference/calculateCoordinates.md
@@ -70,7 +70,7 @@ calculateCoordinates(
 
   ped_packed
 
-  :   Logical, default TRUE. Passed to \`kinship2::align.pedigree\`.
+  :   Logical, default TRUE. Passed to \`kinship2_align.pedigree\`.
 
   ped_align
 
