+++ reference/kinship2_duporder.md
@@ -1,0 +1,13 @@
+# Find all duplicated IDs on a level, and return
+
+Find all duplicated IDs on a level, and return
+
+## Usage
+
+``` r
+kinship2_duporder(idlist, plist, lev, ped)
+```
+
+## Value
+
+A matrix with one row per duplicated ID, columns 1 and 2
