### 2025-11-30 268a5b1
Update GitFlow strategy link in README files
Replaced the outdated GitFlow branching strategy link with the official GitLab documentation link in both README.Rmd and README.md for improved reference accuracy.
