+++ llms.txt
@@ -86,7 +86,7 @@ For support, please use the GitHub issues page.
 ### Branching and Versioning System
 
 The development of ggpedigree follows a [GitFlow branching
-strategy](https://tilburgsciencehub.com/topics/automation/version-control/advanced-git/git-branching-strategies/):
+strategy](https://docs.gitlab.com/user/project/repository/branches/strategies/):
 
 - **Feature Branches**: All major changes and new features should be
   developed on separate branches created from the dev branch. Name these
