### 2025-11-30 22b7c44
updating minor fixes for cran resubmission
