+++ R/calcConnections.R
@@ -187,7 +187,7 @@ calculateConnections <- function(ped,
     dplyr::left_join(mom_connections,
       by = c("personID", "momID")
     ) |>
-    unique() |>
+    unique() |> 
     dplyr::left_join(dad_connections,
       by = c("personID", "dadID")
     ) |>
@@ -351,13 +351,15 @@ buildSpouseSegments <- function(ped, connections_for_FOO, use_hash = TRUE) {
         suffix = c("", "_parent1"),
         multiple = "any"
       ) |>
+      unique()|> 
       dplyr::left_join(
         connections_for_FOO |>
           dplyr::mutate(personID = paste0(.data$personID)),
         by = c("parent2" = "personID"),
         suffix = c("", "_parent2"),
         multiple = "any"
-      ) |>
+      )  |>
+      unique()|> 
       dplyr::mutate(
         x_start = .data$x_pos,
         x_end = .data$x_pos_parent2,
@@ -433,11 +435,11 @@ buildTwinSegments <- function(ped, connections_for_FOO) {
       suffix = c("", "_twin"),
       multiple = "all"
     ) |>
+    unique() |>
     dplyr::rename(
       x_twin = "x_pos_twin",
       y_twin = "y_pos_twin"
     ) |>
-    unique() |>
     dplyr::mutate(
       x_mid_twin = (.data$x_pos + .data$x_twin) / 2,
       y_mid_twin = (.data$y_pos + .data$y_twin) / 2
