+++ news/index.md
@@ -2,6 +2,13 @@
 
 ## ggpedigree (development version)
 
+## ggpedigree 1.0.0.1
+
+### cran resubmission
+
+- Fixed a minor documentation issue with a url whose certificate had
+  expired.
+
 ## ggpedigree 1.0.0
 
 ### cran submission
