+++ llms.txt
@@ -60,7 +60,7 @@ citation(package = "ggpedigree")
 ```
 
 Garrison S (2025). *ggpedigree: Visualizing Pedigrees with ‘ggplot2’ and
-‘plotly’*. R package version 1.0.0,
+‘plotly’*. R package version 1.0.0.1,
 <https://github.com/R-Computing-Lab/ggpedigree/>.
 
 A BibTeX entry for LaTeX users is
@@ -70,7 +70,7 @@ A BibTeX entry for LaTeX users is
   title = {ggpedigree: Visualizing Pedigrees with 'ggplot2' and 'plotly'},
   author = {S. Mason Garrison},
   year = {2025},
-  note = {R package version 1.0.0},
+  note = {R package version 1.0.0.1},
   url = {https://github.com/R-Computing-Lab/ggpedigree/},
 }
 ```
