### 2025-12-02 bbde10c
Refactor kinship2 autohint logic and export functions
Refactored kinship2_autohint by moving the shift logic to a new exported function kinship2_shift, improving maintainability and modularity. Exported kinship2_autohint and pedigree functions, added pedigree.makemissingid helper, and updated documentation accordingly.
