+++ NEWS.md
@@ -1,6 +1,7 @@
 # ggpedigree (development version)
 
 * Saved raw excel data as csv to have a non-proprietary format for data storage
+* Added more refactorings of kinship2 internals for maintainability
 
 # ggpedigree 1.0.0.1
 ## cran resubmission
