### 2025-12-02 95f2f68
Refactor autohint shift logic and update spouse segment filtering
Moved the shift function from kinship2_autohints.R to a standalone kinship2_shift function and updated references to use the new function. Updated comments and removed unique() filtering in spouse and twin segment calculations in calcConnections.R to fix issues with segment generation.
