### 2025-12-07 d23107a
Refactor autohint logic and add helper functions
Extracted hint cleaning and horder setting logic into new helper functions kinship2_autohint.cleanhints and kinship2_autohint.sethorder for improved readability and maintainability. Also updated conditional checks in kinship2_shift to use explicit TRUE comparisons.
