+++ R/calcConnections.R
@@ -22,7 +22,8 @@ calculateConnections <- function(ped,
                                  config = list(),
                                  spouseID = "spouseID",
                                  personID = "personID",
-                                 momID = "momID", famID = "famID",
+                                 momID = "momID",
+                                 famID = "famID",
                                  twinID = "twinID",
                                  dadID = "dadID") {
   # Check inputs -----------------------------------------------------------
@@ -352,15 +353,15 @@ buildSpouseSegments <- function(ped, connections_for_FOO, use_hash = TRUE) {
         suffix = c("", "_parent1"),
         multiple = "any"
       ) |>
-     # unique()|> NOPE
+      # unique()|> NOPE
       dplyr::left_join(
         connections_for_FOO |>
           dplyr::mutate(personID = paste0(.data$personID)),
         by = c("parent2" = "personID"),
         suffix = c("", "_parent2"),
         multiple = "any"
-      )  |>
-   #   unique()|> NOPE
+      ) |>
+      #   unique()|> NOPE
       dplyr::mutate(
         x_start = .data$x_pos,
         x_end = .data$x_pos_parent2,
@@ -376,7 +377,6 @@ buildSpouseSegments <- function(ped, connections_for_FOO, use_hash = TRUE) {
         -"x_pos_parent2",
         -"y_pos_parent2"
       )
-
   } else {
     # spouses
     # Get spouse coordinates
@@ -391,7 +391,7 @@ buildSpouseSegments <- function(ped, connections_for_FOO, use_hash = TRUE) {
         suffix = c("", "_spouse"),
         multiple = "any"
       ) |>
-   #   unique() |> NOPE
+      #   unique() |> NOPE
       dplyr::rename(
         x_spouse = "x_pos_spouse",
         y_spouse = "y_pos_spouse"
@@ -435,7 +435,7 @@ buildTwinSegments <- function(ped, connections_for_FOO) {
       suffix = c("", "_twin"),
       multiple = "all"
     ) |>
-#    unique() |> NOPE
+    #    unique() |> NOPE
     dplyr::rename(
       x_twin = "x_pos_twin",
       y_twin = "y_pos_twin"
