### 2025-12-08 R-Computing-Lab/ggpedigree#92 @magnusdv
`sex` column required?
Context: openjournals/joss-reviews#9434

Is `sex` a required column in `ggpedigree()`? If so, this should be stated in the doc, and ideally validated early in on the function. (The current help page only lists personID, momID, and dadID as required.)

Without `sex` I get this confusing error:
``` r
library(ggpedigree)

df1 = data.frame(personID = 1:3, dadID = c(0,0,1), momID = c(0,0,2))
ggpedigree(df1)
#> Error in pedigree.sexrepair(sex = sex): Invalid values for 'sex'
```
