### 2025-12-08 R-Computing-Lab/ggpedigree#93 @magnusdv
Singleton plot
Context: openjournals/joss-reviews#9434

When plotting a single individual it seems to disappear:

``` r
library(ggpedigree)

df = data.frame(personID = 1, momID = 0, dadID = 0, sex = 1)
ggPedigree(df)
```

![](https://i.imgur.com/HcuUj9z.png)
