### 2025-12-08 R-Computing-Lab/ggpedigree#94 @magnusdv
Default plot appearance
Context: openjournals/joss-reviews#9434

Below is a simple plot, with a few comments on the default styling. I’m sure all of this can be adjusted with additional code, but it might be worthwhile tweaking some of the defaults to improve the first impression.

``` r
library(ggpedigree)
library(ggplot2)

df = data.frame(personID = 1:3, momID = c(0,0,2), dadID = c(0,0,1), sex = c(1,2,1))
ggPedigree(df) + theme(plot.background = element_rect(color = 1))
```

![](https://i.imgur.com/u50MtZi.png)

- Huge top margin: Is this intentional?
- Symbols are too small (for my taste).
- Large distance between labels and symbols.
- Personally I find the yellow color a bit difficult to see.
- Sex legend: Although useful in some cases, I would personally prefer having it disabled by default.

Out of curiosity: How hard is it to obtain a more ‘classical’ look with open symbols, like the example below?

``` r
library(pedtools)
x = nuclearPed()
plot(x)
```
