### 2025-12-09 R-Computing-Lab/ggpedigree#95 @magnusdv
Unexpected behaviour with simple pedigrees
Context: openjournals/joss-reviews#9434

Hi, I wanted to share a few issues I encountered while exploring the package. Overall, the takeaway is that the plotting function might benefit from checking the input a bit more thoroughly.

I also recommend documenting more clearly what the function expects. For instance: can the columns be integer/numeric/character? In particular, I am confused about how the `sex` column is interpreted (see point 6 below).

``` r
library(ggpedigree)
```

**1**) I don’t understand why this fails (and the error message is unhelpful)

``` r
df = data.frame(personID = c("1", "2", "3"), momID = c(NA, NA, "1"),
                dadID = c(NA, NA, "2"), sex = c(2, 1, 2))
df
#>   personID momID dadID sex
#> 1        1  <NA>  <NA>   2
#> 2        2  <NA>  <NA>   1
#> 3        3     1     2   2

ggPedigree(df)
#> Error in `dplyr::mutate()`:
#> ℹ In argument: `couple_hash = .makeSymmetricKey(.data$personID,
#>   .data$spouseID)`.
#> Caused by error in `.makeSymmetricKey()`:
#> ! id1 and id2 must be of the same type. id1 is numeric and id2 is character
```

**2**) Wrong sex, and parents swap position:

``` r
df = data.frame(personID = 1:3, momID = c(0, 0, 1),
                dadID = c(0, 0, 2), sex = c(2, 1, 2))
df
#>   personID momID dadID sex
#> 1        1     0     0   2
#> 2        2     0     0   1
#> 3        3     1     2   2

ggPedigree(df)
```

![](https://i.imgur.com/4hPolC6.png)

**3**) Also wrong sex:

``` r
df = data.frame(personID = 1:3, dadID = c(0, 0, 1),
                momID = c(0, 0, 2), sex = c(1, 2, 1))
df
#>   personID dadID momID sex
#> 1        1     0     0   1
#> 2        2     0     0   2
#> 3        3     1     2   1

ggPedigree(df)
```

![](https://i.imgur.com/veWa1wH.png)

**4**) Individual 3 disappears (I expected a diamond shape):

``` r
df = data.frame(personID = 1:3, dadID = c(0, 0, 1),
                momID = c(0, 0, 2), sex = c(1, 2, NA))
df
#>   personID dadID momID sex
#> 1        1     0     0   1
#> 2        2     0     0   2
#> 3        3     1     2  NA

ggPedigree(df)
#> Warning in pedigree.sexrepair(sex = sex): More than 25% of the sex values are
#> 'unknown'
```

![](https://i.imgur.com/f6qCiQF.png)

**5**) Works, but gives warning that I don’t understand:

``` r
df = data.frame(personID = c("a", "b", "c"), dadID = c(NA, NA, "a"),
                momID = c(NA, NA, "b"), sex = c(1, 0, 1))
df
#>   personID dadID momID sex
#> 1        a  <NA>  <NA>   1
#> 2        b  <NA>  <NA>   0
#> 3        c     a     b   1

ggPedigree(df)
#> Warning in data.frame(V1 = as.numeric(names(wcc$membership)), V2 =
#> wcc$membership): NAs introduced by coercion
```

![](https://i.imgur.com/kDfoTGu.png)

**6**) Inconsistent sex codes

``` r
df = data.frame(personID = 1:3, dadID = c(0,0,1), momID = c(0,0,2))
ggPedigree(cbind(df, sex = c(1,3,1))) # 1 -> female, 3 -> male
ggPedigree(cbind(df, sex = c(1,2,0))) # 0 -> female, 2 -> unknown
ggPedigree(cbind(df, sex = c(1,2,3))) # 2 -> male, 3 -> unknown
```

<img src="https://i.imgur.com/PIWjNKj.png" width="30%" /><img src="https://i.imgur.com/ejEMiZ4.png" width="30%" /><img src="https://i.imgur.com/MX7Nswp.png" width="30%" />
