### 2025-12-10 297adc6
id coocersion to numeric incomplete
still a work in progress, already added recoding 0s to NA. now need to fix why the id variables don't get handled the same
