+++ R/calcConnections.R
@@ -51,6 +51,18 @@ calculateConnections <- function(ped,
   }
   # Capture type-safe NAs for each ID column
   na_person <- ped$personID[NA_integer_]
+  na_mom <- ped$momID[NA_integer_]
+  na_dad <- ped$dadID[NA_integer_]
+
+  if(config$debug==TRUE) {
+    message("NA types - personID: ", class(na_person),
+            ", momID: ", class(na_mom),
+            ", dadID: ", class(na_dad))
+  }
+  if(class(na_person) != class(na_mom) ||
+     class(na_person) != class(na_dad)) {
+    stop("personID, momID, and dadID must be of the same type (e.g., all character or all numeric).")
+  }
 
   # Add spouseID if missing
   if (!all(spouseID %in% names(ped))) {
@@ -70,7 +82,7 @@ calculateConnections <- function(ped,
     )
 
     # Ensure class matches personID exactly (in case factor, character, etc.)
-    attributes(ped$spouseID) <- attributes(ped$personID)
+   # attributes(ped$spouseID) <- attributes(ped$personID)
   } else {
     # rename spouseID to match
     names(ped)[names(ped) == spouseID] <- "spouseID"
