### 2025-12-10 #9434 @magnusdv
I've now read the paper and played around with the package, overall a pleasant experience. The package provides highly customizable pedigree plots based on the `ggplot2` framework, and also `plotly` versions with some degree of interactivity. Although alternatives for this already exist (e.g. https://louislenezet.github.io/Pedixplorer/) the approach here is quite different in practice, with a more a `ggplot2`-centric workflow for customization.

I did have some initial trouble using the package, partly due to insufficient documentation, I've reported this in separate issues in the `ggpedigree` repo. Below are general comments on the manuscript and the software.

* Authorship: The package includes a substantial amount of code copied from `kinship2`, since that package may be retired. I am unsure what the best way to handle this is, but I would at least expect the DESCRIPTION to acknowledge it and give appropriate credit for the borrowed code. As it stands, the DESCRIPTION presents `ggpedigree` as an *alternative* to `kinship2`, which seems only partly true given that much of the plotting logic is inherited from `kinship2`.

* The paper acknowledges that the package adapts algorithms from `kinship2`, "with enhancements for large, multi-generational pedigrees and complex family structures". This is interesting and deserves more explanation. As `kinship2` is the currrent R standard for aligning large & complex pedigrees, such claimed improvements should be clearly described and ideally accompanied by runtime comparisons or similar.

* On the same note, the online vignette https://r-computing-lab.github.io/ggpedigree/articles/v22_plots_morecomplexity.html#visualize-the-pedigree-with-ggpedigree shows a large-ish pedigree plotted by both `kinship2` and `ggpedigree`. The example is surprising, since the `ggpedigree` versions are visually *inferior* to the `kinship2` one. I would recommend reworking this example so that it better highlights the strengths of `ggpedigree`.

* The paper presents `ggpedigree` as a stand-alone package for pedigree plotting, but in reality it is intertwined with another package, `BGmisc` for analysis in behavioural genetics. While this may work fine in principle, the separation between the two feels a bit ad hoc and incomplete. I think the project as a whole would benefit from clarifying the responsibilities of the two packages. A free-standing `ggplot2`-based pedigree plotter would likely appeal to users across many fields, but conceivably less so if it is too closely tied to behaviour genetics applications.

* The package includes two built-in example pedigrees, `redsquirrels` and `ASOIAF`, but both give errors when I try to plot them, and their help pages don’t show any example code. There are working examples elsewhere, but they rely on datasets from `BGmisc`. For a single-purpose package I would expect simple, self-contained examples.

* Several interesting exported functions, like `calculateCoordinates()`, lack examples in their help pages. Please add simple, self-contained examples. (I actually thought examples was a CRAN requirement, but apparently not?)

* Many exported functions of the package have very similar names:
  - ggPedigree
  - ggpedigree
  - ggPedigreeInteractive
  - ggPedigreeinteractive
  - ggpedigreeInteractive
  - ggpedigreeinteractive
  - pedigree
  - plotPedigree

  I understand that many of these are simple aliases, but in my view this has more cons than pros: It clutters tab-completion, makes the documentation hard to navigate, and obscures which function is intended as the main one. For example, the fact that `plotPedigree` is *not* the primary plotting function is quite confusing to me. I recommend to simplify the API and reduce some of the redundant functions.

* The package imports `reshape2` alongside tidyverse package, which seems unneccessary. As I understand it, the authors of `reshape2` have recommended switching to `tidyr` in new code.
