### 2025-12-11 4db9bad
Add War of Roses data and SVG family tree
Added EnglishMonarchy02.svg family tree, raw CSV data files (df_raw_wor.csv, warofroses.csv), R script (df_warroses.R), and warofroses.rda data object. These files provide source data and visualization for the English monarchy and War of the Roses analysis.
