### 2025-12-11 32dff50
Update package description in DESCRIPTION file
Revised the package description to clarify its focus on visualizing pedigrees and family trees, mention integration with BGmisc, and note the inclusion of layout helper functions from kinship2 due to its impending deprecation.
