+++ DESCRIPTION
@@ -6,7 +6,7 @@ Authors@R: c(
     person("S. Mason", "Garrison", email= "garrissm@wfu.edu", role = c("aut", "cre", "cph"),
            comment = c(ORCID = "0000-0002-4804-6003"))
   )
-Description: Provides plotting functions for visualizing pedigrees in behavior genetics and kinship research. The package complements 'BGmisc' [Garrison et al. (2024) <doi:10.21105/joss.06203>] by rendering pedigrees using the 'ggplot2' framework and offers a modern alternative to the base-graphics pedigree plot in 'kinship2' [Sinnwell et al. (2014) <doi:10.1159/000363105>]. Features include support for duplicated individuals, complex mating structures, integration with simulated pedigrees, and layout customization.
+Description: Provides plotting functions for visualizing pedigrees and family trees. in behavior genetics and kinship research. The package complements a behavior genetics package 'BGmisc' [Garrison et al. (2024) <doi:10.21105/joss.06203>] by rendering pedigrees using the 'ggplot2' framework and offers a modern redesign beyond the base-graphics pedigree plot in 'kinship2' [Sinnwell et al. (2014) <doi:10.1159/000363105>]. Features include support for duplicated individuals, complex mating structures, integration with simulated pedigrees, and layout customization. Due to the impending depreciation of kinship2, version 1.0 incorporates the layout helper functions from kinship2. 
 License: GPL-3
 URL: https://github.com/R-Computing-Lab/ggpedigree/,
      https://r-computing-lab.github.io/ggpedigree/
