### 2025-12-11 6d3f74e
Improve pedigree ID type handling and recoding
Enhances type safety and consistency for pedigree ID columns (personID, momID, dadID, spouseID) across functions, ensuring all are of the same class and recoding missing IDs (e.g., 0) to NA. Updates documentation to reflect new config option 'recode_missing_ids', improves error messages, and adds/extends tests for type consistency and missing spouse scenarios.
