### 2025-12-12 257d629
Remove debug print statements from calcConnections.R
Eliminated print statements that output the classes of ped$spouseID, ped$personID, ped$momID, and ped$dadID. This cleans up the function by removing unnecessary debug output.
