### 2025-12-12 R-Computing-Lab/ggpedigree#95 @smasongarrison
# Checklist
- [x] 1) I don’t understand why this fails (and the error message is unhelpful)
- 6d3f74edb162e05d886fa96a609471b7792a2d3d Problem seemed to be from inconsistent handling of string IDs
- [ ] 2) Wrong sex, and parents swap position:
- [ ] 3) Also wrong sex:
- [ ] 4) Individual 3 disappears (I expected a diamond shape):
- [ ] 5) Works, but gives warning that I don’t understand:
- [ ] 6) Inconsistent sex codes
