### 2025-12-12 f1a0bd0
Improve tests for string vs numeric ID handling
Refactored and expanded test cases in test-joss_review.R to thoroughly check consistent handling of string and numeric IDs, including various missing value representations. Updated NEWS.md to document the repair of inconsistent ID handling.
