+++ tests/testthat/test-joss_review.R
@@ -1,36 +1,82 @@
-test_that("checking the input a bit more thoroughly", {
-  df = data.frame(personID = c("1", "2", "3"), momID = c(NA, NA, "1"),
-                  dadID = c(NA, NA, "2"), sex = c(2, 1, 2))
-  df
-  #>   personID momID dadID sex
-  #> 1        1  <NA>  <NA>   2
-  #> 2        2  <NA>  <NA>   1
-  #> 3        3     1     2   2
-  ggPedigree(df,config = list(debug=TRUE))
-  expect_no_error(ggPedigree(df))
 
-  #> Error in `dplyr::mutate()`:
-  #> ℹ In argument: `couple_hash = .makeSymmetricKey(.data$personID,
-  #>   .data$spouseID)`.
-  #> Caused by error in `.makeSymmetricKey()`:
-  #> ! id1 and id2 must be of the same type. id1 is numeric and id2 is character
-})
-test_that("checking the input a bit more thoroughly", {
-  df = data.frame(personID = c("1", "2", "3"), momID = c(NA, NA, "1"),
-                  dadID = c(NA, NA, "2"), sex = c(0, 1, 0))
-  df
-  #>   personID momID dadID sex
-  #>   1        1  <NA>  <NA>   0
-  #>   2        2  <NA>  <NA>   1
-  #>   3        3     1     2   0
-  ggPedigree(df,config = list(debug=TRUE))
+expect_roundtrip <- function(df,
+                             cols = c("personID", "momID", "dadID", "sex"),
+                             NA_id_value = NA
+                             ) {
+  ped_df <- ggPedigree(df, config = list(debug = TRUE))[["data"]]
+
+  if (!is.na(NA_id_value)) {
+    df$momID[df$momID == NA_id_value] <- NA
+    df$dadID[df$dadID == NA_id_value] <- NA
+  }
+  expect_equal(ped_df[, cols, drop = FALSE],
+               df[, cols, drop = FALSE],
+               ignore_attr = TRUE)
+
+  expect_equal(nrow(ped_df), nrow(df))
   expect_no_error(ggPedigree(df))
-#  ebug mode is ON. Debugging information will be printed.
- # Coordinates calculated. Number of individuals: 3
-#  NA types - personID: numeric, momID: character, dadID: character
-  #> Error in `dplyr::mutate()`:
-  #> ℹ In argument: `couple_hash = .makeSymmetricKey(.data$personID,
-  #>   .data$spouseID)`.
-  #> Caused by error in `.makeSymmetricKey()`:
-  #> ! id1 and id2 must be of the same type. id1 is numeric and id2 is character
+}
+
+case_char <- function(sex = c(0,1,0),
+                      missing_parent= NA_character_
+                      ) {
+  data.frame(
+  personID = c("1","2","3"),
+  momID = c(missing_parent, missing_parent, "1"),
+  dadID = c(missing_parent, missing_parent, "2"),
+  sex = sex
+)
+}
+case_num <- function(sex=c(0,1,0),
+                     missing_parent = NA_real_
+                     ) {
+  data.frame(
+  personID = 1:3,
+  dadID = c(missing_parent, missing_parent, 1),
+  momID = c(missing_parent, missing_parent, 2),
+  sex = sex
+)
+}
+
+
+test_that("char ids; 0 missingid; 1/2",  {
+  expect_roundtrip(case_char(c(2,1,2),"0"), NA_id_value = "0")
 })
+test_that("char ids; 0 missingid; 1/2 + NA", {
[…]
