+++ NEWS.md
@@ -3,6 +3,7 @@
 * Saved raw excel data as csv to have a non-proprietary format for data storage
 * Added more refactorings of kinship2 internals for maintainability
 * Harmonize names by changing midparent to mid_parent throughout the codebase
+* Repair inconsistent handling of string IDs vs numeric IDs
 
 # ggpedigree 1.0.0.1
 ## cran resubmission
