### 2025-12-13 615e01e
improve sex code handling
 Updates sex code handling in alignPedigreeWithRelations and pedigree.sexrepair to better support non-standard and numeric codes. Enhances tests to cover new config options and warning handling for sex code edge cases.
