+++ R/calcCoordinates.R
@@ -255,8 +255,16 @@ alignPedigreeWithRelations <- function(ped,
                                        code_male,
                                        sexVar,
                                        config) {
+
+
+  # recodeSex <- function(
+  #  ped, verbose = FALSE, code_male = NULL, code_na = NULL, code_female = NULL,
+ #   recode_male = "M", recode_female = "F", recode_na = NA_character_)
   # Recode sex values in case non-standard codes are used (e.g., "M"/"F")
-  ped_recode <- BGmisc::recodeSex(ped, code_male = code_male)
+  ped_recode <- BGmisc::recodeSex(ped, code_male = code_male
+                                 # code_female = config$code_female,
+                                  #code_na = config$code_na
+                                 )
   if ("relation" %in% names(config) && !is.null(config$relation)) {
     # Construct a pedigree object to compute layout coordinates
     ped_ped <- pedigree(
