### 2025-12-15 d83f557
Rename and update Wars of the Roses dataset
Renamed all references from 'warofroses' to 'warsofroses' throughout the codebase, including data files, variables, and documentation. Fixed and updated entries in the CSV data, added missing URLs, corrected parent IDs, and improved documentation for the new dataset. Removed the old .rda and documentation files and added new ones for 'warsofroses'.
