+++ data-raw/df_warroses.R
@@ -33,7 +33,7 @@ df <- ped2fam(ged, personID = "id") %>%
  # )
 
 
-warofroses <- df %>%
+warsofroses <- df %>%
   select(-famID) %>%
   ped2fam(personID = "personID", famID = "famID") %>%
   rename(
@@ -41,7 +41,7 @@ warofroses <- df %>%
   )
 
 # checks
-df_repaired <- checkSex(warofroses,
+df_repaired <- checkSex(warsofroses,
   code_male = "M",
   code_female = "F",
   verbose = TRUE, repair = TRUE
@@ -95,13 +95,13 @@ checkis_acyclic <- checkPedigreeNetwork(df_repaired,
 checkis_acyclic
 if (checkis_acyclic$is_acyclic) {
   message("The pedigree is acyclic.")
-  write_csv(warofroses, here("data-raw", "warofroses.csv"))
-  usethis::use_data(warofroses, overwrite = TRUE, compress = "xz")
+  write_csv(warsofroses, here("data-raw", "warsofroses.csv"))
+  usethis::use_data(warsofroses, overwrite = TRUE, compress = "xz")
 } else {
   message("The pedigree contains cyclic relationships.")
 }
 
-warofroses %>%
+warsofroses %>%
   filter(is.na(momID) & is.na(dadID)) %>%
   select(id, name, famID, momID, dadID, sex) %>%
   mutate(
