### 2025-12-15 bde6421
Update NEWS.md
