### 2025-12-16 6510061
Update NEWS.md
