### 2025-12-16 2253856
Move and improve recodeMissingIDs; add early column check
Moved the recodeMissingIDs function from ggpedigree.R to ggpedigreeHelpers.R, enhanced it to handle multiple missing code values, and updated its documentation. Added an earlier check in ggPedigree to ensure required columns (personID, sex, dadID, momID) are present. Updated NEWS.md to reflect these changes and made minor editorial improvements.
