+++ R/ggpedigree.R
@@ -93,7 +93,9 @@ ggPedigree <- function(ped,
       stop("ped should be a data.frame or inherit to a data.frame")
     }
   }
-
+  if (!all(c(personID,dadID, momID, "sex") %in% names(ped))) {
+    stop("ped must contain personID, sex, dadID, and momID columns")
+  }
 
   if (interactive == TRUE &&
     requireNamespace("plotly", quietly = TRUE)) {
@@ -1556,74 +1558,3 @@ addTwins <- function(plotObject,
 }
 
 
-# doesn't work@
-recodeMissingIDs <- function(ped, momID = "momID", dadID = "dadID",
-                             personID = "personID",
-                             famID = "famID", matID = "matID", patID = "patID",
-                             missing_code_numeric = 0,
-                             missing_code_character = "0",
-                             config = list()) {
-  if (momID %in% names(ped)) {
-    if (any(ped[[momID]] == missing_code_numeric, na.rm = TRUE)) {
-      ped[[momID]][ped[[momID]] == missing_code_numeric] <- NA
-    }
-
-    if (any(ped[[momID]] == missing_code_character, na.rm = TRUE)) {
-      ped[[momID]][ped[[momID]] == missing_code_character] <- NA
-    }
-  }
-
-  if (dadID %in% names(ped)) {
-    if (any(ped[[dadID]] == missing_code_numeric, na.rm = TRUE)) {
-      ped[[dadID]][ped[[dadID]] == missing_code_numeric] <- NA
-    }
-    if (any(ped[[dadID]] == missing_code_character, na.rm = TRUE)) {
-      ped[[dadID]][ped[[dadID]] == missing_code_character] <- NA
-    }
-  }
-  if (personID %in% names(ped)) {
-    if (any(ped[[personID]] == missing_code_numeric, na.rm = TRUE)) {
-      ped[[personID]][ped[[personID]] == missing_code_numeric] <- NA
-    }
-    if (any(ped[[personID]] == missing_code_character, na.rm = TRUE)) {
-      ped[[personID]][ped[[personID]] == missing_code_character] <- NA
-    }
-  }
-
-  if (famID %in% names(ped)) {
-    if (any(ped[[famID]] == missing_code_numeric, na.rm = TRUE)) {
-      ped[[famID]][ped[[famID]] == missing_code_numeric] <- NA
-    }
-    if (any(ped[[famID]] == missing_code_character, na.rm = TRUE)) {
-      ped[[famID]][ped[[famID]] == missing_code_character] <- NA
-    }
-  }
-  if (matID %in% names(ped)) {
-    if (any(ped[[matID]] == missing_code_numeric, na.rm = TRUE)) {
-      ped[[matID]][ped[[matID]] == missing_code_numeric] <- NA
-    }
-    if (any(ped[[matID]] == missing_code_character, na.rm = TRUE)) {
-      ped[[matID]][ped[[matID]] == missing_code_character] <- NA
-    }
-  }
-  if (patID %in% names(ped)) {
-    if (any(ped[[patID]] == missing_code_numeric, na.rm = TRUE)) {
-      ped[[patID]][ped[[patID]] == missing_code_numeric] <- NA
-    }
-    if (any(ped[[patID]] == missing_code_character, na.rm = TRUE)) {
-      ped[[patID]][ped[[patID]] == missing_code_character] <- NA
-    }
-  }
-
-  # handle twinID as missing IDs
-  if ("twinID" %in% names(ped)) {
-    if (any(ped[["twinID"]] == missing_code_numeric, na.rm = TRUE)) {
-      ped[["twinID"]][ped[["twinID"]] == missing_code_numeric] <- NA
-    }
[…]
