### 2025-12-16 R-Computing-Lab/ggpedigree#94 @smasongarrison
I could definitely get really close to replicating the figure here probably by changing the default shape or by using the outline feature.

I've waffled about whether to include color in the default. I think that fundamentally my worry is that people might not realize that you can use color if its not the default. Let me think about a way to do it... because obviously black/white needs to be an easy option for journal figures.
