### 2025-12-16 #9434 @flyaflya
After reviewing the codebase against JOSS’s “substantial scholarly effort” criteria, it looks like a very strong fit: the project is sizeable (~7,098 LOC across 23 R files), mature (extensive release history + CRAN), well-tested (unit tests + CI/R-CMD-check + Codecov + vdiffr snapshots), and well-documented (man pages with examples, vignettes, pkgdown site, and included datasets). The implementation also appears highly consistent with the paper’s described architecture, algorithms, and feature set.

I was able to run the documentation examples as written. For context, I’m very familiar with R and ggplot-based visualization, but less familiar with genetics and the competing package ecosystem, so I appreciate having @magnusdv on the review team.

One small docs tweak: the GitHub README example should explicitly define the example dataset to avoid the common error `object 'potter' not found`. For readers unfamiliar with BGmisc, hitting that immediately could discourage them from continuing.

```
library(ggpedigree)
potter <- BGmisc::potter ## add this line to instructions to avoid error
ggPedigree(potter,
           famID = "famID",
           personID = "personID"
)
```
