### 2025-12-17 dda990c
Refactor test and example code formatting and logic
Wrapped ggpedigree plotting code in df_warroses.R with if (FALSE) to prevent execution during data processing. Improved formatting, argument alignment, and logic in test-ggpedigree_datainputs.R for better readability and consistency, including clearer handling of expect_warnings and test cases.
