+++ tests/testthat/test-ggpedigree_datainputs.R
@@ -1,11 +1,8 @@
-
 expect_roundtrip <- function(df,
                              cols = c("personID", "momID", "dadID", "sex"),
                              NA_id_value = NA,
-                             expect_warnings = FALSE
-                             ) {
-
-  if (expect_warnings) {
+                             expect_warnings = FALSE) {
+  if (expect_warnings == TRUE) {
     expect_warning(ggPedigree(df), NA)
     ped_df <- suppressWarnings(ggPedigree(df, config = list(debug = TRUE)))
     ped_df <- ped_df[["data"]]
@@ -20,75 +17,83 @@ expect_roundtrip <- function(df,
     df$dadID[df$dadID == NA_id_value] <- NA
   }
   expect_equal(ped_df[, cols, drop = FALSE],
-               df[, cols, drop = FALSE],
-               ignore_attr = TRUE)
+    df[, cols, drop = FALSE],
+    ignore_attr = TRUE
+  )
 
   expect_equal(nrow(ped_df), nrow(df))
 
-  expect_no_error(ggPedigree(df) %>% suppressWarnings())
+  if (expect_warnings == TRUE) {
+    expect_no_error(suppressWarnings(ggPedigree(df)))
+  } else {
+    expect_silent(ggPedigree(df))
+    expect_no_error(ggPedigree(df))
+  }
 }
 # still dealing with elegant handling of sex repair that doesn't break existing functionality
-case_char <- function(sex = c(0,1,0),
-                      missing_parent= NA_character_
-                      ) {
+case_char <- function(sex = c(0, 1, 0),
+                      missing_parent = NA_character_) {
   data.frame(
-  personID = c("1","2","3"),
-  momID = c(missing_parent, missing_parent, "1"),
-  dadID = c(missing_parent, missing_parent, "2"),
-  sex = sex
-)
+    personID = c("1", "2", "3"),
+    momID = c(missing_parent, missing_parent, "1"),
+    dadID = c(missing_parent, missing_parent, "2"),
+    sex = sex
+  )
 }
-case_num <- function(sex=c(0,1,0),
-                     missing_parent = NA_real_
-                     ) {
+case_num <- function(sex = c(0, 1, 0),
+                     missing_parent = NA_real_) {
   data.frame(
-  personID = 1:3,
-  dadID = c(missing_parent, missing_parent, 1),
-  momID = c(missing_parent, missing_parent, 2),
-  sex = sex
-)
+    personID = 1:3,
+    dadID = c(missing_parent, missing_parent, 1),
+    momID = c(missing_parent, missing_parent, 2),
+    sex = sex
+  )
 }
 
 
-test_that("char ids; 0 missingid; 1/2",  {
-  expect_roundtrip(case_char(c(2,1,2),"0"), NA_id_value = "0")
-#  Error in `pedigree(id = ped[[personID]], dadid = ped[[dadID]], momid = ped[[momID]], sex = ped_recode[[sexVar]])`: Id not male, but is a father: 2
+test_that("char ids; 0 missingid; 1/2", {
+  expect_roundtrip(case_char(c(2, 1, 2), "0"), NA_id_value = "0")
+  #  Error in `pedigree(id = ped[[personID]], dadid = ped[[dadID]], momid = ped[[momID]], sex = ped_recode[[sexVar]])`: Id not male, but is a father: 2
 })
[…]
