+++ reference/kinship2_shift.md
@@ -1,0 +1,14 @@
+# Shift a subject and his twins left or right among siblings used in autohint to resolve duplicate positions
+
+Shift a subject and his twins left or right among siblings used in
+autohint to resolve duplicate positions
+
+## Usage
+
+``` r
+kinship2_shift(id, sibs, goleft, hint, twinrel, twinset)
+```
+
+## Value
+
+Updated hint vector
