+++ reference/transformPed.md
@@ -16,7 +16,8 @@ transformPed(
   dadID = "dadID",
   matID = "matID",
   patID = "patID",
-  config = list(focal_fill_include = TRUE, focal_fill_component = "maternal"),
+  config = list(focal_fill_include = TRUE, focal_fill_component = "maternal",
+    recode_missing_ids = TRUE),
   fill_group_paternal = c("paternal", "patID", "paternal line", "paternal lineages",
     "paternal lines"),
   fill_group_maternal = c("maternal", "matID", "maternal line", "maternal lineages",
