### 2025-12-17 63ceeb2
Add code_unknown parameter to plot config and improve sex handling
Introduces a new code_unknown parameter to getDefaultPlotConfig for specifying unknown sex codes. Updates pedigree.sexrepair to provide clearer error and warning messages when unknown or invalid sex values are detected. Adjusts related test to account for the new parameter.
