### 2025-12-18 e6c386a
Update README.Rmd
change how data are loaded in the readme per great suggestion https://github.com/openjournals/joss-reviews/issues/9434#issuecomment-3662418907
