+++ README.Rmd
@@ -66,7 +66,9 @@ library(ggpedigree) # ggPedigree lives here
 library(BGmisc) # helper utilities & example data
 library(tidyverse)
 # if you don't have the most recent version of BGmisc, you may need to install it first as a stop-gap I've added the data loading here
-data("potter") # load example data from BGmisc
+
+potter <- BGmisc::potter ## avoid error
+
 if (!"twinID" %in% names(potter)) {
   # Add twinID and zygosity columns for demonstration purposes
   potter <- potter %>%
