### 2025-12-18 fca948e
Improve error handling for sex coding in pedigree input
Added more informative error messages when constructing pedigree objects fails due to incorrect or non-standard sex coding. Expanded and refactored tests for various sex code and ID input scenarios, including comprehensive cross-product testing of input types, missing values, and configuration options. Updated documentation to include new config option 'code_unknown'.
