+++ tests/testthat/test-ggpedigree_datainputs.R
@@ -1,18 +1,113 @@
-expect_roundtrip <- function(df,
-                             cols = c("personID", "momID", "dadID", "sex"),
-                             NA_id_value = NA,
-                             expect_warnings = FALSE,
-                             expect_errors = FALSE) {
-  if (expect_warnings == TRUE && expect_errors == FALSE) {
-    expect_warning(ggPedigree(df))
-    ped_df <- suppressWarnings(ggPedigree(df, config = list(debug = TRUE)))
-    ped_df <- ped_df[["data"]]
-  } else if (expect_warnings == FALSE && expect_errors == FALSE) {
-    expect_silent(ggPedigree(df))
-    ped_df <- ggPedigree(df, config = list(debug = TRUE))[["data"]]
-  } else if (expect_errors == TRUE) {
-    expect_error(ggPedigree(df))
+# Tests for ggPedigree data input handling
+sex_map <- list(
+  s212 = c(2, 1, 2),
+  s21NA = c(2, 1, NA),
+  s121 = c(1, 2, 1),
+  s12NA = c(1, 2, NA),
+  s101 = c(1, 0, 1),
+  s10NA = c(1, 0, NA),
+  s01NA = c(0, 1, NA),
+  s010 = c(0, 1, 0)
+)
+
+id_map_char <- list(
+  id_char = c("1", "2", "3"),
+  id_char_alpha = c("a", "b", "c")
+)
+id_map_num <- list(
+  id_num = c(1, 2, 3)
+)
+missing_parent_char <- list(
+  mp_char_0 = "0",
+  mp_char_empty = "",
+  mp_char_NA = NA_character_
+)
+
+missing_parent_num <- list(
+  mp_num_0   = 0,
+  mp_num_NA  = NA_real_
+)
+
+config_map <- list(
+  cfg_skip = "cfg_skip",
+  cfg_dbg = list(debug = TRUE),
+  cfg_m2 = list(code_male = 2),
+  cfg_m1 = list(code_male = 1),
+  cfg_m0 = list(code_male = 0)
+)
+
+# 1) Create a full cross product of "dimensions"
+grid_char <- expand.grid(
+  id_type = "char",
+  id_case = names(id_map_char),
+  missing_parent = names(missing_parent_char),
+  sex_case = names(sex_map),
+  config_case = names(config_map),
+  stringsAsFactors = FALSE
+)
+grid_num <- expand.grid(
+  id_type = "num",
+  id_case = names(id_map_num),
+  missing_parent = names(missing_parent_num),
+  sex_case = names(sex_map),
+  config_case = names(config_map),
+  stringsAsFactors = FALSE
+)
+
+# Combine character and numeric grids
+grid <- rbind(grid_char, grid_num)
+# 2) Define exceptions where warnings or errors are expected
+
+grid$expect_warnings <- FALSE
+grid$expect_errors <- FALSE
+
+grid <- grid %>% mutate(
+  expect_warnings = case_when(
+    sex_case %in% c("s12NA", "s10NA", "s01NA", "s21NA") ~ TRUE,
+    TRUE ~ expect_warnings
+  ),
+  expect_errors = case_when(
+    config_case %in% c("cfg_skip", "cfg_dbg", "cfg_m1") &
+      sex_case %in% c("s121", "s101", "s10NA", "s12NA") ~ TRUE,
+    # cfg_m2 fixes 1/2 but not 0/1
+    config_case == "cfg_m2" &
+      sex_case %in% c("s101", "s10NA", "s212", "s21NA", "s01NA", "s010") ~ TRUE,
+    # cfg_m0 fixes 0/1 but not 1/2
[…]
