+++ NEWS.md
@@ -6,11 +6,12 @@
 * Repair inconsistent handling of string IDs vs numeric IDs
 * Added rough cut of Wars of the Roses Pedigree Data Set
 * Required columns check much earlier in ggpedigree function, including sex
-
+* Added more informative error messages when constructing pedigree objects fails due to incorrect or non-standard sex coding. 
+* Expanded and refactored tests for various sex code and ID input scenarios, including comprehensive testing of input types, missing values, and configuration options. 
 
 # ggpedigree 1.0.0.1
 ## cran resubmission
-* Fixed a minor documentation issue with a url whose certificate had expired.
+* Fixed a minor documentation issue with a URL whose certificate had expired.
 
 # ggpedigree 1.0.0
 ## cran submission
