### 2025-12-18 578f26c
Adjust y-axis handling and spacing in pedigree plots
Sets scale_y_reverse limits to use minimum y position and normalizes y positions to start at zero in .adjustSpacing. Also updates test to clarify error handling for character IDs.
