### 2025-12-18 R-Computing-Lab/ggpedigree#94 @smasongarrison
Progress: Fixed the core issue underlying the weird top margin :) 578f26ce35d2429a77a969dcd0475c96316030b8

<img width="586" alt="Image" src="https://github.com/user-attachments/assets/118fe45d-0ce0-47f5-abe6-a051808866bd" /><img width="586"  alt="Image" src="https://github.com/user-attachments/assets/fbe35426-8f5c-4e88-b8e7-bf6fc5945e7b" />
