### 2025-12-19 0ba1f03
Add point scaling by pedigree size to plot config
Introduces a new option to scale point sizes based on pedigree size in plotting functions. Updates documentation, function signatures, and tests to support the new parameter and ensure correct behavior.
