### 2025-12-20 354ed00
Hide sex legend by default and improve point scaling
Set the sex legend to be hidden by default in ggpedigree plots and update the point size scaling logic to better handle different pedigree sizes. Adjusted tests to cover new scaling behavior and made minor improvements to the Game of Thrones dataset demonstration for clearer visualization.
