### 2025-12-20 f352570
Bump version to 1.1.0.0 and update NEWS.md
Updated the package version to 1.1.0.0 in DESCRIPTION and added a new section in NEWS.md to reflect recent changes and improvements.
