+++ R/ggpedigree.R
@@ -1,0 +1,167 @@
+#' Plot a custom pedigree diagram
+#'
+#' Generates a ggplot2-based diagram of a pedigree using custom coordinate layout,
+#' calculated relationship connections, and flexible styling via `config`.
+#' It processes the data using `ped2fam()`. This function
+#' supports multiple families and optionally displays affected status and sex-based color/shape.
+#'
+#' @param ped A data frame containing the pedigree data. Needs personID, momID, and dadID columns
+#' @param famID Character string specifying the column name for family IDs. Defaults to "famID".
+#' @param personID Character string specifying the column name for individual IDs. Defaults to "personID".
+#' @param momID Character string specifying the column name for mother IDs. Defaults to "momID".
+#' @param dadID Character string specifying the column name for father IDs. Defaults to "dadID".
+#' @param spouseID Character string specifying the column name for spouse IDs. Defaults to "spouseID".
+#' @param matID Character string specifying the column name for maternal lines Defaults to "matID".
+#' @param patID Character string specifying the column name for paternal lines Defaults to "patID".
+#' @param twinID Character string specifying the column name for twin IDs. Defaults to "twinID".
+#' @param status_column Character string specifying the column name for affected status. Defaults to NULL.
+#' @param debug Logical. If TRUE, prints debugging information. Default: FALSE.
+#' @param hints Data frame with hints for layout adjustments. Default: NULL.
+#' @param interactive Logical. If TRUE, generates an interactive plot using `plotly`. Default: FALSE.
+#' @param overlay_column Character string specifying the column name for overlay alpha values.
+#' @param tooltip_columns Character vector of column names to show when hovering.
+#'        Defaults to c("personID", "sex").  Additional columns present in `ped`
+#'        can be supplied – they will be added to the Plotly tooltip text.
+#'        Defaults to NULL, which uses the default tooltip columns.
+#' @param return_widget Logical; if TRUE (default) returns a plotly htmlwidget.
+#'        If FALSE, returns the underlying plotly object (useful for further
+#'        customization before printing).
+#' @param focal_fill_column Character string specifying the column name for focal fill color.
+#' @param phantoms Logical. If TRUE, adds phantom parents for individuals without parents.
+#' @param ... Additional arguments passed to `ggplot2` functions.
+#' @param config A list of configuration options for customizing the plot.
+#'        See getDefaultPlotConfig for details. The list can include:
+#'  \describe{
+#'     \item{code_male}{Integer or string. Value identifying males in the sex column. (typically 0 or 1) Default: 1.}
+#'     \item{segment_spouse_color, segment_self_color}{Character. Line colors for respective connection types.}
+#'     \item{segment_sibling_color, segment_parent_color, segment_offspring_color}{Character. Line colors for respective connection types.}
+#'     \item{label_text_size, point_size, segment_linewidth}{Numeric. Controls text size, point size, and line thickness.}
+#'     \item{generation_height}{Numeric. Vertical spacing multiplier between generations. Default: 1.}
+#'     \item{shape_unknown, shape_female, shape_male, status_shape_affected}{Integers. Shape codes for plotting each group.}
+#'     \item{sex_shape_labels}{Character vector of labels for the sex variable. (default: c("Female", "Male", "Unknown"))}
+#'     \item{unaffected, affected}{Values indicating unaffected/affected status.}
+#'     \item{sex_color_include}{Logical. If TRUE, uses color to differentiate sex.}
[…]
