### 2025-12-21 4592d37
split core into core and helpers
