+++ R/ggpedigreeCoreHelpers.R
@@ -1,0 +1,420 @@
+#' @title Prepare Pedigree Data
+#' @description
+#' This function checks and prepares the pedigree data frame for use in ggPedigree.
+#'
+#' @inheritParams ggPedigree
+#' @inheritParams transformPed
+#' @param fill_group_family A character vector specifying the family fill group names.
+#' @keywords internal
+#' @return A data frame with the prepared pedigree data.
+#'
+#'
+preparePedigreeData <- function(ped,
+                                famID = "famID",
+                                personID = "personID",
+                                momID = "momID",
+                                dadID = "dadID",
+                                matID = "matID",
+                                patID = "patID",
+                                config = list(
+                                  focal_fill_include = TRUE,
+                                  focal_fill_component = "maternal",
+                                  recode_missing_ids = TRUE
+                                ),
+                                fill_group_paternal = c(
+                                  "paternal",
+                                  "patID",
+                                  "paternal line",
+                                  "paternal lineages",
+                                  "paternal lines"
+                                ),
+                                fill_group_maternal = c(
+                                  "maternal",
+                                  "matID",
+                                  "maternal line",
+                                  "maternal lineages",
+                                  "maternal lines"
+                                ),
+                                fill_group_family = c(
+                                  "family",
+                                  "famID",
+                                  "family line",
+                                  "family lineages",
+                                  "family lines"
+                                ),
+                                status_column = NULL,
+                                phantoms = FALSE,
+                                focal_fill_column = NULL) {
+  # -----
+  # STEP 2: Pedigree Data Transformation
+  # -----
+
+  # Transform the pedigree data frame to include family, paternal, and maternal IDs
+  ds_ped <- transformPed(
+    ped = ped,
+    famID = famID,
+    personID = personID,
+    momID = momID,
+    dadID = dadID,
+    matID = matID,
+    patID = patID,
+    config = config,
+    fill_group_paternal = fill_group_paternal,
+    fill_group_maternal = fill_group_maternal
+  )
+
+  # ----
+  # STEP 3: Data Cleaning and Recoding
+  # ----
+
+  # Recode affected status into factor, if applicable
+
+  if (!is.null(status_column)) {
+    ds_ped[[status_column]] <- factor(ds_ped[[status_column]],
+      levels = config$status_codes,
+      labels = config$status_labels
+    )
+  }
+
+  # Standardize sex variable using code_male convention
+
+  ds_ped <- BGmisc::recodeSex(ds_ped,
+    recode_male = config$code_male,
+    recode_na = config$code_na,
+    recode_female = config$code_female
+  )
+
+  if (phantoms == TRUE) {
+    # If phantoms are requested, add phantom parents
+    ds_ped <- BGmisc::checkParentIDs(
+      ds_ped,
+      addphantoms = TRUE,
+      repair = TRUE,
+      parentswithoutrow = FALSE,
+      repairsex = FALSE,
+      personID = personID,
+      momID = momID,
+      dadID = dadID,
+      famID = famID
+    )
+  }
+
+  # Add focal fill column if specified
+
+  ds_ped <- addFocalFillColumn(
+    ds_ped = ds_ped,
+    config = config,
+    focal_fill_column = focal_fill_column,
+    famID = famID,
+    matID = matID,
+    patID = patID,
+    personID = personID,
+    fill_group_family = fill_group_family,
+    fill_group_maternal = fill_group_maternal,
[…]
