### 2025-12-21 3a93dcf
Refactor addTwins to internal .addTwins and update docs
Renamed addTwins to .addTwins for internal use and updated its export. Adjusted ggPedigree.core to use .addTwins. Updated documentation to reference ggpedigreeCore.R and renamed addTwins.Rd to dot-addTwins.Rd, reflecting the internal function naming. Also updated documentation source references for several functions to ggpedigreeCoreHelpers.R and ggpedigreeCore.R.
