+++ man/transformPed.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/ggpedigree.R
+% Please edit documentation in R/ggpedigreeCoreHelpers.R
 \name{transformPed}
 \alias{transformPed}
 \title{Process Pedigree Data}
