+++ man/ggPedigree.core.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/ggpedigree.R
+% Please edit documentation in R/ggpedigreeCore.R
 \name{ggPedigree.core}
 \alias{ggPedigree.core}
 \title{Core Function for ggPedigree}
