+++ man/dot-addOverlay.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/ggpedigree.R
+% Please edit documentation in R/ggpedigreeCore.R
 \name{.addOverlay}
 \alias{.addOverlay}
 \alias{addOverlay}
