+++ man/dot-addTwins.Rd
@@ -1,9 +1,18 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/ggpedigree.R
-\name{addTwins}
+% Please edit documentation in R/ggpedigreeCore.R
+\name{.addTwins}
+\alias{.addTwins}
 \alias{addTwins}
 \title{Add Twins to ggplot Pedigree Plot}
 \usage{
+.addTwins(
+  plotObject,
+  connections,
+  config,
+  plot_connections,
+  personID = "personID"
+)
+
 addTwins(
   plotObject,
   connections,
