### 2025-12-21 6f1a61f
Document default values in plot config parameters
Updated Roxygen and Rd documentation for getDefaultPlotConfig to specify default values for many function parameters. This improves clarity for users by making default behaviors explicit in the documentation.
