+++ R/defaultPlotConfig.R
@@ -22,10 +22,10 @@
 #' @param plot_title Main title of the plot.
 #' @param plot_subtitle Subtitle of the plot.
 #' @param value_rounding_digits Number of digits to round displayed values.
-#' @param code_male Integer/string code for males in data.
-#' @param code_na optional Integer/string code for missing values in data.
-#' @param code_female  optional Integer/string code for females in data.
-#' @param code_unknown  optional Integer/string code for unknown
+#' @param code_male Integer/string code for males in data. Default is 1.
+#' @param code_na optional Integer/string code for missing values in data. Default is NA.
+#' @param code_female  optional Integer/string code for females in data. Default is 0.
+#' @param code_unknown  optional Integer/string code for unknown sex in data. Default is NULL.
 #' @param filter_n_pairs Threshold to filter maximum number of pairs.
 #' @param filter_degree_min Minimum degree value used in filtering.
 #' @param filter_degree_max Maximum degree value used in filtering.
@@ -73,26 +73,26 @@
 #' @param ped_packed Whether the pedigree should use packed layout.
 #' @param ped_align Whether to align pedigree generations.
 #' @param ped_width Plot width of the pedigree block.
-#' @param segment_linewidth Line width for segments.
-#' @param segment_linetype Line type for segments.
-#' @param segment_lineend Line end type for segments.
-#' @param segment_linejoin Line join type for segments.
-#' @param segment_offspring_color Color for offspring segments.
-#' @param segment_parent_color Color for parent segments.
-#' @param segment_self_color Color for self-loop segments.
-#' @param segment_sibling_color Color for sibling segments.
-#' @param segment_spouse_color Color for spouse segments.
-#' @param segment_mz_color Color for monozygotic twin segments.
-#' @param segment_mz_linetype Line type for MZ segments.
-#' @param segment_mz_alpha Alpha for MZ segments.
-#' @param segment_mz_t Tuning parameter for MZ segment layout.
-#' @param segment_self_linetype Line type for self-loop segments.
-#' @param segment_self_alpha Alpha value for self-loop segments.
-#' @param segment_self_angle Angle of self-loop segment.
-#' @param segment_self_curvature Curvature of self-loop segment.
-#' @param segment_self_linewidth Width of self-loop segment lines.
-#' @param sex_color_include Whether to color nodes by sex.
-#' @param sex_color_palette A character vector of colors for sex.
+#' @param segment_linewidth Line width for segments. Default is 0.5.
+#' @param segment_linetype Line type for segments. Default is 1 (solid).
+#' @param segment_lineend Line end type for segments. Default is "round".
+#' @param segment_linejoin Line join type for segments. Default is "round".
+#' @param segment_offspring_color Color for offspring segments. Default uses segment_default_color.
+#' @param segment_parent_color Color for parent segments. Default uses segment_default_color.
+#' @param segment_self_color Color for self-loop segments. Default uses segment_default_color.
+#' @param segment_sibling_color Color for sibling segments. Default uses segment_default_color.
+#' @param segment_spouse_color Color for spouse segments. Default uses segment_default_color.
+#' @param segment_mz_color Color for monozygotic twin segments. Default uses segment_default_color.
+#' @param segment_mz_linetype Line type for MZ segments. Default uses segment_linetype.
+#' @param segment_mz_alpha Alpha for MZ segments. Default is 1.
+#' @param segment_mz_t Tuning parameter for MZ segment layout. Default is 0.6.
+#' @param segment_self_linetype Line type for self-loop segments. Default is "dotdash".
[…]
