+++ man/getDefaultPlotConfig.Rd
@@ -213,13 +213,13 @@ getDefaultPlotConfig(
 
 \item{value_rounding_digits}{Number of digits to round displayed values.}
 
-\item{code_male}{Integer/string code for males in data.}
+\item{code_male}{Integer/string code for males in data. Default is 1.}
 
-\item{code_na}{optional Integer/string code for missing values in data.}
+\item{code_na}{optional Integer/string code for missing values in data. Default is NA.}
 
-\item{code_unknown}{optional Integer/string code for unknown}
+\item{code_unknown}{optional Integer/string code for unknown sex in data. Default is NULL.}
 
-\item{code_female}{optional Integer/string code for females in data.}
+\item{code_female}{optional Integer/string code for females in data. Default is 0.}
 
 \item{label_include}{Whether to display labels on plot points.}
 
@@ -287,49 +287,49 @@ getDefaultPlotConfig(
 
 \item{ped_width}{Plot width of the pedigree block.}
 
-\item{segment_linewidth}{Line width for segments.}
+\item{segment_linewidth}{Line width for segments. Default is 0.5.}
 
-\item{segment_linetype}{Line type for segments.}
+\item{segment_linetype}{Line type for segments. Default is 1 (solid).}
 
-\item{segment_lineend}{Line end type for segments.}
+\item{segment_lineend}{Line end type for segments. Default is "round".}
 
-\item{segment_linejoin}{Line join type for segments.}
+\item{segment_linejoin}{Line join type for segments. Default is "round".}
 
-\item{segment_offspring_color}{Color for offspring segments.}
+\item{segment_offspring_color}{Color for offspring segments. Default uses segment_default_color.}
 
-\item{segment_parent_color}{Color for parent segments.}
+\item{segment_parent_color}{Color for parent segments. Default uses segment_default_color.}
 
-\item{segment_self_color}{Color for self-loop segments.}
+\item{segment_self_color}{Color for self-loop segments. Default uses segment_default_color.}
 
-\item{segment_sibling_color}{Color for sibling segments.}
+\item{segment_sibling_color}{Color for sibling segments. Default uses segment_default_color.}
 
-\item{segment_spouse_color}{Color for spouse segments.}
+\item{segment_spouse_color}{Color for spouse segments. Default uses segment_default_color.}
 
-\item{segment_mz_color}{Color for monozygotic twin segments.}
+\item{segment_mz_color}{Color for monozygotic twin segments. Default uses segment_default_color.}
 
-\item{segment_mz_linetype}{Line type for MZ segments.}
+\item{segment_mz_linetype}{Line type for MZ segments. Default uses segment_linetype.}
 
-\item{segment_mz_alpha}{Alpha for MZ segments.}
+\item{segment_mz_alpha}{Alpha for MZ segments. Default is 1.}
 
-\item{segment_mz_t}{Tuning parameter for MZ segment layout.}
+\item{segment_mz_t}{Tuning parameter for MZ segment layout. Default is 0.6.}
 
-\item{segment_self_linetype}{Line type for self-loop segments.}
+\item{segment_self_linetype}{Line type for self-loop segments. Default is "dotdash".}
 
-\item{segment_self_linewidth}{Width of self-loop segment lines.}
+\item{segment_self_linewidth}{Width of self-loop segment lines. Default is half of segment_linewidth.}
 
-\item{segment_self_alpha}{Alpha value for self-loop segments.}
[…]
