+++ man/buildPlotConfig.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/defaultPlotConfig.R
+% Please edit documentation in R/buildPlotConfig.R
 \name{buildPlotConfig}
 \alias{buildPlotConfig}
 \title{build Config}
