### 2025-12-21 f86b540
Switch data import from Excel to CSV files
Replaced read_excel with read.csv for Pedigree, LRS, and ARS datasets in df_redsquirrels.R, reflecting a move to CSV data sources. Also removed readxl from DESCRIPTION 'Suggests' as it is no longer required.
