### 2025-12-21 cd4ad5a
eliminate reshape2 dependency
