+++ R/ggRelatednessMatrix.R
@@ -167,7 +167,19 @@ ggRelatednessMatrix.core <- function(
     mat_plot[upper.tri(mat_plot)] <- NA
   }
   # Melt the matrix into long format for ggplot2
-  df_melted <- reshape2::melt(mat_plot)
+  # df_melted <- r e s h a p e 2 :: m e l t(mat_plot)
+  rnames <- rownames(mat_plot)
+  cnames <- colnames(mat_plot)
+
+  # Handle case where matrix has no row/column names
+  if (is.null(rnames)) {
+    rnames <- seq_len(nrow(mat_plot))
+  }
+  if (is.null(cnames)) {
+    cnames <- seq_len(ncol(mat_plot))
+  }
+  df_melted <- expand.grid(ID1 = rnames, ID2 = cnames, stringsAsFactors = FALSE)
+  df_melted$value <- as.vector(mat_plot)
 
   colnames(df_melted) <- c("ID1", "ID2", "value")
 
