+++ .github/workflows/R-CMD-devcheck.yaml
@@ -31,24 +31,29 @@ jobs:
       R_KEEP_PKG_SOURCE: yes
 
     steps:
-      - uses: actions/checkout@v3
+      - name: Checkout code
+        uses: actions/checkout@v3
 
-      - uses: r-lib/actions/setup-pandoc@v2
+      - name: Install Pandoc for vignette building
+        uses: r-lib/actions/setup-pandoc@v2
 
-      - uses: r-lib/actions/setup-r@v2
+      - name: Set up R
+        uses: r-lib/actions/setup-r@v2
         with:
           r-version: ${{ matrix.config.r }}
           rtools-version: ${{ matrix.config.rtools }}
           http-user-agent: ${{ matrix.config.http-user-agent }}
           use-public-rspm: true
 
-      - uses: r-lib/actions/setup-r-dependencies@v2
+      - name: Set up R dependencies
+        uses: r-lib/actions/setup-r-dependencies@v2
         with:
-          cache-version: 3
+          cache-version: 4
           extra-packages: |
             any::rcmdcheck
           upgrade: 'TRUE'
           needs: check
-      - uses: r-lib/actions/check-r-package@v2
+      - name: Check R Package
+        uses: r-lib/actions/check-r-package@v2
         with:
           upload-snapshots: true
