### 2025-12-21 0dadb1c
Refactor phantom parent logic to config option
Removed the 'phantoms' argument from ggPedigree, ggPedigreeInteractive, ggPedigree.core, and preparePedigreeData, replacing it with a new 'add_phantoms' option in the config list. Updated documentation and tests to reflect this change, allowing phantom parent addition to be controlled via configuration rather than function arguments.
