+++ reference/ggPedigree.core.md
@@ -24,7 +24,6 @@ ggPedigree.core(
   debug = FALSE,
   hints = NULL,
   function_name = "ggPedigree",
-  phantoms = FALSE,
   ...
 )
 ```
@@ -92,12 +91,12 @@ ggPedigree.core(
 - config:
 
   A list of configuration options for customizing the plot. See
-  getDefaultPlotConfig for details. The list can include:
+  getDefaultPlotConfig for details of each option. The list can include:
 
   code_male
 
   :   Integer or string. Value identifying males in the sex column.
-      (typically 0 or 1) Default: 1.
+      (typically 0 or 1) Default: 1
 
   segment_spouse_color, segment_self_color
 
@@ -148,11 +147,6 @@ ggPedigree.core(
 - hints:
 
   Data frame with hints for layout adjustments. Default: NULL.
-
-- phantoms:
-
-  Logical. If TRUE, adds phantom parents for individuals without
-  parents.
 
 - ...:
 
