+++ reference/getDefaultPlotConfig.md
@@ -30,6 +30,7 @@ getDefaultPlotConfig(
   value_rounding_digits = 5,
   code_male = 1,
   code_na = NA,
+  code_unknown = NULL,
   code_female = 0,
   label_include = TRUE,
   label_column = "personID",
@@ -44,6 +45,7 @@ getDefaultPlotConfig(
   label_text_color = "black",
   label_text_family = "sans",
   point_size = 4,
+  point_scale_by_pedigree = TRUE,
   outline_include = FALSE,
   outline_multiplier = 1.25,
   outline_additional_size = 0,
@@ -88,8 +90,9 @@ getDefaultPlotConfig(
   sex_shape_female = 16,
   sex_shape_male = 15,
   sex_shape_unknown = 18,
+  sex_shape_values = NULL,
   sex_shape_include = TRUE,
-  sex_legend_show = TRUE,
+  sex_legend_show = FALSE,
   status_include = TRUE,
   status_code_affected = 1,
   status_code_unaffected = 0,
@@ -178,6 +181,7 @@ getDefaultPlotConfig(
   override_many2many = FALSE,
   optimize_plotly = TRUE,
   recode_missing_ids = TRUE,
+  add_phantoms = FALSE,
   ...
 )
 ```
@@ -254,15 +258,20 @@ getDefaultPlotConfig(
 
 - code_male:
 
-  Integer/string code for males in data.
+  Integer/string code for males in data. Default is 1.
 
 - code_na:
 
-  optional Integer/string code for missing values in data.
+  optional Integer/string code for missing values in data. Default is
+  NA.
+
+- code_unknown:
+
+  optional Integer/string code for unknown sex in data. Default is NULL.
 
 - code_female:
 
-  optional Integer/string code for females in data.
+  optional Integer/string code for females in data. Default is 0.
 
 - label_include:
 
@@ -317,6 +326,10 @@ getDefaultPlotConfig(
 - point_size:
 
   Size of points drawn in plot.
+
+- point_scale_by_pedigree:
+
+  Whether to scale point sizes by pedigree size.
 
 - outline_include:
 
@@ -396,79 +409,81 @@ getDefaultPlotConfig(
 
 - segment_linewidth:
 
-  Line width for segments.
+  Line width for segments. Default is 0.5.
 
 - segment_linetype:
 
-  Line type for segments.
+  Line type for segments. Default is 1 (solid).
 
 - segment_lineend:
 
-  Line end type for segments.
+  Line end type for segments. Default is "round".
 
 - segment_linejoin:
 
-  Line join type for segments.
+  Line join type for segments. Default is "round".
 
 - segment_offspring_color:
 
-  Color for offspring segments.
+  Color for offspring segments. Default uses segment_default_color.
 
 - segment_parent_color:
 
-  Color for parent segments.
+  Color for parent segments. Default uses segment_default_color.
 
 - segment_self_color:
 
-  Color for self-loop segments.
+  Color for self-loop segments. Default uses segment_default_color.
 
 - segment_sibling_color:
 
-  Color for sibling segments.
+  Color for sibling segments. Default uses segment_default_color.
 
 - segment_spouse_color:
 
-  Color for spouse segments.
+  Color for spouse segments. Default uses segment_default_color.
 
 - segment_mz_color:
 
-  Color for monozygotic twin segments.
+  Color for monozygotic twin segments. Default uses
+  segment_default_color.
 
 - segment_mz_linetype:
 
-  Line type for MZ segments.
+  Line type for MZ segments. Default uses segment_linetype.
 
[…]
