+++ reference/ggPedigree.md
@@ -28,7 +28,6 @@ ggPedigree(
   debug = FALSE,
   hints = NULL,
   interactive = FALSE,
-  phantoms = FALSE,
   ...
 )
 
@@ -51,7 +50,6 @@ ggpedigree(
   debug = FALSE,
   hints = NULL,
   interactive = FALSE,
-  phantoms = FALSE,
   ...
 )
 ```
@@ -132,12 +130,12 @@ ggpedigree(
 - config:
 
   A list of configuration options for customizing the plot. See
-  getDefaultPlotConfig for details. The list can include:
+  getDefaultPlotConfig for details of each option. The list can include:
 
   code_male
 
   :   Integer or string. Value identifying males in the sex column.
-      (typically 0 or 1) Default: 1.
+      (typically 0 or 1) Default: 1
 
   segment_spouse_color, segment_self_color
 
@@ -193,11 +191,6 @@ ggpedigree(
 
   Logical. If TRUE, generates an interactive plot using \`plotly\`.
   Default: FALSE.
-
-- phantoms:
-
-  Logical. If TRUE, adds phantom parents for individuals without
-  parents.
 
 - ...:
 
