### 2025-12-21 f257db9
Increase default point size and update related docs/tests
Raised the default `point_size` from 4 to 5 in `getDefaultPlotConfig` and updated documentation, tests, and vignettes to reflect this change. Also added explicit `point_scale_by_pedigree = FALSE` in several vignette examples for clarity, and made minor adjustments to vignette text and plot settings.
