+++ vignettes/articles/v22_plots_morecomplexity.Rmd
@@ -84,7 +84,7 @@ df_repaired <- checkParentIDs(df_got,
   parentswithoutrow = FALSE,
   repairsex = FALSE
 ) %>% mutate(
-  famID_mulit =famID,
+  famID_mulit = famID,
   famID = 1,
   affected = case_when(
     ID %in% c(jon_id, dany_id, "365") ~ 1,
@@ -115,7 +115,8 @@ Here is the same pedigree, but using `ggPedigree()` from {ggpedigree}. This func
 
 ```{r message=FALSE, warning=FALSE}
 df_repaired <- df_repaired %>% mutate(
-  famID =famID_mulit)
+  famID = famID_mulit
+)
 
 pltstatic <- ggPedigree(df_repaired,
   status_column = "affected",
@@ -141,7 +142,7 @@ pltstatic <- ggPedigree(df_repaired,
 
 pltstatic
 
-#pltstatic+ facet_wrap(~famID_mulit, drop=TRUE,scales = "free")
+# pltstatic+ facet_wrap(~famID_mulit, drop=TRUE,scales = "free")
 ```
 
 
@@ -160,7 +161,7 @@ plt <- ggPedigreeInteractive(df_repaired_renamed,
     focal_fill_include = TRUE,
     code_male = "M",
     point_size = 10,
-        segment_linewidth = 0.25,
+    segment_linewidth = 0.25,
     status_include = FALSE,
     status_label_affected = 1,
     status_shape_affected = 4,
