### 2025-12-23 07949e2
Add segment scaling by pedigree and update defaults
Introduces segment_scale_by_pedigree to allow segment linewidths to scale with pedigree size, similar to point scaling. Updates default point_size to 6 and segment_linewidth to 0.75 in documentation and vignettes. Adjusts buildPlotConfig logic and tests to support new scaling behavior.
