+++ R/defaultPlotConfig.R
@@ -226,7 +226,7 @@ getDefaultPlotConfig <- function(function_name = "getDefaultPlotConfig",
                                  label_text_color = "black",
                                  label_text_family = "sans",
                                  # --- POINT / OUTLINE AESTHETICS ---------------------------------------
-                                 point_size = 5,
+                                 point_size = 6,
                                  point_scale_by_pedigree = TRUE,
                                  outline_include = FALSE,
                                  outline_multiplier = 1.25,
@@ -255,6 +255,7 @@ getDefaultPlotConfig <- function(function_name = "getDefaultPlotConfig",
                                  segment_linetype = 1,
                                  segment_lineend = "round",
                                  segment_linejoin = "round",
+                                 segment_scale_by_pedigree = FALSE,
                                  segment_offspring_color = segment_default_color,
                                  segment_parent_color = segment_default_color,
                                  segment_self_color = segment_default_color,
@@ -502,6 +503,7 @@ getDefaultPlotConfig <- function(function_name = "getDefaultPlotConfig",
     segment_linetype = segment_linetype,
     segment_lineend = segment_lineend,
     segment_linejoin = segment_linejoin,
+    segment_scale_by_pedigree = segment_scale_by_pedigree,
     segment_offspring_color = ifelse(segment_default_color == "black", segment_offspring_color, segment_default_color),
     segment_parent_color = ifelse(segment_default_color == "black", segment_parent_color, segment_default_color),
     segment_self_color = ifelse(segment_default_color == "black", segment_self_color, segment_default_color),
