### 2025-12-24 151943a
Update plot config docs and vignette examples
Added documentation for the new `segment_scale_by_pedigree` parameter and clarified default values for sex shape parameters in R and Rd files. Updated vignette plot examples to use new default values for `segment_linewidth`, `point_size`, and `sex_legend_show`. Minor HTML vignette style and code block adjustments.
