### 2025-12-24 R-Computing-Lab/ggpedigree#94 @smasongarrison
I've got a scaling approach working... in theory 😂, using the total pedigree size to scale the dot size.

Combined with a larger default dot size, I think it works ok under most circumstances.

It kicks in after 50 people; otherwise ... it looks very silly.
