### 2025-12-24 aeb7661
add examples
