+++ man/ASOIAF.Rd
@@ -28,13 +28,36 @@ dataset is not intended to represent any real individuals or families.
 The variables are as follows:
 \itemize{
   \item \code{id}:  Person identification variable
+  \item \code{famID}:  Family identification variable
   \item \code{momID}:  ID of the mother
   \item \code{dadID}:  ID of the father
   \item \code{name}:  Name of the person
-  \item \code{sex}: Biological sex
+  \item \code{sex}: Biological sex (M/F)
+  \item \code{url}:  URL to a wiki page about the character
   \item \code{twinID}:  ID of the twin, if applicable
   \item \code{zygosity}: Zygosity of the twin, if applicable. mz is monozygotic; dz is dizygotic
-  \item \code{url}:  URL to a wiki page about the character
   }
+}
+\examples{
+# Load the ASOIAF dataset
+data(ASOIAF)
+
+# View the structure of the dataset
+str(ASOIAF)
+
+# Plot a pedigree for a specific family
+if (requireNamespace("ggplot2", quietly = TRUE)) {
+  # Create a pedigree plot for House Stark
+  ggPedigree(ASOIAF,
+    famID = "famID",
+    personID = "id",
+    momID = "momID",
+    dadID = "dadID",
+    config = list(
+      add_phantoms = TRUE,
+      code_male = "M"
+    )
+  )
+}
 }
 \keyword{datasets}
