### 2025-12-24 c035be6
add examples
