### 2025-12-25 36e52a8
Update ASOIAF example to use House Tarth subset
Modified the example code in both R/documentData.R and man/ASOIAF.Rd to filter the ASOIAF dataset for House Tarth (famID == 26) and updated the pedigree plot example accordingly. This provides a more specific and relevant demonstration for users.
