+++ R/calcCoordinates.R
@@ -37,9 +37,7 @@ utils::globalVariables(c(":="))
 #'   personID = "personID",
 #'   momID = "momID",
 #'   dadID = "dadID",
-#'   config = list(
-#'     code_male = 1
-#'   )
+#'   code_male = 1
 #' )
 #'
 #' # View the coordinates
@@ -51,11 +49,13 @@ utils::globalVariables(c(":="))
 #'   personID = "personID",
 #'   momID = "momID",
 #'   dadID = "dadID",
+#'   code_male = 1,
 #'   config = list(
 #'     ped_packed = FALSE,
 #'     ped_width = 20
 #'   )
 #' )
+
 calculateCoordinates <- function(ped,
                                  personID = "personID",
                                  momID = "momID",
