### 2025-12-25 0caf09e
Update redsquirrels example to include config argument
Added a 'config' list argument with 'add_phantoms' and 'code_male' options to the redsquirrels example in both documentation and Rd file for improved clarity on function usage.
