+++ man/redsquirrels.Rd
@@ -60,7 +60,11 @@ if (requireNamespace("ggplot2", quietly = TRUE)) {
     personID = "personID",
     momID = "momID",
     dadID = "dadID",
-    sex = "sex"
+    sex = "sex",
+    config = list(
+      add_phantoms = TRUE,
+      code_male = "M"
+  )
   )
 }
 }
