### 2025-12-25 0b809e6
Improve error and warning messages for pedigree construction
Enhanced the error and warning messages in calcCoordinates.R to include underlying error details when pedigree construction fails, providing more informative feedback to users. Updated documentation examples in documentData.R and redsquirrels.Rd to reflect changes in config usage.
