+++ R/calcCoordinates.R
@@ -302,8 +302,13 @@ alignPedigreeWithRelations <- function(ped,
         relation = config$relation
       ),
       error = function(e) {
-        stop("Error in constructing pedigree object. Please check that you've
-           correctly specified the sex of individuals. Setting code_male may help if non-standard codes are used (e.g., 'M'/'F'; '1,2').")
+        stop(
+          "Error in constructing pedigree object. Please check that you've ",
+          "correctly specified the sex of individuals. Setting code_male may help ",
+          "if non-standard codes are used (e.g., 'M'/'F'; '1,2').\n\n",
+          "Underlying error: ", conditionMessage(e),
+          call. = FALSE
+        )
       }
     )
   } else {
@@ -315,8 +320,13 @@ alignPedigreeWithRelations <- function(ped,
         sex = ped_recode[[sexVar]]
       ),
       error = function(e) {
-        stop("Error in constructing pedigree object. Please check that you've
-           correctly specified the sex of individuals. Setting code_male may help if non-standard codes are used (e.g., 'M'/'F'; '1,2').")
+        stop(
+          "Error in constructing pedigree object. Please check that you've ",
+          "correctly specified the sex of individuals. Setting code_male may help ",
+          "if non-standard codes are used (e.g., 'M'/'F'; '1,2').\n\n",
+          "Underlying error: ", conditionMessage(e),
+          call. = FALSE
+        )
       }
     )
   }
@@ -343,8 +353,12 @@ alignPedigreeWithHints <- function(ped_ped, config) {
         packed = config$ped_packed
       ),
       error = function(e) {
-        warning("Your hints caused an error and were not used.
-                Using default hints instead.")
+        warning(
+          "Your hints caused an error and were not used.\n",
+          "Using default hints instead.\n\n",
+          "Underlying error from kinship2_autohint(): ", conditionMessage(e),
+          call. = FALSE
+        )
         kinship2_autohint(ped_ped,
           align = config$ped_align,
           packed = config$ped_packed
