+++ R/documentData.R
@@ -52,9 +52,10 @@
 #'     momID = "momID",
 #'     dadID = "dadID",
 #'     sex = "sex",
+#'     code_male="M",
 #'     config = list(
-#'       add_phantoms = TRUE,
-#'       code_male = "M"
+#'       add_phantoms = TRUE#,
+#'     #  code_male = "M"
 #'   )
 #'   )
 #' }
