+++ man/redsquirrels.Rd
@@ -61,9 +61,10 @@ if (requireNamespace("ggplot2", quietly = TRUE)) {
     momID = "momID",
     dadID = "dadID",
     sex = "sex",
+    code_male="M",
     config = list(
-      add_phantoms = TRUE,
-      code_male = "M"
+      add_phantoms = TRUE#,
+    #  code_male = "M"
   )
   )
 }
