### 2025-12-25 c3d83c7
Add code_male parameter for male identification
Introduces a new code_male parameter to ggPedigree, ggPedigreeInteractive, and ggPedigree.core functions, allowing users to specify the value identifying males in the sex column. Updates documentation and internal handling to support this parameter, improving flexibility for datasets with different male coding schemes.
