### 2025-12-25 4f7bee5
Add sexVar argument and improve sex code handling
Introduces a new 'sexVar' argument to ggPedigree, ggPedigreeInteractive, and ggPedigree.core functions, allowing users to specify the column name for sex coding. Also updates helper functions to accept explicit male, female, and NA codes, and adjusts internal checks to use the configurable sex variable. This enhances flexibility for datasets with non-standard sex column names or codes.
