+++ R/ggpedigree.R
@@ -80,7 +80,8 @@ ggPedigree <- function(ped,
                        hints = NULL,
                        interactive = FALSE,
                        code_male = NULL,
-                       ...) {
+                       sexVar = "sex"
+                       ) {
   if (!inherits(ped, "data.frame")) {
     if (rlang::inherits_any(ped, c("ped", "pedigree", "kinship2.pedigree"))) {
       # Convert ped object to data.frame
@@ -93,7 +94,7 @@ ggPedigree <- function(ped,
       stop("ped should be a data.frame or inherit to a data.frame")
     }
   }
-  if (!all(c(personID, dadID, momID, "sex") %in% names(ped))) {
+  if (!all(c(personID, dadID, momID, sexVar) %in% names(ped))) {
     stop("ped must contain personID, sex, dadID, and momID columns")
   }
 
@@ -119,7 +120,9 @@ ggPedigree <- function(ped,
       hints = hints,
       return_widget = return_widget,
       tooltip_columns = tooltip_columns,
-      ...
+      code_male = code_male,
+      sexVar = sexVar
+
     )
   } else {
     if (interactive == TRUE &&
@@ -161,7 +164,7 @@ ggPedigree <- function(ped,
       config = config,
       debug = debug,
       hints = hints,
-      ...
+      sexVar = sexVar
     )
   }
 }
