### 2025-12-25 c46375c
Add sexVar and spouseID parameters, update code_male handling
Introduces the sexVar parameter to relevant functions and documentation, adds spouseID to ggPedigreeInteractive, and improves code_male handling by allowing it to be set via config or argument. Updates tests and documentation to reflect these changes.
