+++ tests/testthat/test-calcConnections.R
@@ -8,7 +8,7 @@ test_that("calculateConnections returns expected columns and structure", {
   )
 
   ped <- calculateCoordinates(ped,
-    code_male = "M",
+
     personID = "personID",
     momID = "momID",
     dadID = "dadID",
@@ -16,7 +16,8 @@ test_that("calculateConnections returns expected columns and structure", {
     config = list(
       ped_align = TRUE,
       ped_packed = TRUE,
-      return_mid_parent = TRUE
+      return_mid_parent = TRUE,
+      code_male = "M"
     )
   )
 
@@ -47,7 +48,7 @@ test_that("calculateConnections returns expected columns and structure with no s
   )
 
   ped <- calculateCoordinates(ped,
-    code_male = "M",
+
     personID = "personID",
     momID = "momID",
     dadID = "dadID",
@@ -55,7 +56,8 @@ test_that("calculateConnections returns expected columns and structure with no s
     config = list(
       ped_align = TRUE,
       ped_packed = TRUE,
-      return_mid_parent = TRUE
+      return_mid_parent = TRUE,
+      code_male = "M"
     )
   )
 
@@ -89,11 +91,13 @@ test_that("calculateConnections returns correct parent coordinates", {
   )
 
   ped <- calculateCoordinates(ped,
-    code_male = "M",
     personID = "personID",
     momID = "momID",
     dadID = "dadID",
-    spouseID = "spouseID"
+    spouseID = "spouseID",
+    config = list(
+      code_male = "M"
+  )
   )
   conn_out <- calculateConnections(ped)
   conns <- conn_out$connections
@@ -119,11 +123,12 @@ test_that("mid_parent coordinates are correct", {
   )
 
   ped <- calculateCoordinates(ped,
-    code_male = "M",
+
     personID = "personID",
     momID = "momID",
     dadID = "dadID",
-    spouseID = "spouseID"
+    spouseID = "spouseID",
+    config = list(code_male = "M")
   )
   conn_out <- calculateConnections(ped)
   conns <- conn_out$connections
