+++ man/ggPedigree.Rd
@@ -25,7 +25,7 @@ ggPedigree(
   hints = NULL,
   interactive = FALSE,
   code_male = NULL,
-  ...
+  sexVar = "sex"
 )
 
 ggpedigree(
@@ -48,7 +48,7 @@ ggpedigree(
   hints = NULL,
   interactive = FALSE,
   code_male = NULL,
-  ...
+  sexVar = "sex"
 )
 }
 \arguments{
@@ -108,6 +108,8 @@ customization before printing).}
 \item{interactive}{Logical. If TRUE, generates an interactive plot using `plotly`. Default: FALSE.}
 
 \item{code_male}{Integer or string. Value identifying males in the sex column. (typically 0 or 1) Default: 1}
+
+\item{sexVar}{Character string specifying the column name for sex. Defaults to "sex".}
 
 \item{...}{Additional arguments passed to `ggplot2` functions.}
 }
