### 2025-12-25 1e0147a
Update code_male usage in examples and tests
Standardized the use of the code_male parameter in documentation and examples, moving it into the config list. Updated test data to use numeric sex codes to match default code_male = 1 behavior.
