### 2025-12-25 39f3e6d
Remove unused ellipsis argument from ggPedigree docs
Eliminated the documentation for the '...' argument in both R/ggpedigree.R and man/ggPedigree.Rd, as it is no longer used or relevant for the ggPedigree function.
