### 2025-12-25 8c3a0b3
Refactor sex recoding
Moved the standardization of the sex variable in preparePedigreeData to occur after phantom parent addition
