### 2025-12-25 d14a5a4
Add full redsquirrels dataset and update documentation
Introduced the full redsquirrels dataset (redsquirrels_full) and corresponding CSV and RDA files. Updated data processing scripts to generate both the full and subset datasets, and improved documentation to describe both datasets and their usage.
