+++ man/redsquirrels.Rd
@@ -3,9 +3,10 @@
 \docType{data}
 \name{redsquirrels}
 \alias{redsquirrels}
+\alias{redsquirrels_full}
 \title{Kluane Red Squirrel Data}
 \format{
-## `redsquirrels`
+## `redsquirrels_full`
 A data frame with 7799 rows and 16 columns:
 \describe{
   \item{personID}{Unique identifier for each squirrel}
@@ -25,35 +26,54 @@ A data frame with 7799 rows and 16 columns:
   \item{year_last}{Last year of ARS data for the squirrel}
   ...
 }
+## `redsquirrels`
+A data frame with 5251 rows and 16 columns:
+A subset of \code{redsquirrels_full} intended for convenient analysis and examples.
+(Same variables as \code{redsquirrels_full}, with fewer rows.)
+
+An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7799 rows and 16 columns.
 }
 \source{
 <https://doi.org/10.5061/dryad.n5q05>
 }
 \usage{
 data(redsquirrels)
+
+data(redsquirrels_full)
 }
 \description{
 A tidy data frame of life‐history and reproductive metrics for 7,799 individual red squirrels
 from the Kluane Red Squirrel Project (1987–present).
-Each row corresponds to one squirrel with associated pedigree links and reproductive success summaries.
-The original data are published under a CC0 1.0 Universal Public Domain Dedication:
 }
 \details{
+#' This package provides two related datasets:
+\itemize{
+  \item \code{redsquirrels_full}: the complete dataset from the published source
+  \item \code{redsquirrels}: a more workable subset derived from \code{redsquirrels_full}
+}
+
+Each row corresponds to one squirrel with associated pedigree links and reproductive success summaries.
+
+The original data are published under a CC0 1.0 Universal Public Domain Dedication:
+
 McFarlane, S. Eryn; Boutin, Stan; Humphries, Murray M. et al. (2015). Data from:
 Very low levels of direct additive genetic variance in fitness and fitness components in a red squirrel population [Dataset].
 Dryad. <https://doi.org/10.5061/dryad.n5q05>
 }
 \examples{
-# Load the red squirrels dataset
-data(redsquirrels)
 
-# View the structure of the dataset
+# Load the red squirrels datasets
+data(redsquirrels)
+data(redsquirrels_full)
+
+# View the structure of the dataset(s)
 str(redsquirrels)
+str(redsquirrels_full)
 
 # Plot a pedigree for a single family
 if (requireNamespace("ggplot2", quietly = TRUE)) {
   # Select one family to plot
-  family_data <- subset(redsquirrels, famID == 1)
+  family_data <- subset(redsquirrels, famID == 8)
 
   # Create a pedigree plot
   ggPedigree(family_data,
