### 2025-12-26 d21f1cd
Update default parameters and example data in documentation
Set 'repairsex' to FALSE in preparePedigreeData, update default 'sexVar' to 'sex' in alignPedigreeWithRelations documentation, and change example family ID in redsquirrels documentation from 8 to 160.
