+++ R/ggpedigreeCoreHelpers.R
@@ -96,7 +96,7 @@ preparePedigreeData <- function(ped,
       addphantoms = TRUE,
       repair = TRUE,
       parentswithoutrow = FALSE,
-      repairsex = T,
+      repairsex = FALSE,
       personID = personID,
       momID = momID,
       dadID = dadID,
