### 2025-12-26 e6c2b7d
refactoring nodes
