+++ vignettes/v00_plots.R
@@ -69,7 +69,7 @@ ggPedigree(potter,
   famID = "famID",
   personID = "personID"
 ) +
-  theme_bw(base_size = 12)
+  theme_bw(base_size = 12) + scale_colour_brewer(palette = "Set2")
 
 ## -----------------------------------------------------------------------------
 ggPedigree(
@@ -103,7 +103,7 @@ p <- ggPedigree(
     status_code_unaffected = FALSE,
     status_shape_affected = 4
   )
-)
+) 
 
 p
 
