+++ vignettes/v00_plots.Rmd
@@ -123,7 +123,7 @@ ggPedigree(potter,
   famID = "famID",
   personID = "personID"
 ) +
-  theme_bw(base_size = 12)
+  theme_bw(base_size = 12) + scale_colour_brewer(palette = "Set2")
 ```
 
 This can be particularly useful for matching house styles in academic publications or removing clutter for presentations.
@@ -169,7 +169,7 @@ p <- ggPedigree(
     status_code_unaffected = FALSE,
     status_shape_affected = 4
   )
-)
+) 
 
 p
 ```
