### 2025-12-26 2916e41
refactoring nodes
