+++ R/ggpedigreeCoreHelpers.R
@@ -423,3 +423,39 @@ addFocalFillColumn <- function(ds_ped,
   # -----
   return(ds_ped)
 }
+#' @title Pick First Matching Rule
+#' @description
+#' This function evaluates a list of rules and returns the action associated with the first rule that matches.
+#' If no rules match, it returns a default value.
+#' @param rules A list of rules, where each rule is a list with `when` and `do` elements.
+#' @param default The default value to return if no rules match (default is NULL).
+#' @return The action associated with the first matching rule, or the default value.
+#' @keywords internal
+.pick_first <- function(rules, default = NULL) {
+  for (r in rules) {
+    if (isTRUE(r$when())) return(r$do)
+  }
+  default
+}
+
+
+.should_add_overlay <- function(config,
+                                overlay_column = NULL,
+                                status_column= NULL,
+                                focal_fill_column = NULL
+                                ) {
+  isTRUE(config$overlay_include) && !is.null(overlay_column) ||
+    (isTRUE(config$status_include) && !is.null(status_column) && isTRUE(config$sex_color_include)) ||
+    (isTRUE(config$focal_fill_include) && !is.null(focal_fill_column) && !isTRUE(config$sex_color_include))
+}
+
+.get_color_mode <- function(config, status_column, focal_fill_column) {
+  .pick_first(
+    rules = list(
+      list(when = function() isTRUE(config$sex_color_include), do = "sex"),
+      list(when = function() isTRUE(config$focal_fill_include), do = "focal_fill"),
+      list(when = function() !is.null(status_column) && isTRUE(config$status_include), do = "status")
+    ),
+    default = "none"
+  )
+}
