+++ R/ggpedigreeCore.R
@@ -295,10 +295,9 @@ ggPedigree.core <- function(ped,
   )
 
   # Add overlay points for affected status if applicable
-  if (
-    config$focal_fill_include == TRUE && config$sex_color_include == FALSE ||
-      config$overlay_include == TRUE && !is.null(overlay_column) ||
-      !is.null(status_column) && config$status_include == TRUE && config$sex_color_include == TRUE) {
+
+  if (.should_add_overlay(config, overlay_column, status_column, focal_fill_column)) {
+
     # If overlay_column is specified, use it for alpha aesthetic
 
     p <- .addOverlay(
@@ -404,7 +403,7 @@ ggPedigree.core <- function(ped,
     message("Status column: ", status_column)
   }
 
-  if (config$outline_include == TRUE) {
+  if (isTRUE(config$outline_include)) {
     plotObject <- plotObject +
       ggplot2::geom_point(
         ggplot2::aes(shape = as.factor(.data$sex)),
@@ -416,8 +415,30 @@ ggPedigree.core <- function(ped,
       )
   }
 
-  if (config$sex_color_include == TRUE) {
-    # Use color and shape to represent sex
+
+  # 2) Determine which "node mode" to use (exactly one)
+  node_mode <- .pick_first(
+    rules = list(
+      list(
+        when = function() isTRUE(config$sex_color_include),
+        do   = "sex_color"
+      ),
+      list(
+        when = function() isTRUE(config$focal_fill_include),
+        do   = "focal_fill"
+      ),
+      list(
+        when = function() isTRUE(config$status_include) && !is.null(status_column),
+        do   = "status"
+      )
+    ),
+    default = "shape_only"
+  )
+
+
+
+  # 3) Add the selected layer
+  if (node_mode == "sex_color") {
     plotObject <- plotObject +
       ggplot2::geom_point(
         ggplot2::aes(
@@ -427,34 +448,22 @@ ggPedigree.core <- function(ped,
         size = config$point_size,
         na.rm = TRUE
       )
-  } else if (config$focal_fill_include == TRUE) {
-    # If status_column is not present but status_include is TRUE,
-    # use alpha aesthetic to show affected status
-    if (is.null(focal_fill_column)) {
-      plotObject <- plotObject +
-        ggplot2::geom_point(
-          ggplot2::aes(
-            color = .data$focal_fill,
-            shape = as.factor(.data$sex)
-          ),
-          size = config$point_size,
-          na.rm = TRUE
-        )
-    } else {
-      plotObject <- plotObject +
-        ggplot2::geom_point(
-          ggplot2::aes(
-            color = !!rlang::sym(focal_fill_column),
-            shape = as.factor(.data$sex)
-          ),
-          size = config$point_size,
-          na.rm = TRUE
-        )
-    }
-  } else if (config$status_include == TRUE &&
-    !is.null(status_column)) {
-    # If status_column is present but sex_color_include is FALSE,
-    # use shape for sex and color for affected status
+
+  } else if (node_mode == "focal_fill") {
+    # Preserve your original "if focal_fill_column is NULL, use .data$focal_fill"
+    color_expr <- if (is.null(focal_fill_column)) rlang::expr(.data$focal_fill) else rlang::sym(focal_fill_column)
+
+    plotObject <- plotObject +
+      ggplot2::geom_point(
+        ggplot2::aes(
+          color = !!color_expr,
+          shape = as.factor(.data$sex)
+        ),
+        size = config$point_size,
+        na.rm = TRUE
+      )
+
+  } else if (node_mode == "status") {
     plotObject <- plotObject +
       ggplot2::geom_point(
         ggplot2::aes(
@@ -464,18 +473,18 @@ ggPedigree.core <- function(ped,
         size = config$point_size,
         na.rm = TRUE
[…]
