### 2025-12-27 3afdc99
Refactor addScales for modular color/shape scale logic
Refactored the addScales function in ggpedigreeCore.R to delegate color and shape scale logic to dedicated helper functions for sex, status, and focal fill modes. This improves code readability, maintainability, and extensibility. Updated NEWS.md to reflect the refactor.
