### 2025-12-27 9a38f2a
Add apply_default_color option to plot config
Introduces the apply_default_color parameter to getDefaultPlotConfig, allowing users to toggle between the default color palette and black for plot elements. Updates related logic, documentation, and tests to support this new option and ensure consistent color handling.
