### 2025-12-27 f1f7217
Add option for black coloring when default color is off
Updated plot config logic to set all color palettes to black when apply_default_color is FALSE. Refactored default and effective palette handling in getDefaultPlotConfig, and updated documentation accordingly. This improves flexibility for users who want monochrome plots.
