+++ man/getDefaultPlotConfig.Rd
@@ -82,12 +82,7 @@ getDefaultPlotConfig(
   sex_color_include = TRUE,
   sex_legend_title = "Sex",
   sex_shape_labels = c("Female", "Male", "Unknown"),
-  sex_color_palette = if (apply_default_color) {
-     color_palette_default
- } else {
-   
-     c("black", "black", "black")
- },
+  sex_color_palette = NULL,
   sex_shape_female = 16,
   sex_shape_male = 15,
   sex_shape_unknown = 18,
@@ -101,9 +96,9 @@ getDefaultPlotConfig(
   status_label_unaffected = "Unaffected",
   status_alpha_affected = 1,
   status_alpha_unaffected = 0,
-  status_color_palette = c(color_palette_default[1], color_palette_default[2]),
+  status_color_palette = NULL,
   status_color_affected = "black",
-  status_color_unaffected = color_palette_default[2],
+  status_color_unaffected = NULL,
   status_shape_affected = 4,
   status_legend_title = "Affected",
   status_legend_show = FALSE,
