+++ NEWS.md
@@ -5,6 +5,7 @@
 * Tweak code_male handling to be more robust about using config values
 * Slice up squirrel dataset into smaller pedigrees for testing
 * Refactored add node
+* set option to include black coloring if apply_default_color is FALSE
 
 # ggpedigree 1.1.0.0
 * Saved raw excel data as csv to have a non-proprietary format for data storage
