### 2025-12-27 fdc7d79
Refactor and clean up code formatting and logic
This commit standardizes code formatting, improves indentation, and clarifies logic across multiple R scripts and test files. It also removes unnecessary blank lines, aligns function arguments, and updates comments for better readability and maintainability. No functional changes are introduced.
