+++ R/defaultPlotConfig.R
@@ -292,7 +292,7 @@ getDefaultPlotConfig <- function(function_name = "getDefaultPlotConfig",
                                  status_label_unaffected = "Unaffected",
                                  status_alpha_affected = 1,
                                  status_alpha_unaffected = 0,
-                                 status_color_palette =  NULL,
+                                 status_color_palette = NULL,
                                  # Use first color for affected,
                                  status_color_affected = "black",
                                  status_color_unaffected = NULL,
