+++ R/ggpedigreeCore.R
@@ -94,7 +94,6 @@ ggPedigree.core <- function(ped,
   }
 
 
-
   # -----
   # STEP 4: Coordinate Generation
   # -----
@@ -297,7 +296,6 @@ ggPedigree.core <- function(ped,
   # Add overlay points for affected status if applicable
 
   if (.should_add_overlay(config, overlay_column, status_column, focal_fill_column)) {
-
     # If overlay_column is specified, use it for alpha aesthetic
 
     p <- .addOverlay(
@@ -429,7 +427,6 @@ ggPedigree.core <- function(ped,
         when = function() isTRUE(config$focal_fill_include) && !is.null(focal_fill_column),
         do   = "focal_fill"
       ),
-
       list(
         when = function() isTRUE(config$status_include) && !is.null(status_column),
         do   = "status"
@@ -437,7 +434,6 @@ ggPedigree.core <- function(ped,
     ),
     default = "shape_only"
   )
-
 
 
   # 3) Add the selected layer
@@ -451,14 +447,13 @@ ggPedigree.core <- function(ped,
         size = config$point_size,
         na.rm = TRUE
       )
-
   } else if (node_mode == "focal_fill") {
     # Preserve your original "if focal_fill_column is NULL, use .data$focal_fill"
-    color_expr <- if (is.null(focal_fill_column)){
+    color_expr <- if (is.null(focal_fill_column)) {
       rlang::expr(.data$focal_fill)
-      } else {
+    } else {
       rlang::sym(focal_fill_column)
-      }
+    }
 
     plotObject <- plotObject +
       ggplot2::geom_point(
@@ -469,7 +464,6 @@ ggPedigree.core <- function(ped,
         size = config$point_size,
         na.rm = TRUE
       )
-
   } else if (node_mode == "status") {
     plotObject <- plotObject +
       ggplot2::geom_point(
@@ -480,7 +474,6 @@ ggPedigree.core <- function(ped,
         size = config$point_size,
         na.rm = TRUE
       )
-
   } else { # "shape_only"
     plotObject <- plotObject +
       ggplot2::geom_point(
@@ -491,7 +484,6 @@ ggPedigree.core <- function(ped,
   }
 
   plotObject
-
 }
 
 #' @rdname dot-addNodes
@@ -505,11 +497,12 @@ addNodes <- .addNodes
 #' @return A ggplot object with added overlay.
 #'
 .addOverlay <- function(plotObject,
-                        config = list(overlay_include = FALSE,
-                                      status_include = FALSE,
-                                      focal_fill_include = FALSE,
-                                      sex_color_include = FALSE),
-
+                        config = list(
+                          overlay_include = FALSE,
+                          status_include = FALSE,
+                          focal_fill_include = FALSE,
+                          sex_color_include = FALSE
+                        ),
                         focal_fill_column = NULL,
                         status_column = NULL,
                         overlay_column = NULL) {
@@ -519,27 +512,31 @@ addNodes <- .addNodes
     rules = list(
       list(
         when = function() isTRUE(config$overlay_include) && !is.null(overlay_column),
-        do   = list(
+        do = list(
           alpha_var = overlay_column,
-          shape      = config$overlay_shape,
-          color      = config$overlay_color
+          shape = config$overlay_shape,
+          color = config$overlay_color
         )
       ),
       list(
-        when = function() isTRUE(config$status_include) &&
-          !is.null(status_column) &&
-          isTRUE(config$sex_color_include),
-        do   = list(
+        when = function() {
+          isTRUE(config$status_include) &&
+            !is.null(status_column) &&
+            isTRUE(config$sex_color_include)
+        },
+        do = list(
           alpha_var = status_column,
-          shape      = config$status_shape_affected,
[…]
