### 2025-12-29 d7a2d52
added stop when insufficient data to graph
This seems to be a deeper issue about handling singletons. For now I've set this to have a warning https://github.com/R-Computing-Lab/ggpedigree/issues/93
