+++ R/calcConnectionsHelpers.R
@@ -263,8 +263,8 @@ getRelativeCoordinates <- function(ped,
   # Require same type
   if (mode(id1) != mode(id2)) {
     # if they're both a number type, allow it
-    if (mode(id1) %in% c("integer", "double") &&
-      mode(id2) %in% c("integer", "double")) {
+    if (mode(id1) %in% c("integer", "double","logical","numeric") &&
+      mode(id2) %in% c("integer", "double","logical","numeric")) {
       # Numeric comparison
     } else {
       stop(paste0(
