### 2025-12-29 78d0535
Update ggpedigreeCore.R
