### 2025-12-29 0cb4b96
Update ggpedigreeCore.R
