+++ R/ggpedigreeCore.R
@@ -334,14 +334,15 @@ ggPedigree.core <- function(ped,
 # scale min
 
 
- if(is.na(min(ds$y_pos)) ) {
-   stop("y_pos contains all NA values, cannot set y-axis limits. This occurs when there are less than two people in the pedigree.")
+ if(is.na(min(ds$y_pos, na.rm = TRUE)) ) {
+   warning("y_pos contains all NA values, cannot set y-axis limits. This occurs when there are less than two people in the pedigree.")
+
  } else {
 
   p <- p +
     ggplot2::scale_y_reverse(limits = c(
       NA,
-      min(ds$y_pos)
+      min(ds$y_pos, na.rm = TRUE)
     ))
 }
   if (config$apply_default_theme == TRUE) {
