### 2025-12-29 R-Computing-Lab/ggpedigree#93 @smasongarrison
I've now added a warning when there's only a single person to graph. 0cb4b96193b31f0088dff6b9ef8edd028a4c0ef8 I thought about manually forcing coordinates when there's only one person to plot, but that seemed like it might have unintended consequences with faceting and some other ggplot-specific features.
