### 2025-12-29 a47faf4
tweaked default colors
still using viridis, but now using less yellow
