### 2025-12-29 d19d6cd
tweaked default colors
still using viridis, but now using less yellow
