### 2025-12-29 99d5ff1
Update buildPlotConfig.R
