+++ reference/ggPedigreeInteractive.md
@@ -17,6 +17,7 @@ ggPedigreeInteractive(
   patID = "patID",
   matID = "matID",
   twinID = "twinID",
+  spouseID = "spouseID",
   status_column = NULL,
   tooltip_columns = NULL,
   focal_fill_column = NULL,
@@ -24,7 +25,9 @@ ggPedigreeInteractive(
   config = list(optimize_plotly = TRUE),
   debug = FALSE,
   return_widget = TRUE,
-  ...
+  hints = NULL,
+  code_male = NULL,
+  sexVar = "sex"
 )
 
 ggpedigreeInteractive(
@@ -36,6 +39,7 @@ ggpedigreeInteractive(
   patID = "patID",
   matID = "matID",
   twinID = "twinID",
+  spouseID = "spouseID",
   status_column = NULL,
   tooltip_columns = NULL,
   focal_fill_column = NULL,
@@ -43,7 +47,9 @@ ggpedigreeInteractive(
   config = list(optimize_plotly = TRUE),
   debug = FALSE,
   return_widget = TRUE,
-  ...
+  hints = NULL,
+  code_male = NULL,
+  sexVar = "sex"
 )
 
 ggpedigreeinteractive(
@@ -55,6 +61,7 @@ ggpedigreeinteractive(
   patID = "patID",
   matID = "matID",
   twinID = "twinID",
+  spouseID = "spouseID",
   status_column = NULL,
   tooltip_columns = NULL,
   focal_fill_column = NULL,
@@ -62,7 +69,9 @@ ggpedigreeinteractive(
   config = list(optimize_plotly = TRUE),
   debug = FALSE,
   return_widget = TRUE,
-  ...
+  hints = NULL,
+  code_male = NULL,
+  sexVar = "sex"
 )
 ```
 
@@ -107,6 +116,11 @@ ggpedigreeinteractive(
 
   Character string specifying the column name for twin IDs. Defaults to
   "twinID".
+
+- spouseID:
+
+  Character string specifying the column name for spouse IDs. Defaults
+  to "spouseID".
 
 - status_column:
 
@@ -190,9 +204,19 @@ ggpedigreeinteractive(
   returns the underlying plotly object (useful for further customization
   before printing).
 
-- ...:
+- hints:
 
-  Additional arguments passed to \`ggplot2\` functions.
+  Data frame with hints for layout adjustments. Default: NULL.
+
+- code_male:
+
+  Integer or string. Value identifying males in the sex column.
+  (typically 0 or 1) Default: 1
+
+- sexVar:
+
+  Character string specifying the column name for sex. Defaults to
+  "sex".
 
 ## Value
 
@@ -205,5 +229,5 @@ library(BGmisc)
 data("potter")
 ggPedigreeInteractive(potter, famID = "famID", personID = "personID")
 
[…]
