+++ reference/ggpedigree-package.md
@@ -1,19 +1,16 @@
 # ggpedigree: Visualizing Pedigrees with 'ggplot2' and 'plotly'
 
 Provides plotting functions for visualizing pedigrees and family trees.
-The package complements a behavior genetics package 'BGmisc' \[Garrison
-et al. (2024)
+in behavior genetics and kinship research. The package complements a
+behavior genetics package 'BGmisc' \[Garrison et al. (2024)
 [doi:10.21105/joss.06203](https://doi.org/10.21105/joss.06203) \] by
-rendering pedigrees using the 'ggplot2' framework. Features include
-support for duplicated individuals, complex mating structures,
-integration with simulated pedigrees, and layout customization. Due to
-the impending deprecation of kinship2, version 1.0 incorporates the
-layout helper functions from kinship2. The pedigree alignment algorithms
-are adapted from 'kinship2' \[Sinnwell et al. (2014)
-[doi:10.1159/000363105](https://doi.org/10.1159/000363105) \]. We
-gratefully acknowledge the original authors: Jason Sinnwell, Terry
-Therneau, Daniel Schaid, and Elizabeth Atkinson for their foundational
-work.
+rendering pedigrees using the 'ggplot2' framework and offers a modern
+redesign beyond the base-graphics pedigree plot in 'kinship2' \[Sinnwell
+et al. (2014) [doi:10.1159/000363105](https://doi.org/10.1159/000363105)
+\]. Features include support for duplicated individuals, complex mating
+structures, integration with simulated pedigrees, and layout
+customization. Due to the impending depreciation of kinship2, version
+1.0 incorporates the layout helper functions from kinship2.
 
 ## See also
 
