+++ reference/preparePedigreeData.md
@@ -15,8 +15,7 @@ preparePedigreeData(
   matID = "matID",
   patID = "patID",
   config = list(focal_fill_include = TRUE, focal_fill_component = "maternal",
-    recode_missing_ids = TRUE, add_phantoms = FALSE, code_male = "M", code_female = "F",
-    code_na = NA),
+    recode_missing_ids = TRUE, add_phantoms = FALSE),
   fill_group_paternal = c("paternal", "patID", "paternal line", "paternal lineages",
     "paternal lines"),
   fill_group_maternal = c("maternal", "matID", "maternal line", "maternal lineages",
