+++ reference/ggPedigree.md
@@ -28,8 +28,7 @@ ggPedigree(
   debug = FALSE,
   hints = NULL,
   interactive = FALSE,
-  code_male = NULL,
-  sexVar = "sex"
+  ...
 )
 
 ggpedigree(
@@ -51,8 +50,7 @@ ggpedigree(
   debug = FALSE,
   hints = NULL,
   interactive = FALSE,
-  code_male = NULL,
-  sexVar = "sex"
+  ...
 )
 ```
 
@@ -194,15 +192,9 @@ ggpedigree(
   Logical. If TRUE, generates an interactive plot using \`plotly\`.
   Default: FALSE.
 
-- code_male:
+- ...:
 
-  Integer or string. Value identifying males in the sex column.
-  (typically 0 or 1) Default: 1
-
-- sexVar:
-
-  Character string specifying the column name for sex. Defaults to
-  "sex".
+  Additional arguments passed to \`ggplot2\` functions.
 
 ## Value
 
