### 2025-12-30 df65777
Update calculateCoordinates.Rd
