+++ man/calculateCoordinates.Rd
@@ -83,4 +83,31 @@ coords_custom <- calculateCoordinates(
     ped_width = 20
   )
 )
+# Load example data
+data(potter, package = "BGmisc")
+
+# Calculate coordinates for the pedigree
+coords <- calculateCoordinates(
+  ped = potter,
+  personID = "personID",
+  momID = "momID",
+  dadID = "dadID",
+  config  = list(
+   code_male = 1)
+)
+
+# View the coordinates
+head(coords)
+
+# Example with custom configuration
+coords_custom <- calculateCoordinates(
+  ped = potter,
+  personID = "personID",
+  momID = "momID",
+  dadID = "dadID",
+  config = list(
+    ped_packed = FALSE,
+    ped_width = 20
+  )
+)
 }
