+++ reference/ASOIAF.md
@@ -53,24 +53,24 @@ The variables are as follows:
 ``` r
 # Load the ASOIAF dataset
 data(ASOIAF)
-
+df_ASOIAF <-  ASOIAF[ASOIAF$famID == 26,]  # Subset to House Tarth
 # View the structure of the dataset
-str(ASOIAF)
-#> 'data.frame':    679 obs. of  9 variables:
-#>  $ id      : num  1 2 3 4 5 6 7 8 9 10 ...
-#>  $ famID   : num  1 1 1 1 1 1 1 1 1 1 ...
-#>  $ momID   : num  566 NA 2 2 2 NA 6 NA NA NA ...
-#>  $ dadID   : num  564 NA 1 1 1 NA 3 7 7 7 ...
-#>  $ name    : chr  "Walder Frey" "Perra Royce" "Stevron Frey" "Emmon Frey" ...
-#>  $ sex     : chr  "M" "F" "M" "M" ...
-#>  $ url     : chr  "https://awoiaf.westeros.org/index.php/Walder_Frey" "https://awoiaf.westeros.org/index.php/Perra_Royce" "https://awoiaf.westeros.org/index.php/Stevron_Frey" "https://awoiaf.westeros.org/index.php/Emmon_Frey" ...
-#>  $ twinID  : num  NA NA NA NA NA NA NA NA NA NA ...
+str(df_ASOIAF)
+#> 'data.frame':    6 obs. of  9 variables:
+#>  $ id      : num  654 655 656 657 658 659
+#>  $ famID   : num  26 26 26 26 26 26
+#>  $ momID   : num  NA NA 655 655 655 655
+#>  $ dadID   : num  NA NA 654 654 654 654
+#>  $ name    : chr  "Selwyn Tarth" "Unknown wife of Selwyn Tarth" "Galladon Tarth" "Arianne Tarth" ...
+#>  $ sex     : chr  "M" "F" "M" "F" ...
+#>  $ url     : chr  "https://awoiaf.westeros.org/index.php/Selwyn_Tarth" "https://awoiaf.westeros.org/index.php/Selwyn_Tarth#Family" "https://awoiaf.westeros.org/index.php/Galladon_Tarth" "https://awoiaf.westeros.org/index.php/Arianne_Tarth" ...
+#>  $ twinID  : num  NA NA NA NA NA NA
 #>  $ zygosity: chr  NA NA NA NA ...
 
-# Plot a pedigree for a specific family
+# Plot a pedigree for House Tarth
 if (requireNamespace("ggplot2", quietly = TRUE)) {
-  # Create a pedigree plot for House Stark
-  ggPedigree(ASOIAF,
+  # Create a pedigree plot for House Tarth
+  ggPedigree(df_ASOIAF,
     famID = "famID",
     personID = "id",
     momID = "momID",
