+++ llms.txt
@@ -176,6 +176,7 @@ file.
   package `kinship2`.
 
 - [`redsquirrels`](https://r-computing-lab.github.io/ggpedigree/reference/redsquirrels.md)
+  [`redsquirrels_full`](https://r-computing-lab.github.io/ggpedigree/reference/redsquirrels.md)
   : Kluane Red Squirrel Data
 
 - [`warsofroses`](https://r-computing-lab.github.io/ggpedigree/reference/warsofroses.md)
