+++ reference/dot-addOverlay.md
@@ -7,7 +7,8 @@ Add Overlay to ggplot Pedigree Plot
 ``` r
 .addOverlay(
   plotObject,
-  config,
+  config = list(overlay_include = FALSE, status_include = FALSE, focal_fill_include =
+    FALSE, sex_color_include = FALSE),
   focal_fill_column = NULL,
   status_column = NULL,
   overlay_column = NULL
@@ -15,7 +16,8 @@ Add Overlay to ggplot Pedigree Plot
 
 addOverlay(
   plotObject,
-  config,
+  config = list(overlay_include = FALSE, status_include = FALSE, focal_fill_include =
+    FALSE, sex_color_include = FALSE),
   focal_fill_column = NULL,
   status_column = NULL,
   overlay_column = NULL

[…]
