### 2025-12-31 4327619
Add greyscale color theme support to plot config
Introduces a 'color_theme' parameter to getDefaultPlotConfig and related plotting functions, allowing users to switch between color and greyscale palettes. Updates documentation and tests to cover greyscale theming, ensuring all relevant color settings adapt accordingly when greyscale is selected.
