+++ tests/testthat/test-ggPedigree.R
@@ -287,7 +287,8 @@ test_that("focal fill works with ID and different methods", {
       focal_fill_include = TRUE,
       sex_color_include = FALSE,
       focal_fill_personID = 1,
-      focal_fill_method = "steps"
+      focal_fill_method = "steps",
+      color_theme = "greyscale"
     )
   )
   expect_s3_class(p, "gg") # Should return a ggplot object
