+++ R/defaultPlotConfig.R
@@ -5,6 +5,11 @@
 #' plotting functions such as `ggPedigree()`, `ggRelatednessMatrix()`, and `ggPhenotypeByDegree()`.
 #' Each key corresponds to a configurable plot, layout, or aesthetic behavior.
 #' @param color_palette_default A character vector of default colors for the plot.
+#' @param color_theme Theme mode controlling default palettes. Options: "color" (default) or "greyscale".
+#' @param greyscale_palette_default Default discrete greyscale palette used when color_theme="greyscale".
+#' @param greyscale_low Greyscale low color for continuous gradients when color_theme="greyscale".
+#' @param greyscale_mid Greyscale midpoint color for continuous gradients when color_theme="greyscale".
+#' @param greyscale_high Greyscale high color for continuous gradients when color_theme="greyscale".
 #' @param segment_default_color A character string for the default color of segments in the plot.
 #' @param function_name The name of the function calling this configuration.
 #' @param personID The column name for person identifiers in the data.
@@ -189,6 +194,12 @@ getDefaultPlotConfig <- function(function_name = "getDefaultPlotConfig",
                                  apply_default_scales = TRUE,
                                  apply_default_theme = TRUE,
                                  segment_default_color = "black",
+                                 # Greyscale theming option
+                                 color_theme = "color", # c("color", "greyscale"),
+                                 greyscale_palette_default = c("#1A1A1A", "#7F7F7F", "#D9D9D9"),
+                                 greyscale_low = "#000000",
+                                 greyscale_mid = "#7F7F7F",
+                                 greyscale_high = "#FFFFFF",
                                  color_palette_default = c(
                                    "#440154FF",
                                    "#7fd34e",
@@ -412,12 +423,55 @@ getDefaultPlotConfig <- function(function_name = "getDefaultPlotConfig",
       " is not supported by getDefaultPlotConfig."
     ))
   }
+  # color_theme <- match.arg(color_theme)
 
+  grey_color_names <- c(
+    "greyscale",
+    "grayscale",
+    "grey",
+    "gray"
+  )
+  black_color_names <- c(
+    "blackscale",
+    "blackandwhite",
+    "black & white",
+    "black white",
+    "bw",
+    "black"
+  )
+
+  if (stringr::str_to_lower(color_theme) %in% c(grey_color_names, black_color_names)) {
+    color_palette_default <- greyscale_palette_default
+    color_palette_low <- greyscale_low
+    color_palette_mid <- greyscale_mid
+    color_palette_high <- greyscale_high
+
+    segment_default_color <- "black"
+    label_text_color <- "black"
+    outline_color <- "black"
+    overlay_color <- "black"
+
+    status_color_palette <- c("#1A1A1A", "#BDBDBD")
+    status_color_affected <- "#1A1A1A"
+    status_color_unaffected <- "#BDBDBD"
+
+    focal_fill_high_color <- "grey80"
+    focal_fill_mid_color <- "grey50"
+    focal_fill_low_color <- "#1A1A1A"
+    focal_fill_na_value <- "black"
+    tile_color_palette <- c("#FFFFFF", "#BDBDBD", "#000000")
+
+    # Make sex palette greyscale too
+    sex_color_palette <- rep("black", length(sex_color_palette))
+
+    focal_fill_color_values <- c("#1A1A1A", "#7F7F7F", "#D9D9D9")
+  }
   core_list <- list(
     # ---- General Appearance ----
     apply_default_scales = apply_default_scales,
     segment_default_color = segment_default_color,
     apply_default_theme = apply_default_theme,
+    color_theme = color_theme,
     color_palette_default = color_palette_default,
[…]
