### 2025-12-31 a833caa
Update README with greyscale example and bump version
Added a new example in the README and README.Rmd demonstrating the use of the greyscale color theme in ggPedigree, along with the corresponding output image. Updated NEWS.md to reflect recent changes and bumped the package version to 1.1.0.1 in DESCRIPTION.
