+++ NEWS.md
@@ -1,12 +1,9 @@
 # ggpedigree (development version)
 # ==============================
-* Improved documentation of datasets included in the package
-* Added smarter warning message for tryCatch
-* Tweak code_male handling to be more robust about using config values
-* Slice up squirrel dataset into smaller pedigrees for testing
-* Refactored add node
+
 
 # ggpedigree 1.1.0.0
+
 * Saved raw excel data as csv to have a non-proprietary format for data storage
 * Added more refactorings of kinship2 internals for maintainability
 * Harmonize names by changing midparent to mid_parent throughout the codebase
@@ -20,7 +17,12 @@
 * Set Sex key to be hidden by default in ggpedigree
 * Slight tweaks to GoT dataset demonstration to improve visualization
 * Remove reshape2 dependency by using base R functions
-
+* Improved documentation of datasets included in the package
+* Added smarter warning message for tryCatch
+* Tweak code_male handling to be more robust about using config values
+* Slice up squirrel dataset into smaller pedigrees for testing
+* Refactored add node
+* Added greyscale option
 
 # ggpedigree 1.0.0.1
 ## cran resubmission
