+++ README.md
@@ -68,6 +68,9 @@ pedigree structure. The `potter` dataset contains simulated pedigree
 data for the Weasley family from the Harry Potter series.
 
 ``` r
+library(ggpedigree) # ggPedigree lives here
+library(BGmisc) # helper utilities & example data
+potter <- BGmisc::potter # load example data
 ggPedigree(potter,
   famID = "famID",
   personID = "personID"
@@ -75,6 +78,17 @@ ggPedigree(potter,
 ```
 
 <img src="man/figures/README-basic-usage-1.png" width="80%" />
+
+``` r
+ggPedigree(potter,
+  famID = "famID",
+  personID = "personID",
+  config = list(
+    color_theme = "greyscale")
+)
+```
+
+<img src="man/figures/README-slightly-usage-1.png" width="80%" />
 
 ## Citation
 
@@ -84,7 +98,7 @@ cite the following:
     citation(package = "ggpedigree")
 
 Garrison S (2025). *ggpedigree: Visualizing Pedigrees with ‘ggplot2’ and
-‘plotly’*. R package version 1.0.0.1,
+‘plotly’*. R package version 1.1.0.1,
 <https://github.com/R-Computing-Lab/ggpedigree/>.
 
 A BibTeX entry for LaTeX users is
@@ -93,7 +107,7 @@ A BibTeX entry for LaTeX users is
       title = {ggpedigree: Visualizing Pedigrees with 'ggplot2' and 'plotly'},
       author = {S. Mason Garrison},
       year = {2025},
-      note = {R package version 1.0.0.1},
+      note = {R package version 1.1.0.1},
       url = {https://github.com/R-Computing-Lab/ggpedigree/},
     }
 
