### 2025-12-31 844eb10
Update NEWS.md
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
