### 2025-12-31 6828a39
Update R/defaultPlotConfig.R
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
