### 2025-12-31 7e5e48a
Convert all hex codes to R color names in greyscale theme
Co-authored-by: smasongarrison <6001608+smasongarrison@users.noreply.github.com>
