### 2025-12-31 7a20f33
Update getDefaultPlotConfig.Rd
