+++ man/getDefaultPlotConfig.Rd
@@ -13,10 +13,10 @@ getDefaultPlotConfig(
   apply_default_theme = TRUE,
   segment_default_color = "black",
   color_theme = "color",
-  greyscale_palette_default = c("#1A1A1A", "#7F7F7F", "#D9D9D9"),
-  greyscale_low = "#000000",
-  greyscale_mid = "#7F7F7F",
-  greyscale_high = "#FFFFFF",
+  greyscale_palette_default = c("grey10", "grey50", "grey85"),
+  greyscale_low = "black",
+  greyscale_mid = "grey50",
+  greyscale_high = "white",
   color_palette_default = c("#440154FF", "#7fd34e", "#f1e51d"),
   color_palette_low = "#000004FF",
   color_palette_mid = "#56106EFF",
