### 2025-12-31 f954554
Add greyscale theme assertions to focal fill test
Restored color_theme = "greyscale" parameter and added proper test
assertions to validate that greyscale colors are applied correctly.
The test now verifies:
- A colour scale exists for focal fill
- The plot can be built successfully
- The scale has the correct aesthetic mapping
- The scale is a continuous scale (steps2/gradient2)

Co-authored-by: smasongarrison <6001608+smasongarrison@users.noreply.github.com>
