+++ tests/testthat/test-ggPedigree.R
@@ -280,6 +280,7 @@ test_that("focal fill works with ID and different methods", {
   library(BGmisc)
   data("potter") # load example data from BGmisc
 
+  # Test with greyscale theme
   p <- ggPedigree(potter,
     famID = "famID",
     personID = "personID",
@@ -287,12 +288,28 @@ test_that("focal fill works with ID and different methods", {
       focal_fill_include = TRUE,
       sex_color_include = FALSE,
       focal_fill_personID = 1,
-      focal_fill_method = "steps"
+      focal_fill_method = "steps",
+      color_theme = "greyscale"
     )
   )
   expect_s3_class(p, "gg") # Should return a ggplot object
   expect_true("focal_fill" %in% names(p$data)) # focal_fill column should be present
   expect_true(all(p$data$focal_fill >= 0 & p$data$focal_fill <= 1)) # focal_fill values should be between 0 and 1
+  
+  # Test that greyscale color scale is applied
+  # The plot should have a colour scale for the focal fill
+  colour_scales <- p$scales$find("colour")
+  expect_true(length(colour_scales) > 0, "Plot should have a colour scale for focal fill")
+  
+  # Build the plot to verify greyscale colors are used
+  built <- ggplot2::ggplot_build(p)
+  expect_s3_class(built, "ggplot_built")
+  
+  # Verify that the scale uses greyscale by checking the scale aesthetics
+  scale <- p$scales$scales[[colour_scales[1]]]
+  expect_true("colour" %in% scale$aesthetics)
+  # The scale should be a gradient2 or steps2 type scale
+  expect_true(inherits(scale, "ScaleContinuous"))
 
   p2 <- ggPedigree(potter,
     famID = "famID",
