### 2025-12-31 14b097b
Fix scale access guard in greyscale test
Added proper guard check before accessing scale object to prevent
indexing errors if colour_scales is empty. Updated comment to
accurately describe the test being performed.

Co-authored-by: smasongarrison <6001608+smasongarrison@users.noreply.github.com>
