### 2025-12-31 efc80b7
Add greyscale theme assertions to focal fill test (#114)
* Initial plan

* Remove unnecessary greyscale parameter from focal fill test

The test 'focal fill works with ID and different methods' is specifically
testing focal fill functionality with different methods, not greyscale
theming. The color_theme parameter was added but no assertions about
greyscale colors were included. Since there's already a dedicated
greyscale test in test-defaultPlotConfig.R, this parameter is unnecessary
and potentially misleading.

Co-authored-by: smasongarrison <6001608+smasongarrison@users.noreply.github.com>

[…]
