+++ reference/getDefaultPlotConfig.md
@@ -18,6 +18,11 @@ getDefaultPlotConfig(
   apply_default_scales = TRUE,
   apply_default_theme = TRUE,
   segment_default_color = "black",
+  color_theme = "color",
+  greyscale_palette_default = c("grey10", "grey50", "grey85"),
+  greyscale_low = "black",
+  greyscale_mid = "grey50",
+  greyscale_high = "white",
   color_palette_default = c("#440154FF", "#7fd34e", "#f1e51d"),
   color_palette_low = "#000004FF",
   color_palette_mid = "#56106EFF",
@@ -216,6 +221,30 @@ getDefaultPlotConfig(
 - segment_default_color:
 
   A character string for the default color of segments in the plot.
+
+- color_theme:
+
+  Theme mode controlling default palettes. Options: "color" (default) or
+  "greyscale".
+
+- greyscale_palette_default:
+
+  Default discrete greyscale palette used when color_theme="greyscale".
+
+- greyscale_low:
+
+  Greyscale low color for continuous gradients when
+  color_theme="greyscale".
+
+- greyscale_mid:
+
+  Greyscale midpoint color for continuous gradients when
+  color_theme="greyscale".
+
+- greyscale_high:
+
+  Greyscale high color for continuous gradients when
+  color_theme="greyscale".
 
 - color_palette_default:
 
