+++ reference/calculateCoordinates.md
@@ -119,20 +119,20 @@ head(coords)
 #> 4        4     1       Lily Evans       Lily   Evans   1   103   104        5
 #> 5        5     1     James Potter      James  Potter   1    NA    NA        4
 #> 6        6     1   Dudley Dursley     Dudley Dursley   2     3     1       NA
-#>   sex twinID zygosity nid x_pos x_order y_order y_pos parent_fam spousehint
-#> 1   1     NA     <NA>   1   1.0       2       2     2          1          1
-#> 2   0     NA     <NA>   2   0.0       1       2     2          1          0
-#> 3   0     NA     <NA>   3   2.0       3       2     2          3          0
-#> 4   0     NA     <NA>   4   4.0       5       2     2          3          0
-#> 5   1     NA     <NA>   5   3.0       4       2     2          0          1
-#> 6   1     NA     <NA>   6   1.5       1       3     3          2          0
-#>   y_fam extra x_fam
-#> 1     1 FALSE   0.5
-#> 2     1 FALSE   0.5
-#> 3     1 FALSE   3.0
-#> 4     1 FALSE   3.0
-#> 5    NA FALSE    NA
-#> 6     2 FALSE   1.5
+#>   sex twinID zygosity nid        x_pos x_order y_order y_pos parent_fam
+#> 1   1     NA     <NA>   1 1.000000e+00       2       2     2          1
+#> 2   0     NA     <NA>   2 5.551115e-17       1       2     2          1
+#> 3   0     NA     <NA>   3 2.000000e+00       3       2     2          3
+#> 4   0     NA     <NA>   4 4.000000e+00       5       2     2          3
+#> 5   1     NA     <NA>   5 3.000000e+00       4       2     2          0
+#> 6   1     NA     <NA>   6 1.500000e+00       1       3     3          2
+#>   spousehint y_fam extra x_fam
+#> 1          1     1 FALSE   0.5
+#> 2          0     1 FALSE   0.5
+#> 3          0     1 FALSE   3.0
+#> 4          0     1 FALSE   3.0
+#> 5          1    NA FALSE    NA
+#> 6          0     2 FALSE   1.5
 
 # Example with custom configuration
 coords_custom <- calculateCoordinates(
@@ -168,20 +168,20 @@ head(coords)
 #> 4        4     1       Lily Evans       Lily   Evans   1   103   104        5
 #> 5        5     1     James Potter      James  Potter   1    NA    NA        4
 #> 6        6     1   Dudley Dursley     Dudley Dursley   2     3     1       NA
-#>   sex twinID zygosity nid x_pos x_order y_order y_pos parent_fam spousehint
-#> 1   1     NA     <NA>   1   1.0       2       2     2          1          1
-#> 2   0     NA     <NA>   2   0.0       1       2     2          1          0
-#> 3   0     NA     <NA>   3   2.0       3       2     2          3          0
-#> 4   0     NA     <NA>   4   4.0       5       2     2          3          0
-#> 5   1     NA     <NA>   5   3.0       4       2     2          0          1
-#> 6   1     NA     <NA>   6   1.5       1       3     3          2          0
-#>   y_fam extra x_fam
-#> 1     1 FALSE   0.5
-#> 2     1 FALSE   0.5
-#> 3     1 FALSE   3.0
-#> 4     1 FALSE   3.0
-#> 5    NA FALSE    NA
-#> 6     2 FALSE   1.5
+#>   sex twinID zygosity nid        x_pos x_order y_order y_pos parent_fam
+#> 1   1     NA     <NA>   1 1.000000e+00       2       2     2          1
+#> 2   0     NA     <NA>   2 5.551115e-17       1       2     2          1
[…]
