### 2025-12-31 7449387
add more explanations
