+++ vignettes/articles/v01_plots_extended.Rmd
@@ -1,8 +1,8 @@
 ---
-title: "Extended Plotting pedigrees with ggPedigree()"
+title: "Extended: Plotting pedigrees with ggPedigree()"
 output: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{Extended Plotting pedigrees with ggPedigree()}
+  %\VignetteIndexEntry{Extended: Plotting pedigrees with ggPedigree()}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
@@ -23,6 +23,7 @@ library(viridis) # viridis for color palettes
 library(tidyverse) # for data wrangling
 ```
 
+This vignette demonstrates advanced plotting capabilities of the `ggpedigree` package, focusing on creating custom pedigrees for publication and handling complex pedigree structures. We will walk through the steps to prepare pedigree data, customize plot aesthetics, and generate publication-quality figures. It extends the basic examples found in the main package documentation.
 
 # Constructing Custom Pedigrees for Publication
 
@@ -142,7 +143,7 @@ df_fig1 <- tribble(
 
 ## Plotting the pedigree
 
-This example shows how to create a custom pedigree plot highlighting individuals from a specific mitochondrial lineage (in blue). The plot uses various configuration options to adjust the appearance of the pedigree, including point size, outline, and segment colors. 
+This example shows how to create a custom pedigree plot highlighting individuals from a specific mitochondrial lineage (in blue). The plot uses various configuration options to adjust the appearance of the pedigree, including point size, outline, and segment colors.  Debug mode is enabled to facilitate fine-tuning of the layout as well as to extract coordinates for overlaying additional information.
 
 ```{r, dimensions = c(6, 4)}
 fig1 <- ggPedigree(
@@ -268,6 +269,9 @@ p <- ggPedigree(
     status_code_unaffected = FALSE,
     generation_height = 1,
     point_size = 2,
+    label_text_size = 2,
+    label_method = "ggrepel",
+    label_segment_color = "gray",
     point_scale_by_pedigree = FALSE,
     generation_width = 1,
     status_shape_affected = 4,
