+++ vignettes/articles/v21_phenotypicbydegree.Rmd
@@ -1,10 +1,10 @@
 ---
-title: "Extended Plotting Phenotypic Correlation by Degree of Relatedness"
+title: "Extended: Plotting Phenotypic Correlation by Degree of Relatedness"
 author: "S. Mason Garrison"
 date: "`r Sys.Date()`"
 output: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{Plotting Phenotypic Correlation by Degree of Relatedness}
+  %\VignetteIndexEntry{Extended: Plotting Phenotypic Correlation by Degree of Relatedness}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
@@ -16,7 +16,7 @@ knitr::opts_chunk$set(
 )
 ```
 
-This vignette demonstrates how to visualize phenotypic correlation by degree of relatedness using the `ggPhenotypeByDegree` function from the `ggpedigree` package. This function is particularly useful for analyzing and visualizing the relationship between phenotypic traits and genetic relatedness in pedigree data.
+This vignette demonstrates how to visualize phenotypic correlation by degree of relatedness using the `ggPhenotypeByDegree` function from the `ggpedigree` package. This function is particularly useful for analyzing and visualizing the relationship between phenotypic traits and genetic relatedness in pedigree data. It extends the basic examples found in the main package documentation.
 
 The example below uses a squirrel pedigree dataset to illustrate how to create a plot showing the phenotypic correlation across different degrees of relatedness.
 <details>
