### 2025-12-31 704649c
Add vignette on configuration options and update docs
Added a new vignette 'v02_configuration.Rmd' demonstrating how to use the config argument to control ggpedigree plots. Updated .gitignore to exclude vignette config files, incremented package version to 1.1.0.2, and updated NEWS.md to document the new vignette and recent changes.
