### 2025-12-31 R-Computing-Lab/ggpedigree#94 @smasongarrison
I've now updated the color defaults to be more asectic as well as adding a new color_theme option to allow gray scale. Thanks for nudging me to revisit the defaults!
