### 2025-12-31 2c2d87f
sex_legend_show is FALSE
added missing guides for shape legend when sex_legend_show is FALSE.
