+++ R/ggpedigreeCore.R
@@ -94,7 +94,6 @@ ggPedigree.core <- function(ped,
   }
 
 
-
   # -----
   # STEP 4: Coordinate Generation
   # -----
@@ -297,7 +296,6 @@ ggPedigree.core <- function(ped,
   # Add overlay points for affected status if applicable
 
   if (.should_add_overlay(config, overlay_column, status_column, focal_fill_column)) {
-
     # If overlay_column is specified, use it for alpha aesthetic
 
     p <- .addOverlay(
@@ -333,21 +331,20 @@ ggPedigree.core <- function(ped,
   # -----
   # STEP 11: Scales, Theme
   # -----
-# scale min
+  # scale min
 
 
- if(is.na(min(ds$y_pos, na.rm = TRUE)) ) {
-   warning("y_pos contains all NA values, cannot set y-axis limits. This occurs when there are less than two people in the pedigree.")
-p <- p +
-    ggplot2::scale_y_reverse()
- } else {
-
-  p <- p +
-    ggplot2::scale_y_reverse(limits = c(
-      NA,
-      min(ds$y_pos, na.rm = TRUE)
-    ))
- }
+  if (is.na(min(ds$y_pos, na.rm = TRUE))) {
+    warning("y_pos contains all NA values, cannot set y-axis limits. This occurs when there are less than two people in the pedigree.")
+    p <- p +
+      ggplot2::scale_y_reverse()
+  } else {
+    p <- p +
+      ggplot2::scale_y_reverse(limits = c(
+        NA,
+        min(ds$y_pos, na.rm = TRUE)
+      ))
+  }
   if (config$apply_default_theme == TRUE) {
     p <- p +
       ggplot2::theme_minimal() +
@@ -444,7 +441,6 @@ p <- p +
   )
 
 
-
   # 3) Add the selected layer
   if (node_mode == "sex_color") {
     plotObject <- plotObject +
@@ -456,7 +452,6 @@ p <- p +
         size = config$point_size,
         na.rm = TRUE
       )
-
   } else if (node_mode == "focal_fill") {
     # Preserve your original "if focal_fill_column is NULL, use .data$focal_fill"
     color_expr <- if (is.null(focal_fill_column)) rlang::expr(.data$focal_fill) else rlang::sym(focal_fill_column)
@@ -470,7 +465,6 @@ p <- p +
         size = config$point_size,
         na.rm = TRUE
       )
-
   } else if (node_mode == "status") {
     plotObject <- plotObject +
       ggplot2::geom_point(
@@ -481,7 +475,6 @@ p <- p +
         size = config$point_size,
         na.rm = TRUE
       )
-
   } else { # "shape_only"
     plotObject <- plotObject +
       ggplot2::geom_point(
@@ -492,7 +485,6 @@ p <- p +
   }
 
   plotObject
-
 }
 
 #' @rdname dot-addNodes
@@ -506,11 +498,12 @@ addNodes <- .addNodes
 #' @return A ggplot object with added overlay.
 #'
 .addOverlay <- function(plotObject,
-                        config = list(overlay_include = FALSE,
-                                      status_include = FALSE,
-                                      focal_fill_include = FALSE,
-                                      sex_color_include = FALSE),
-
+                        config = list(
+                          overlay_include = FALSE,
+                          status_include = FALSE,
+                          focal_fill_include = FALSE,
+                          sex_color_include = FALSE
+                        ),
                         focal_fill_column = NULL,
                         status_column = NULL,
                         overlay_column = NULL) {
@@ -520,27 +513,31 @@ addNodes <- .addNodes
     rules = list(
       list(
         when = function() isTRUE(config$overlay_include) && !is.null(overlay_column),
-        do   = list(
+        do = list(
[…]
