### 2025-12-31 33808e8
Add recode_missing_sex option to config and handling
Introduces a new config parameter, recode_missing_sex, to control recoding of missing sex values in pedigrees. Updates ggPedigree.core to handle both character and numeric sex codes based on this option, updates documentation, and adjusts tests to reflect the new parameter.
