### 2025-12-31 29d8c9a
Improve missing sex recoding and update label method
Refactored missing sex recoding logic in ggPedigree.core to handle both character and numeric sex variables more robustly. Updated vignette to use 'geom_text_repel' instead of 'ggrepel' for label_method.

Co-Authored-By: Copilot <175728472+Copilot@users.noreply.github.com>
