+++ R/ggpedigreeCore.R
@@ -166,21 +166,26 @@ ggPedigree.core <- function(ped,
   config$gap_woff <- 0.5 * config$generation_width # single constant for all “stub” offsets
 
   # recode missing sex to "unknown"
-  if (config$recode_missing_sex == TRUE) {
-    if (any(is.na(ds$sex)) && is.character(ds$sex)) {
+  if (config$recode_missing_sex == TRUE && any(is.na(ds$sex))) {
+    non_na_sex <- unique(ds$sex)[!is.na(ds$sex)]
+    n_unique_sex <- length(non_na_sex)
+
+    if (is.character(ds$sex)) {
       ds <- ds |>
         dplyr::mutate(sex = dplyr::case_when(
           is.na(.data$sex) ~ "unknown",
           TRUE ~ as.character(.data$sex)
         ))
-    } else if (any(is.na(ds$sex)) && is.numeric(ds$sex)) {
-      if (length(unique(ds$sex)[!is.na(unique(ds$sex))]) == 3 && (max(ds$sex, na.rm = TRUE) == 3 || max(ds$sex, na.rm = TRUE) == 2)) {
+    } else if (is.numeric(ds$sex)) {
+      max_sex <- max(ds$sex, na.rm = TRUE)
+
+      if (n_unique_sex == 3 && (max_sex == 3 || max_sex == 2)) {
         ds <- ds |>
           dplyr::mutate(sex = dplyr::case_when(
             is.na(.data$sex) ~ 3,
             TRUE ~ as.numeric(.data$sex)
           ))
-      } else if (length(unique(ds$sex)[!is.na(unique(ds$sex))]) == 2) {
+      } else if (n_unique_sex == 2) {
         ds <- ds |>
           dplyr::mutate(sex = dplyr::case_when(
             is.na(.data$sex) ~ max(ds$sex, na.rm = TRUE) + 1,
