### 2025-12-31 4c40e19
Update vignettes/articles/v02_configuration.Rmd
Update vignettes/articles/v02_configuration.Rmd

Update vignettes/articles/v02_configuration.Rmd

Co-Authored-By: Copilot <175728472+Copilot@users.noreply.github.com>
