### 2026-01-01 e660e10
Update documentation and vignettes for 1.1.0.2 release
Updated DESCRIPTION, NEWS.md, and README files to reflect version 1.1.0.2, including new features, bug fixes, and improved documentation. Enhanced vignettes with additional configuration examples, added a custom emoji pedigree image, and made minor improvements to plot configuration and labeling in vignette code.
