+++ README.Rmd
@@ -32,17 +32,11 @@ options(citation.bibtex.max = 0)
 
 <!-- badges: end -->
 
-`ggpedigree` provides modern tools for visualizing pedigree structures using both 'ggplot2' and 'plotly'. It is designed to work seamlessly with the `BGmisc` package for simulated or empirical pedigree data, and extends the plotting capabilities of the base-graphics functions in `kinship2`.
+`ggpedigree` provides modern tools for visualizing family trees and pedigree structures using both `ggplot2` and `plotly`. Specifically, it enables the creation of static and interactive pedigree plots that can accommodate complex family relationships, including duplicated individuals and various mating structures.
 
-The package supports:
+The package complements a behavior genetics package `BGmisc` [Garrison et al. (2024) <doi:10.21105/joss.06203>] by rendering pedigrees using the 'ggplot2' framework. Features include support for duplicated individuals, complex mating structures, integration with simulated pedigrees, and layout customization. Due to the impending deprecation of `kinship2`, version 1.0 incorporates the layout helper functions from kinship2. The pedigree alignment algorithms are adapted from 'kinship2' [Sinnwell et al. (2014) <doi:10.1159/000363105>]. We gratefully acknowledge the original authors: Jason Sinnwell, Terry Therneau, Daniel Schaid, and Elizabeth Atkinson for their foundational work. 
 
-- Static, publication-ready pedigree plots using the `ggplot2` framework
 
-- Interactive pedigree visualizations using `plotly`
-
-- Layout customization, complex mating structures, and duplicated individuals
-
-`ggpedigree` is useful in behavior genetics, kinship analysis, and any research involving complex pedigree data.
 
 ## Installation
 
