+++ authors.md
@@ -10,14 +10,14 @@
 Source:
 [`DESCRIPTION`](https://github.com/R-Computing-Lab/ggpedigree/blob/main/DESCRIPTION)
 
-Garrison S (2025). *ggpedigree: Visualizing Pedigrees with 'ggplot2' and
+Garrison S (2026). *ggpedigree: Visualizing Pedigrees with 'ggplot2' and
 'plotly'*. R package version 1.1.0.2,
 <https://github.com/R-Computing-Lab/ggpedigree/>.
 
     @Manual{,
       title = {ggpedigree: Visualizing Pedigrees with 'ggplot2' and 'plotly'},
       author = {S. Mason Garrison},
-      year = {2025},
+      year = {2026},
       note = {R package version 1.1.0.2},
       url = {https://github.com/R-Computing-Lab/ggpedigree/},
     }
