### 2026-01-02 c939a72
Clarify BGmisc dependency, keep aliases, and add self-contained examples
Co-authored-by: smasongarrison <6001608+smasongarrison@users.noreply.github.com>
