### 2026-01-02 64f5e85
make it clearer how bgmisc and ggpedigree differ
