### 2026-01-02 35e5605
Document canonical API and add optimiseStaticPedigree alias
Documented ggPedigree(), ggPedigreeInteractive(), and ggRelatednessMatrix() as canonical entry points in NEWS.md. Added and exported the optimiseStaticPedigree alias for optimizeStaticPedigree, updated NAMESPACE and documentation accordingly. Improved README and vignette for clarity and updated example and feature descriptions.

Add British English aliases for pedigree functions

Introduces 'optimisePedigree' and 'optimiseStaticPedigree' as British English aliases for 'optimizePedigree' and 'optimizeStaticPedigree'. Updates NAMESPACE 
[…]
