+++ README.Rmd
@@ -34,9 +34,10 @@ options(citation.bibtex.max = 0)
 
 `ggpedigree` provides modern tools for visualizing family trees and pedigree structures using both `ggplot2` and `plotly`. Specifically, it enables the creation of static and interactive pedigree plots that can accommodate complex family relationships, including duplicated individuals and various mating structures.
 
-The package complements a behavior genetics package `BGmisc` [Garrison et al. (2024) <doi:10.21105/joss.06203>] by rendering pedigrees using the 'ggplot2' framework. Features include support for duplicated individuals, complex mating structures, integration with simulated pedigrees, and layout customization. Due to the impending deprecation of `kinship2`, version 1.0 incorporates the layout helper functions from kinship2. The pedigree alignment algorithms are adapted from 'kinship2' [Sinnwell et al. (2014) <doi:10.1159/000363105>]. We gratefully acknowledge the original authors: Jason Sinnwell, Terry Therneau, Daniel Schaid, and Elizabeth Atkinson for their foundational work. 
+The package complements the behavior genetics toolkit `BGmisc` [Garrison et al. (2024) <doi:10.21105/joss.06203>] for tasks such as pedigree validation and computing relatedness matrices. Core plotting functions (`ggPedigree()`, `ggPedigreeInteractive()`, `calculateCoordinates()`) work on any rectangular pedigree and ship with built-in example data, while relatedness workflows reuse matrix builders from `BGmisc`. Due to the impending deprecation of `kinship2`, version 1.0 incorporates the layout helper functions from kinship2. The pedigree alignment algorithms are adapted from 'kinship2' [Sinnwell et al. (2014) <doi:10.1159/000363105>]. We gratefully acknowledge the original authors: Jason Sinnwell, Terry Therneau, Daniel Schaid, and Elizabeth Atkinson for their foundational work. 
 
 
+See the [package vignettes](https://r-computing-lab.github.io/ggpedigree/articles/) for end-to-end tutorials, including relatedness matrices and interactive plots.
 
 ## Installation
 
