+++ man/optimizeStaticPedigree.Rd
@@ -11,6 +11,13 @@ optimizeStaticPedigree(
   variable_drop = c("parent_hash", "couple_hash", "gen", "spousehint", "parent_fam",
     "nid", "x_order", "y_order", "y_fam", "zygosity", "extra", "x_fam")
 )
+
+optimiseStaticPedigree(
+  p,
+  config = list(),
+  variable_drop = c("parent_hash", "couple_hash", "gen", "spousehint", "parent_fam",
+    "nid", "x_order", "y_order", "y_fam", "zygosity", "extra", "x_fam")
+)
 }
 \arguments{
 \item{p}{A ggplot object representing the pedigree plot.}
