### 2026-01-02 R-Computing-Lab/ggpedigree#95 @smasongarrison
with sex coding defined
``` r
library(ggpedigree)
df <- data.frame(
  personID = c("1", "2", "3"), momID = c(NA, NA, "1"),
  dadID = c(NA, NA, "2"), sex = c(2, 1, 2)
)

ggPedigree(df,
           config = list(code_male = 1, code_female = 2))
```

![](https://i.imgur.com/Atw8tnh.png)

``` r

df <- data.frame(
  personID = 1:3, momID = c(0, 0, 1),
  dadID = c(0, 0, 2), sex = c(2, 1, 2)
)
df
#>   personID momID dadID sex
#> 1        1     0     0   2
#> 2        2     0     0   1
#> 3        3     1     2   2

ggPedigree(df,
           config = list(code_male = 1, code_female = 2))
```

![](https://i.imgur.com/vRuU5cj.png)

``` r

df <- data.frame(
  personID = 1:3, dadID = c(0, 0, 1),
  momID = c(0, 0, 2), sex = c(1, 2, 1)
)
df
#>   personID dadID momID sex
#> 1        1     0     0   1
#> 2        2     0     0   2
#> 3        3     1     2   1

ggPedigree(df,
           config = list(code_male = 1, code_female = 2))
```

![](https://i.imgur.com/gdpUVQA.png)

``` r

df <- data.frame(
  personID = 1:3, dadID = c(0, 0, 1),
  momID = c(0, 0, 2), sex = c(1, 2, NA)
)
df
#>   personID dadID momID sex
#> 1        1     0     0   1
#> 2        2     0     0   2
#> 3        3     1     2  NA

ggPedigree(df,
           config = list(code_male = 1, code_female = 2))
#> Warning in pedigree.sexrepair(sex = sex): More than 25% of the sex values are
#> 'unknown'. Please try using config options to specify male and female labels.
```

![](https://i.imgur.com/GLtqTbr.png)

``` r

df <- data.frame(
  personID = c("a", "b", "c"), dadID = c(NA, NA, "a"),
  momID = c(NA, NA, "b"), sex = c(1, 0, 1)
)
df
#>   personID dadID momID sex
#> 1        a  <NA>  <NA>   1
#> 2        b  <NA>  <NA>   0
#> 3        c     a     b   1

ggPedigree(df,
           config = list(code_male = 0, code_female = 1)
           ) + ggplot2::scale_size_area()
#> Warning in .ped2id(ped = ped, personID = personID, momID = momID, dadID =
#> dadID, : After converting IDs to numeric, all IDs became NA. This indicates ID
#> coercion collapsed IDs. Please ensure IDs aren't character or factor variables.
#> Error: Error in constructing pedigree object. Please check that you've correctly specified the sex of individuals. Setting code_male may help if non-standard codes are used (e.g., 'M'/'F'; '1,2').
#>
#> Underlying error: Id not male, but is a father: a

df <- data.frame(personID = 1:3, dadID = c(0, 0, 1), momID = c(0, 0, 2))
ggPedigree(cbind(df, sex = c(1, 3, 1)),  # 1 -> female, 3 -> male
           config = list(code_male = 3, code_female = 1, code_unknown = 2))
#> Error: Error in constructing pedigree object. Please check that you've correctly specified the sex of individuals. Setting code_male may help if non-standard codes are used (e.g., 'M'/'F'; '1,2').
#>
#> Underlying error: Id not male, but is a father: 1
ggPedigree(cbind(df, sex = c(1, 2, 0)),# 0 -> female, 2 -> unknown
           config = list(code_male = 1, code_female = 0, code_unknown = 2))
```

![](https://i.imgur.com/cPM30Jh.png)

``` r
ggPedigree(cbind(df, sex = c(1, 2, 3)),# 2 -> male, 3 -> unknown
           config = list(code_male = 2, code_female = 1, code_unknown = 3)
           )
#> Error: Error in constructing pedigree object. Please check that you've correctly specified the sex of individuals. Setting code_male may help if non-standard codes are used (e.g., 'M'/'F'; '1,2').
#>
#> Underlying error: Id not male, but is a father: 1
```
