### 2026-01-02 20fb7cf
Improve sex label handling and messaging in pedigree functions
Expanded sex value handling in ggPedigree.core to support factors and integers, and standardized NA recoding to 'Unknown'. Updated error and warning messages in pedigree.sexrepair to mention all label config options for clarity.
