### 2026-01-02 R-Computing-Lab/ggpedigree#95 @smasongarrison
Ok so with the latest BGmisc update this will be fixed https://github.com/R-Computing-Lab/BGmisc/commit/63182abea6980832954772abe99e7c5b175c452c
