+++ NAMESPACE
@@ -17,6 +17,7 @@ export(ggpedigreeInteractive)
 export(ggpedigreeinteractive)
 export(ggrelatednessmatrix)
 export(kinship2_autohint)
+export(kinship2_plotPedigree)
 export(optimisePedigree)
 export(optimiseStaticPedigree)
 export(pedigree)
