### 2026-01-03 6bd5ef4
Remove deprecated aliases and fix function naming
Deprecated aliases such as ggpedigreeInteractive, ggpedigreeinteractive, ggrelatednessmatrix, optimisePedigree, and optimiseStaticPedigree were removed from the NAMESPACE, R scripts, and documentation. Function calls and configuration references were updated to use the correct camelCase naming (e.g., ggPedigree, ggPedigreeInteractive). This improves consistency and reduces confusion for users.
