+++ R/documentData.R
@@ -72,7 +72,7 @@
 #'     config = list(
 #'       add_phantoms = TRUE,
 #'       code_male = "M"
-#'   )
+#'     )
 #'   )
 #' }
 "redsquirrels"
@@ -120,7 +120,7 @@
 #' @examples
 #' # Load the ASOIAF dataset
 #' data(ASOIAF)
-#' df_ASOIAF <-  ASOIAF[ASOIAF$famID == 26,]  # Subset to House Tarth
+#' df_ASOIAF <- ASOIAF[ASOIAF$famID == 26, ] # Subset to House Tarth
 #' # View the structure of the dataset
 #' str(df_ASOIAF)
 #'
