+++ R/calcCoordinates.R
@@ -65,8 +65,9 @@ utils::globalVariables(c(":="))
 #'   personID = "personID",
 #'   momID = "momID",
 #'   dadID = "dadID",
-#'   config  = list(
-#'    code_male = 1)
+#'   config = list(
+#'     code_male = 1
+#'   )
 #' )
 #'
 #' # View the coordinates
@@ -83,7 +84,6 @@ utils::globalVariables(c(":="))
 #'     ped_width = 20
 #'   )
 #' )
-
 calculateCoordinates <- function(ped,
                                  personID = "personID",
                                  momID = "momID",
@@ -106,7 +106,7 @@ calculateCoordinates <- function(ped,
   # -----
   # Set up
   # -----
-  if(!is.null(code_male)){
+  if (!is.null(code_male)) {
     config$code_male <- code_male
   }
   # Fill missing configuration values with defaults
@@ -319,7 +319,7 @@ alignPedigreeWithRelations <- function(ped,
   #   recode_male = "M", recode_female = "F", recode_na = NA_character_)
   # Recode sex values in case non-standard codes are used (e.g., "M"/"F")
 
-  if(!is.null(code_male)){
+  if (!is.null(code_male)) {
     config$code_male <- code_male
     code_male <- NULL
   }
