+++ articles/v02_configuration.md
@@ -41,11 +41,11 @@ A basic pedigree plot uses defaults:
 
 ``` r
 ggPedigree(
-potter,
-famID = "famID",
-personID = "personID",
-momID = "momID",
-dadID = "dadID"
+  potter,
+  famID = "famID",
+  personID = "personID",
+  momID = "momID",
+  dadID = "dadID"
 )
 ```
 
@@ -125,17 +125,17 @@ This example customizes labels. Here we label individuals by
 
 ``` r
 ggPedigree(
-potter,
-famID = "famID",
-personID = "personID",
-momID = "momID",
-dadID = "dadID",
-config = list(
-label_include = TRUE,
-label_column = "first_name",
-label_text_size = 3.2,
-label_nudge_y = 0.15
-)
+  potter,
+  famID = "famID",
+  personID = "personID",
+  momID = "momID",
+  dadID = "dadID",
+  config = list(
+    label_include = TRUE,
+    label_column = "first_name",
+    label_text_size = 3.2,
+    label_nudge_y = 0.15
+  )
 )
 ```
 
@@ -145,12 +145,12 @@ To turn labels off completely:
 
 ``` r
 ggPedigree(
-potter,
-famID = "famID",
-personID = "personID",
-momID = "momID",
-dadID = "dadID",
-config = list(label_include = FALSE)
+  potter,
+  famID = "famID",
+  personID = "personID",
+  momID = "momID",
+  dadID = "dadID",
+  config = list(label_include = FALSE)
 )
 ```
 
@@ -160,18 +160,18 @@ You can also use repelled labels to avoid overlaps:
 
 ``` r
 ggPedigree(
-potter,
-famID = "famID",
-personID = "personID",
-momID = "momID",
-dadID = "dadID",
-config = list(
-label_include = TRUE,
-label_method = "geom_text_repel",
-label_max_overlaps = 10,
-label_text_size = 9,
-label_segment_color = "grey50"
-)
+  potter,
+  famID = "famID",
+  personID = "personID",
+  momID = "momID",
+  dadID = "dadID",
+  config = list(
+    label_include = TRUE,
+    label_method = "geom_text_repel",
+    label_max_overlaps = 10,
+    label_text_size = 9,
+    label_segment_color = "grey50"
+  )
 )
 ```
 
@@ -200,19 +200,19 @@ points.
 
 ``` r
 ggPedigree(
-potter,
-famID = "famID",
-personID = "personID",
-momID = "momID",
-dadID = "dadID",
-config = list(
-point_scale_by_pedigree = FALSE,
-point_size = 6,
-outline_include = TRUE,
-outline_color = "maroon",
-outline_multiplier = 1.5,
-outline_alpha = 1
-)
+  potter,
+  famID = "famID",
+  personID = "personID",
+  momID = "momID",
+  dadID = "dadID",
+  config = list(
+    point_scale_by_pedigree = FALSE,
+    point_size = 6,
+    outline_include = TRUE,
+    outline_color = "maroon",
+    outline_multiplier = 1.5,
+    outline_alpha = 1
+  )
 )
 ```
 
@@ -233,17 +233,17 @@ segment color.
 
 ``` r
 ggPedigree(
-potter,
-famID = "famID",
-personID = "personID",
-momID = "momID",
-dadID = "dadID",
-config = list(
-segment_linewidth = 2,
-segment_spouse_color = "steelblue",
-segment_parent_color = "black",
-segment_offspring_color = "black"
-)
+  potter,
+  famID = "famID",
+  personID = "personID",
+  momID = "momID",
+  dadID = "dadID",
+  config = list(
+    segment_linewidth = 2,
[…]
