+++ reference/ggRelatednessMatrix.md
@@ -13,15 +13,6 @@ ggRelatednessMatrix(
   personID = "personID",
   ...
 )
-
-ggrelatednessmatrix(
-  mat,
-  config = list(),
-  interactive = FALSE,
-  tooltip_columns = NULL,
-  personID = "personID",
-  ...
-)
 ```
 
 ## Arguments
