+++ reference/ASOIAF.md
@@ -53,7 +53,7 @@ The variables are as follows:
 ``` r
 # Load the ASOIAF dataset
 data(ASOIAF)
-df_ASOIAF <-  ASOIAF[ASOIAF$famID == 26,]  # Subset to House Tarth
+df_ASOIAF <- ASOIAF[ASOIAF$famID == 26, ] # Subset to House Tarth
 # View the structure of the dataset
 str(df_ASOIAF)
 #> 'data.frame':    6 obs. of  9 variables:
