+++ reference/dot-addScales.md
@@ -19,12 +19,12 @@ addScales(plotObject, config, status_column = NULL, focal_fill_column = NULL)
 - config:
 
   A list of configuration options for customizing the plot. See
-  getDefaultPlotConfig for details of each option. The list can include:
+  getDefaultPlotConfig for details. The list can include:
 
   code_male
 
   :   Integer or string. Value identifying males in the sex column.
-      (typically 0 or 1) Default: 1
+      (typically 0 or 1) Default: 1.
 
   segment_spouse_color, segment_self_color
 
