### 2026-01-03 14e234c
add other rendering of emoji
