### 2026-01-03 555b751
Update NEWS and clarify vignette introduction
Updated NEWS.md to mention two vignettes for ggpedigree configuration options. Improved clarity in the introduction of the extended configuration vignette by referencing previous explanations and rephrasing dataset usage.
