+++ articles/v32_plots_morecomplexity.md
@@ -1,0 +1,264 @@
+# Extended: Plotting more complicated pedigrees with \`ggPedigree()\`
+
+## Introduction
+
+This vignette demonstrates some of the more complex family trees you can
+visualization with
+[`ggPedigree()`](https://r-computing-lab.github.io/ggpedigree/reference/ggPedigree.md).
+It extends the basic examples found in the main package documentation.
+
+We illustrate usage with a more complicated data set from ggpedigree:
+`asoiaf` a fictional pedigree based on characters from the *A Song of
+Ice and Fire* universe by George R. R. Martin. This dataset includes
+complex relationships, such as half-siblings, multiple marriages, and
+consanguineous relationships, making it an excellent example for
+showcasing the capabilities of
+[`ggPedigree()`](https://r-computing-lab.github.io/ggpedigree/reference/ggPedigree.md).
+
+## Basic usage
+
+``` r
+# Install your package from GitHub
+
+library(ggpedigree) # ggPedigree lives here
+library(BGmisc) # helper utilities & example data
+library(ggplot2) # ggplot2 for plotting
+library(viridis) # viridis for color palettes
+library(tidyverse) # for data wrangling
+```
+
+### Load Packages and Data
+
+We begin by loading the required libraries and examining the structure
+of the built-in `ASOIAF` pedigree.
+
+``` r
+data(ASOIAF)
+```
+
+The ASOIAF dataset includes character IDs, names, family identifiers,
+and parent identifiers for a subset of characters drawn from the *A Song
+of Ice and Fire* canon.
+
+``` r
+head(ASOIAF)
+#>   id famID momID dadID          name sex
+#> 1  1     1   566   564   Walder Frey   M
+#> 2  2     1    NA    NA   Perra Royce   F
+#> 3  3     1     2     1  Stevron Frey   M
+#> 4  4     1     2     1    Emmon Frey   M
+#> 5  5     1     2     1    Aenys Frey   M
+#> 6  6     1    NA    NA Corenna Swann   F
+#>                                                   url twinID zygosity
+#> 1   https://awoiaf.westeros.org/index.php/Walder_Frey     NA     <NA>
+#> 2   https://awoiaf.westeros.org/index.php/Perra_Royce     NA     <NA>
+#> 3  https://awoiaf.westeros.org/index.php/Stevron_Frey     NA     <NA>
+#> 4    https://awoiaf.westeros.org/index.php/Emmon_Frey     NA     <NA>
+#> 5    https://awoiaf.westeros.org/index.php/Aenys_Frey     NA     <NA>
+#> 6 https://awoiaf.westeros.org/index.php/Corenna_Swann     NA     <NA>
+```
+
+## Data Cleaning
+
+``` r
+df_got <- checkSex(ASOIAF,
+  code_male = "M",
+  code_female = "F",
+  repair = TRUE
+)
+
+
+# Find the IDs of Jon Snow and Daenerys Targaryen
+
+jon_id <- df_got %>%
+  filter(name == "Jon Snow") %>%
+  pull(ID)
+
+dany_id <- df_got %>%
+  filter(name == "Daenerys Targaryen") %>%
+  pull(ID)
+```
+
+Here we’ve added phantom IDs to the dataset, which are used to represent
+individuals who are not directly related to the family tree but are
+still part of the pedigree structure. This is useful for visualizing
+complex relationships, such as half-siblings or step-siblings.
+
+``` r
+df_repaired <- checkParentIDs(df_got,
+  addphantoms = TRUE,
+  repair = TRUE,
+  parentswithoutrow = FALSE,
[…]
