+++ reference/buildPlotConfig.md
@@ -7,7 +7,12 @@ all necessary parameters are set.
 ## Usage
 
 ``` r
-buildPlotConfig(default_config, config, function_name = "ggPedigree")
+buildPlotConfig(
+  default_config,
+  config,
+  function_name = "ggPedigree",
+  pedigree_size = NULL
+)
 ```
 
 ## Arguments
@@ -23,6 +28,10 @@ buildPlotConfig(default_config, config, function_name = "ggPedigree")
 - function_name:
 
   The name of the function for which the configuration is being built.
+
+- pedigree_size:
+
+  Size of the pedigree, used for point scaling.
 
 ## Value
 
