+++ pkgdown.yml
@@ -11,7 +11,7 @@ articles:
   v30_matrix: v30_matrix.html
   articles/v31_phenotypicbydegree: v31_phenotypicbydegree.html
   articles/v32_plots_morecomplexity: v32_plots_morecomplexity.html
-last_built: 2026-01-04T05:33Z
+last_built: 2026-01-04T05:37Z
 urls:
   reference: https://r-computing-lab.github.io/ggpedigree/reference
   article: https://r-computing-lab.github.io/ggpedigree/articles
