### 2026-01-04 488f25c
Update v11_configuration_extended.Rmd
