+++ vignettes/v10_configuration.R
@@ -25,4 +25,3 @@ ggPedigree(
   momID = "momID",
   dadID = "dadID"
 )
-
