+++ tests/testthat/test-plotPedigree.R
@@ -9,13 +9,13 @@ test_that("simulated pedigree plots correctly", {
 
   results <- simulatePedigree(kpc = kpc, Ngen = Ngen, sexR = sexR, marR = marR)
 
-  expect_no_error(plotPedigree(results, verbose = FALSE))
+  expect_no_error(kinship2_plotPedigree(results, verbose = FALSE))
 
   kpc <- 2
   results2 <- simulatePedigree(kpc = kpc, Ngen = Ngen, sexR = sexR, marR = marR)
   results2$fam <- paste0("fam 2")
   results <- rbind(results, results2)
-  expect_output(plotPedigree(results, verbose = TRUE))
+  expect_output(kinship2_plotPedigree(results, verbose = TRUE))
 })
 
 
@@ -32,10 +32,10 @@ test_that("pedigree plots correctly with affected variables", {
     n = nrow(results),
     size = 1, prob = .1
   )
-  expect_error(plotPedigree(results, verbose = TRUE, affected = affected[1:10]))
+  expect_error(kinship2_plotPedigree(results, verbose = TRUE, affected = affected[1:10]))
   results$affected <- affected
-  expect_output(plotPedigree(results, verbose = TRUE, affected = "affected"))
-  expect_output(plotPedigree(results, verbose = TRUE, affected = results$affected))
+  expect_output(kinship2_plotPedigree(results, verbose = TRUE, affected = "affected"))
+  expect_output(kinship2_plotPedigree(results, verbose = TRUE, affected = results$affected))
 
   # file.remove("Rplots.pdf")
 })
@@ -45,7 +45,7 @@ test_that("pedigree errs when affected variables named", {
   library(BGmisc)
   data(inbreeding)
 
-  expect_error(plotPedigree(data, verbose = TRUE, affected = "affected"))
+  expect_error(kinship2_plotPedigree(data, verbose = TRUE, affected = "affected"))
 })
 
 
@@ -53,5 +53,5 @@ test_that("pedigree plots multiple families", {
   library(BGmisc)
   data(inbreeding)
   plotPedigree(inbreeding, verbose = TRUE)
-  expect_message(plotPedigree(inbreeding, verbose = TRUE))
+  expect_message(kinship2_plotPedigree(inbreeding, verbose = TRUE))
 })
