### 2026-01-04 938e1ac
Update test-plotPedigree.R and more examples
