+++ R/ggPedigreeInteractive.R
@@ -11,6 +11,32 @@
 #' library(BGmisc)
 #' data("potter")
 #' ggPedigreeInteractive(potter, famID = "famID", personID = "personID")
+#'
+#'
+#' data(hazard)
+#' ggPedigreeInteractive(
+#'   hazard,
+#'   famID = "famID",
+#'   personID = "ID",
+#'   momID = "momID",
+#'   dadID = "dadID",
+#'   config = list(
+#'     code_male = 0,
+#'     status_column = "affected",
+#'     label_nudge_y = .25,
+#'     label_include = TRUE,
+#'     include_tooltip = TRUE,
+#'     label_method = "geom_text",
+#'     sex_color_include = TRUE
+#'   ),
+#'   tooltip_columns = c("personID", "birthYr", "onsetYr", "deathYr")
+#' ) %>%
+#'   plotly::layout(
+#'     title = "Hazard Pedigree (interactive)",
+#'     hoverlabel = list(bgcolor = "white"),
+#'     margin = list(l = 50, r = 50, t = 50, b = 50)
+#'   ) %>%
+#'   plotly::config(displayModeBar = TRUE)
 #' @export
 #' @seealso ggPedigree.core, ggPedigree, vignette("v20_interactiveplots")
 
