### 2026-01-04 c33604f
Tidy up vignette code and formatting
Rounded and simplified numeric values in v31_phenotypicbydegree.Rmd for readability, improved comments and line breaks in v00_plots.R, adjusted ggplot2 theme and scale formatting in v00_plots.Rmd, and updated HTML outputs for v00_plots.html and v20_interactiveplots.html to reflect these changes and improve code block formatting.

Refactor code style and improve test clarity

Updated .lintr to exclude data-raw/ and ignore pipe consistency linter. Improved code clarity and comments in R scripts, reformatted and clarified test cases in test-ggPedigree.R and te
[…]
