### 2026-01-05 3b3dee7
Refactor relink and closest_dup functions in processExtras
Moved closest_dup and relink functions out of processExtras, making them standalone with explicit dup_xy_list argument. Updated relink call to pass dup_xy_list, improving clarity and modularity.
