+++ man/closest_dup.Rd
@@ -1,0 +1,22 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/processExtras.R
+\name{closest_dup}
+\alias{closest_dup}
+\title{Find the closest duplicate appearance based on coordinates}
+\usage{
+closest_dup(target_core, x0, y0, dup_xy_list)
+}
+\arguments{
+\item{target_core}{The coreID of the individual to find.}
+
+\item{x0}{The x-coordinate of the reference point.}
+
+\item{y0}{The y-coordinate of the reference point.}
+}
+\value{
+The personID of the closest duplicate appearance.
+}
+\description{
+Find the closest duplicate appearance based on coordinates
+}
+\keyword{internal}
