### 2026-01-05 1b02859
Update documentation and references for clarity
Expanded @seealso sections in ggPedigreeInteractive and ggRelatednessMatrix to reference additional vignettes. Fixed minor formatting issues in documentation for warsofroses and related .Rd files to improve consistency and clarity.
