+++ man/warsofroses.Rd
@@ -29,7 +29,7 @@ The variables are as follows:
  \item \code{dadID}:  ID of the father
  \item \code{name}:  Name of the person
  \item \code{sex}: Biological sex
-  \item \code{url}:  URL to a wiki page about the character
+ \item \code{url}:  URL to a wiki page about the character
   }
 }
 \keyword{datasets}
