+++ articles/v31_phenotypicbydegree.md
@@ -218,10 +218,10 @@ df <- pedigree_df <- tribble(
   ~USA_flag_10_polychorFunction_chisq,
   ~USA_flag_10_polychorFunction_df,
   ~USA_flag_10_ml_polychorFunction,
-  3617250, 0.225, 0.275, 0.2255859375, 0.247734086859983, 0.25, 0.27490234375, 1, 0, 59.2335021779651, 0.480619365892598, NA, NA, 0.249700992238048, 0.04517140756813, 0.000691314990032303, 0.00000799819827079773, 0, 0.0451713825840523,
-  4983424, 0.225, 0.275, 0.2255859375, 0.247985118108249, 0.25, 0.27490234375, 0, 0, 60.2924621468198, 0.54683237363258, NA, NA, 0.247046760504304, 0.0409988164718674, 0.000599673001112523, 0.00000799819827079773, 0, 0.040998793040341,
-  120024, 0.275, 0.45, 0.359375, 0.42720031928978, 0.4375, 0.44921875, 1, 1, 55.3786442198885, 0.519857671410953, NA, NA, 0.208202301154754, 0.169242204331722, 0.00393269696620211, 0.00000814738450571895, 0, 0.169234306738862,
-  137947, 0.275, 0.45, 0.275146484375, 0.336850848231778, 0.34375, 0.447265625, 1, 0, 58.7404851093794, 0.436444779593302, NA, NA, 0.230130064360623, 0.0979941906533943, 0.00359788288853913, 0.00000800052657723427, 0, 0.0979945954952227,
+  3617250, 0.225, 0.275, 0.2255, 0.2477, 0.25, 0.2749, 1, 0, 59.2335, 0.4806, NA, NA, 0.2497, 0.0451, 0.0006, 0.00000799, 0, 0.0451,
+  4983424, 0.225, 0.275, 0.2255, 0.2479, 0.25, 0.2749, 0, 0, 60.2924, 0.5468, NA, NA, 0.2470, 0.0409, 0.0005, 0.00000799, 0, 0.0409,
+  120024, 0.275, 0.45, 0.359375, 0.42720031928978, 0.4375, 0.4492, 1, 1, 55.3786, 0.519857671410953, NA, NA, 0.2082, 0.1692, 0.00393269696620211, 0.0000081, 0, 0.169234306738862,
+  137947, 0.275, 0.45, 0.2751, 0.3368, 0.34375, 0.4472, 1, 0, 58.7404, 0.4364, NA, NA, 0.230130064360623, 0.0979941906533943, 0.00359788288853913, 0.00000800052657723427, 0, 0.0979945954952227,
   108989, 0.275, 0.45, 0.275390625, 0.328092281014401, 0.3125, 0.4453125, 0, 0, 60.2531852003308, 0.619355993501189, NA, NA, 0.233078768161892, 0.0774003012451702, 0.00404657995476565, 0.00000800139969214797, 0, 0.0774005589379321,
   1668205, 0.45, 0.55, 0.4501953125, 0.495319721631096, 0.4990234375, 0.549560546875, 1, 1, 57.9295757847709, 0.522333747318167, NA, NA, 0.257514897453126, 0.160645710133276, 0.000975187176863543, 0.0000904696062207222, 0, 0.16065612400207,
   729441, 0.45, 0.55, 0.451171875, 0.496536543871464, 0.5, 0.5498046875, 1, 0, 66.1279781018233, 0.275059425002599, NA, NA, 0.265100488202374, 0.150191012774079, 0.00152996759330951, 0.00000815372914075851, 0, 0.150193105587773,
[…]
