+++ reference/closest_dup.md
@@ -1,0 +1,27 @@
+# Find the closest duplicate appearance based on coordinates
+
+Find the closest duplicate appearance based on coordinates
+
+## Usage
+
+``` r
+closest_dup(target_core, x0, y0, dup_xy_list)
+```
+
+## Arguments
+
+- target_core:
+
+  The coreID of the individual to find.
+
+- x0:
+
+  The x-coordinate of the reference point.
+
+- y0:
+
+  The y-coordinate of the reference point.
+
+## Value
+
+The personID of the closest duplicate appearance.
