### 2026-01-06 R-Computing-Lab/ggpedigree#93 @magnusdv
I don’t think the intended warning is showing yet. I’m seeing this:

``` r
library(ggpedigree)

df = data.frame(personID = 1, momID = 0, dadID = 0, sex = 1)
df
#>   personID momID dadID sex
#> 1        1     0     0   1

ggPedigree(df)
#> Warning in min(ds$y_pos, na.rm = TRUE): no non-missing arguments to min;
#> returning Inf
#> Warning in min(ds$y_pos, na.rm = TRUE): no non-missing arguments to min;
#> returning Inf
#> Warning in min(ds$y_pos, na.rm = TRUE): no non-missing arguments to min;
#> returning Inf
```

![](https://i.imgur.com/4gyIFkq.png)
