### 2026-01-06 R-Computing-Lab/ggpedigree#123 @magnusdv
Sex column name
Context: openjournals/joss-reviews#9434

Is this supposed to work, or am I doing something wrong?

``` r
library(ggpedigree)

df = data.frame(personID = 1:3, dadID = c(0,0,1), momID = c(0,0,2), foo = c(1,2,1))
df
#>   personID dadID momID foo
#> 1        1     0     0   1
#> 2        2     0     0   2
#> 3        3     1     2   1

ggpedigree(df, sexVar = "foo")
#> Error:
#> ! Error in constructing pedigree object. Please check that you've correctly specified the sex of individuals. Setting code_male may help if non-standard codes are used (e.g., 'M'/'F'; '1,2').
#>
#> Underlying error: All sex values are labeled as unknown. Please try using config options to specify male, female, and unknown labels (code_male, code_female, code_unknown)
```
