### 2026-01-06 28d6ad6
updated documentation about required variables
updated documentation about required variables https://github.com/R-Computing-Lab/ggpedigree/issues/92
