+++ man/ggPedigreeInteractive.Rd
@@ -27,7 +27,7 @@ ggPedigreeInteractive(
 )
 }
 \arguments{
-\item{ped}{A data frame containing the pedigree data. Needs personID, momID, and dadID columns}
+\item{ped}{A data frame containing the pedigree data. Needs personID, momID, dadID, and sex columns.}
 
 \item{famID}{Character string specifying the column name for family IDs. Defaults to "famID".}
 
