+++ man/preparePedigreeData.Rd
@@ -26,7 +26,7 @@ preparePedigreeData(
 )
 }
 \arguments{
-\item{ped}{A data frame containing the pedigree data. Needs personID, momID, and dadID columns}
+\item{ped}{A data frame containing the pedigree data. Needs personID, momID, dadID, and sex columns.}
 
 \item{famID}{Character string specifying the column name for family IDs. Defaults to "famID".}
 
