+++ vignettes/articles/paper.bib
@@ -229,7 +229,7 @@
 }
 @Article{bgmisc,
   title = {BGmisc: An R Package for Extended Behavior Genetics Analysis},
-  author = {{{Garrison}, S. M.} and {{Hunter}, Michael D.} and {{Lyu}, X.} and {{Trattner}, J. D.} and {{Burt}, S. A.}},
+  author = {{Garrison}, S. M. and {Hunter}, Michael D. and {Lyu}, X. and {Trattner}, J. D. and {Burt}, S. A.},
   journal = {Journal of Open Source Software},
   year = {2024},
   volume = {9},
@@ -396,8 +396,7 @@
 	author = {{Martin}, George R. R.},
 	year = {1997},
 	date = {1997},
-	publisher = {Bantam},
-	volume = {1}
+	publisher = {Bantam}
 }
 
 @book{martin2014,
@@ -515,7 +514,7 @@ year = {2025},
 issn = {2352-3964},
 doi = {10.1016/j.ebiom.2025.105911},
 url = {https://www.sciencedirect.com/science/article/pii/S235239642500355X},
-author = {S. Alexandra Burt and S. Mason Garrison and Xuanyu Lyu and Joseph L. Rodgers and Sarah L. Carroll and Ken R. Smith and Michael D. Hunter},
+author = {S. Alexandra {Burt} and S. Mason {Garrison} and Xuanyu {Lyu} and Joseph L. {Rodgers} and Sarah L. {Carroll} and Ken R. {Smith} and Michael D. {Hunter}},
 keywords = {mtDNA, Longevity, Extended pedigrees, Matrilineal relatedness, Heritability},
 abstract = {}
 }
