+++ reference/ggPedigree.core.md
@@ -33,8 +33,8 @@ ggPedigree.core(
 
 - ped:
 
-  A data frame containing the pedigree data. Needs personID, momID, and
-  dadID columns
+  A data frame containing the pedigree data. Needs personID, momID,
+  dadID, and sex columns.
 
 - famID:
 
