+++ reference/calculateCoordinates.md
@@ -25,8 +25,8 @@ calculateCoordinates(
 
 - ped:
 
-  A data frame containing the pedigree data. Needs personID, momID, and
-  dadID columns
+  A data frame containing the pedigree data. Needs personID, momID,
+  dadID, and sex columns.
 
 - personID:
 
