+++ reference/preparePedigreeData.md
@@ -32,8 +32,8 @@ preparePedigreeData(
 
 - ped:
 
-  A data frame containing the pedigree data. Needs personID, momID, and
-  dadID columns
+  A data frame containing the pedigree data. Needs personID, momID,
+  dadID, and sex columns.
 
 - famID:
 
