+++ reference/transformPed.md
@@ -29,8 +29,8 @@ transformPed(
 
 - ped:
 
-  A data frame containing the pedigree data. Needs personID, momID, and
-  dadID columns
+  A data frame containing the pedigree data. Needs personID, momID,
+  dadID, and sex columns.
 
 - famID:
 
