+++ reference/ggPedigree.md
@@ -60,8 +60,8 @@ ggpedigree(
 
 - ped:
 
-  A data frame containing the pedigree data. Needs personID, momID, and
-  dadID columns
+  A data frame containing the pedigree data. Needs personID, momID,
+  dadID, and sex columns.
 
 - famID:
 
