+++ reference/calculateConnections.md
@@ -24,8 +24,8 @@ calculateConnections(
 
 - ped:
 
-  A data frame containing the pedigree data. Needs personID, momID, and
-  dadID columns
+  A data frame containing the pedigree data. Needs personID, momID,
+  dadID, and sex columns.
 
 - config:
 
