### 2026-01-06 14ed2b4
Update paper text and Potter pedigree image
Refined wording and citation formatting in _paper.Rmd and paper.md, including improved references to BGmisc and funding details. Updated wfu_potter_pedigree.png with a new image version.
