### 2026-01-06 a612fbf
Revise ggpedigree feature descriptions in paper
Updated wording in the vignette and paper to clarify feature descriptions and improve readability. Minor edits include rephrasing example data references and simplifying technical explanations. Regenerated PDF and updated pedigree image to reflect these changes.
