### 2026-01-06 7e22783
Update paper text and figures for clarity
Revised wording in _paper.Rmd and paper.md for improved clarity and accuracy, including citation formatting and acknowledgments. Updated paper.pdf and wfu_potter_pedigree.png to reflect these changes.
