### 2026-01-06 96fe8bc
Update default segment linewidth and documentation
Increased the default segment_linewidth for better visibility in plots and updated related documentation, examples, and vignettes. Also updated the package version to 1.1.0.3 and improved README and vignette text for clarity.
