### 2026-01-06 20d8e26
Update relink and closest_dup to accept dup_xy argument
Added the dup_xy argument to the relink and closest_dup functions, updated their documentation, and modified processExtras to pass the new argument. This change improves flexibility and clarity in handling duplicate appearances in pedigree layouts.
