+++ vignettes/articles/paper.bib
@@ -28,7 +28,7 @@
 
 @article{kinship2,
 	author = {{Sinnwell}, J. P. and {Therneau}, T. M. and {Schaid}, D. J.},
-	title = {The kinship2 r package for pedigree data},
+	title = {The kinship2 {R} package for pedigree data},
 	journal = {Human Heredity},
 	year = {2014},
 	volume = {78}, issue = {2},
@@ -228,7 +228,7 @@
   year          = {1984},
 }
 @Article{bgmisc,
-  title = {BGmisc: An R Package for Extended Behavior Genetics Analysis},
+  title = {BGmisc: An {R} Package for Extended Behavior Genetics Analysis},
   author = {{Garrison}, S. M. and {Hunter}, Michael D. and {Lyu}, X. and {Trattner}, J. D. and {Burt}, S. A.},
   journal = {Journal of Open Source Software},
   year = {2024},
@@ -326,7 +326,7 @@
   url = {https://juliengamartin.github.io/pedtricks/},
 }
 @inproceedings{garrison2024,
-	title = {Charting New Paths in Behavior Genetics: Developing a Comprehensive R Visualization Atlas},
+	title = {Charting New Paths in Behavior Genetics: Developing a Comprehensive {R} Visualization Atlas},
 	author = {{Garrison}, S. M.},
 	year = {2024},
 	journal = {Behavior Genetics},
@@ -353,6 +353,8 @@
 	author = {{Lyu}, X. and {Hunter}, Michael D. and {Burt}, S. A. and {Good}, R. and {Carroll}, S. L. and {Garrison}, S. M.},
 	year = {2025},
 	journal = {Behavior Genetics},
+	pages = {320-337},
+	volume = {55},
 	doi= {10.1007/s10519-025-10225-1},
 	url = {https://link.springer.com/article/10.1007/s10519-025-10225-1}
 }
@@ -436,7 +438,7 @@ PMCID: PMC4133633}
 }
 
 @book{vigeland2021,
-	title = {Pedigree analysis in R},
+	title = {Pedigree analysis in {R}},
 	author = {{Vigeland}, Magnus D.},
 	year = {2021},
 	month = {04},
@@ -463,7 +465,7 @@ PMCID: PMC4133633}
 }
 
 @Manual{pedtools,
-	title = {Pedigree analysis in {\textbraceleft}R{\textbraceright}},
+	title = {Pedigree analysis in {R}},
 	author = {{Vigeland}, Magnus D.},
 	year = {2021},
 	doi = {10.32614/CRAN.package.pedtools},
@@ -492,7 +494,7 @@ PMCID: PMC4133633}
 }
 
 @article{morrissey2010,
-	title = {pedantics: an r package for pedigree-based genetic simulation and pedigree manipulation, characterization and viewing},
+	title = {pedantics: an {R} package for pedigree-based genetic simulation and pedigree manipulation, characterization and viewing},
 	author = {{Morrissey}, Michael B. and {Wilson}, Alastair J.},
 	year = {2010},
 	month = {07},
@@ -520,7 +522,7 @@ abstract = {}
 }
 @article{pedixplorer,
     author = {Le Nézet, Louis and Sinnwell, Jason and Letko, Anna and André, Catherine and Quignon, Pascale},
-    title = {Pedixplorer: a Bioconductor package to streamline pedigree design and visualization},
+    title = {Pedixplorer: a {Bioconductor} package to streamline pedigree design and visualization},
     journal = {Bioinformatics},
     volume = {41},
     number = {6},
