### 2026-01-07 #9434 @csoneson
Thanks @smasongarrison. I have gone through the submission and overall it looks good - I made a small [PR](https://github.com/R-Computing-Lab/ggpedigree/pull/126) fixing some capitalization and formatting in the paper, please check and merge if you agree.

I will create another post below with the remaining steps (for me and for you). In particular, note that your Zenodo archive metadata needs an update to agree with the title of the paper.
