### 2026-01-07 20534db
Modify license and enhance DESCRIPTION file
Updated license information and improved package description.
