+++ DESCRIPTION
@@ -7,7 +7,7 @@ Authors@R: c(
            comment = c(ORCID = "0000-0002-4804-6003"))
   )
 Description: Provides plotting functions for visualizing pedigrees and family trees. The package complements a behavior genetics package 'BGmisc' [Garrison et al. (2024) <doi:10.21105/joss.06203>] by rendering pedigrees using the 'ggplot2' framework. Features include support for duplicated individuals, complex mating structures, integration with simulated pedigrees, and layout customization. Due to the impending deprecation of kinship2, version 1.0 incorporates the layout helper functions from kinship2. The pedigree alignment algorithms are adapted from 'kinship2' [Sinnwell et al. (2014) <doi:10.1159/000363105>]. We gratefully acknowledge the original authors: Jason Sinnwell, Terry Therneau, Daniel Schaid, and Elizabeth Atkinson for their foundational work. 
-License: GPL-3
+License: GPL (>= 3)
 URL: https://github.com/R-Computing-Lab/ggpedigree/,
      https://r-computing-lab.github.io/ggpedigree/
 BugReports: https://github.com/R-Computing-Lab/ggpedigree/issues
