### 2026-01-08 d5ae2da
Improve robustness of sexVar handling in pedigree functions
Refactor core and helper functions to more robustly handle cases where the sex variable is named differently than 'sex'. Add tests to ensure correct behavior when sexVar is not 'sex'. Update documentation and internal logic to support flexible sex variable names.
