### 2026-01-21 22ca714
Update point size expectation in tests
